Trait parquet::arrow::array_reader::ArrayReader

pub trait ArrayReader {
    fn as_any(&self) -> &dyn Any;

    fn get_data_type(&self) -> &ArrowType;

    fn next_batch(&mut self, batch_size: usize) -> Result<ArrayRef>;

    fn get_def_levels(&self) -> Option<&[i16]>;

    fn get_rep_levels(&self) -> Option<&[i16]>;
}

Array reader reads parquet data into arrow array.

Required methods

fn as_any(&self) -> &dyn Any

fn get_data_type(&self) -> &ArrowType

Returns the arrow type of this array reader.

fn next_batch(&mut self, batch_size: usize) -> Result<ArrayRef>

Reads at most batch_size records into an arrow array and return it.

fn get_def_levels(&self) -> Option<&[i16]>

Returns the definition levels of data from last call of next_batch. The result is used by parent array reader to calculate its own definition levels and repetition levels, so that its parent can calculate null bitmap.

fn get_rep_levels(&self) -> Option<&[i16]>

Return the repetition levels of data from last call of next_batch. The result is used by parent array reader to calculate its own definition levels and repetition levels, so that its parent can calculate null bitmap.

Loading content...