Module parquet::arrow::arrow_reader

Contains reader which reads parquet data into arrow RecordBatch

Structs

ArrowPredicateFn

An ArrowPredicate created from an FnMut

ArrowReaderBuilder

A generic builder for constructing sync or async arrow parquet readers. This is not intended to be used directly, instead you should use the specialization for the type of reader you wish to use

ArrowReaderMetadata

The cheaply clone-able metadata necessary to construct a ArrowReaderBuilder

ArrowReaderOptions

Options that control how metadata is read for a parquet file

ParquetRecordBatchReader

An Iterator<Item = ArrowResult<RecordBatch>> that yields RecordBatch read from a parquet data source

RowFilter

A RowFilter allows pushing down a filter predicate to skip IO and decode

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

Traits

ArrowPredicate

A predicate operating on RecordBatch

RowGroups

A collection of row groups

Type Aliases

ParquetRecordBatchReaderBuilder

A synchronous builder used to construct ParquetRecordBatchReader for a file