Trait parquet::arrow::arrow_reader::filter::ArrowPredicate

pub trait ArrowPredicate: Send + 'static {
    // Required methods
    fn projection(&self) -> &ProjectionMask;
    fn evaluate(
        &mut self,
        batch: RecordBatch
    ) -> Result<BooleanArray, ArrowError>;
}

A predicate operating on RecordBatch

Required Methods

fn projection(&self) -> &ProjectionMask

Returns the ProjectionMask that describes the columns required to evaluate this predicate. All projected columns will be provided in the batch passed to evaluate

fn evaluate(&mut self, batch: RecordBatch) -> Result<BooleanArray, ArrowError>

Evaluate this predicate for the given RecordBatch containing the columns identified by Self::projection

Rows that are true in the returned BooleanArray will be returned by the parquet reader, whereas rows that are false or Null will not be