parquet::arrow::arrow_reader

Function evaluate_predicate

pub(crate) fn evaluate_predicate(
    batch_size: usize,
    array_reader: Box<dyn ArrayReader>,
    input_selection: Option<RowSelection>,
    predicate: &mut dyn ArrowPredicate,
) -> Result<RowSelection>

Evaluates an ArrowPredicate, returning a RowSelection indicating which rows to return.

input_selection: Optional pre-existing selection. If Some, then the final RowSelection will be the conjunction of it and the rows selected by predicate.

Note: A pre-existing selection may come from evaluating a previous predicate or if the ParquetRecordBatchReader specified an explicit RowSelection in addition to one or more predicates.