parquet::arrow::arrow_reader

Module read_plan

ReadPlan and ReadPlanBuilder for determining which rows to read from a Parquet file

Structs

LimitedReadPlanBuilder 🔒

Builder for ReadPlan that applies a limit and offset to the read plan

ReadPlan

A plan reading specific rows from a Parquet Row Group.

ReadPlanBuilder

A builder for ReadPlan