parquet::arrow::arrow_reader

Module statistics

StatisticsConverter to convert statistics in parquet format to arrow [ArrayRef].

Macros

get_data_page_statistics 🔒

get_statistics 🔒

Special macro to combine the statistics iterators for min and max using the paste macro. This is used to avoid repeating the same code for min and max statistics extractions

make_decimal_stats_iterator 🔒

Special iterator adapter for extracting i128 values from from an iterator of ParquetStatistics

make_stats_iterator 🔒

Define an adapter iterator for extracting statistics from an iterator of ParquetStatistics

Structs

MaxBooleanStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxByteArrayStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxDecimal32StatsIterator 🔒

MaxDecimal64StatsIterator 🔒

MaxDecimal128StatsIterator 🔒

MaxDecimal256StatsIterator 🔒

MaxDoubleStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxFixedLenByteArrayStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxFloatStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxInt32StatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MaxInt64StatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinBooleanStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinByteArrayStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinDecimal32StatsIterator 🔒

MinDecimal64StatsIterator 🔒

MinDecimal128StatsIterator 🔒

MinDecimal256StatsIterator 🔒

MinDoubleStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinFixedLenByteArrayStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinFloatStatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinInt32StatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

MinInt64StatsIterator 🔒

Maps an iterator of ParquetStatistics into an iterator of `&$stat_value_type``

StatisticsConverter

Extracts Parquet statistics as Arrow arrays

Functions

from_bytes_to_f16 🔒

from_bytes_to_i32 🔒

from_bytes_to_i64 🔒

from_bytes_to_i128 🔒

from_bytes_to_i256 🔒

max_page_statistics 🔒

Extracts the max statistics from an iterator of parquet page ColumnIndexMetaData’s to an [ArrayRef]

max_statistics 🔒

Extracts the max statistics from an iterator of ParquetStatistics to an [ArrayRef]

min_page_statistics 🔒

Extracts the min statistics from an iterator of parquet page ColumnIndexMetaData’s to an [ArrayRef]

min_statistics 🔒

Extracts the min statistics from an iterator of ParquetStatistics to an [ArrayRef]

null_counts_page_statistics 🔒

Extracts the null count statistics from an iterator of parquet page ColumnIndexMetaData’s to an [ArrayRef]