parquet::arrow::arrow_reader::statistics

Macro get_statistics

macro_rules! get_statistics {
    (Min, $data_type: ident, $iterator: ident, $physical_type: ident) => { ... };
    (Max, $data_type: ident, $iterator: ident, $physical_type: ident) => { ... };
    (
        $data_type: ident,
        $iterator: ident,
        $physical_type: ident,
        $boolean_iter: ident,
        $int32_iter: ident,
        $int64_iter: ident,
        $float_iter: ident,
        $double_iter: ident,
        $byte_array_iter: ident,
        $fixed_len_byte_array_iter: ident,
        $decimal32_iter: ident,
        $decimal64_iter: ident,
        $decimal128_iter: ident,
        $decimal256_iter: ident,
        $dictionary_statistics: ident
    ) => { ... };
}

Special macro to combine the statistics iterators for min and max. This is used to avoid repeating the same code for min and max statistics extractions

Parameters: data_type: The data type of the statistics (e.g. DataType::Int32) iterator: The iterator of ParquetStatistics to extract the statistics from.