Trait parquet::arrow::arrow_reader::ArrowPredicate

pub trait ArrowPredicate: Send + 'static {
    // Required methods
    fn projection(&self) -> &ProjectionMask;
    fn evaluate(
        &mut self,
        batch: RecordBatch
    ) -> Result<BooleanArray, ArrowError>;
}

A predicate operating on RecordBatch

Required Methods

fn projection(&self) -> &ProjectionMask

Returns the ProjectionMask that describes the columns required to evaluate this predicate. All projected columns will be provided in the batch passed to evaluate

fn evaluate(&mut self, batch: RecordBatch) -> Result<BooleanArray, ArrowError>

Evaluate this predicate for the given RecordBatch containing the columns identified by Self::projection

Must return a BooleanArray that has the same length as the input batch where each row indicates whether the row should be returned:

true:the row should be returned

false or null: the row should not be returned