parquet::arrow

Module arrow_writer

Contains writer which writes arrow data into parquet data.

Structs

ArrowColumnChunk

The data for a single column chunk, see ArrowColumnWriter

ArrowColumnWriter

Encodes ArrowLeafColumn to ArrowColumnChunk

ArrowLeafColumn

A leaf column that can be encoded by ArrowColumnWriter

ArrowWriter

Encodes [RecordBatch] to parquet

ArrowWriterOptions

Arrow-specific configuration settings for writing parquet files.

Functions

compute_leaves

Computes the ArrowLeafColumn for a potentially nested [ArrayRef]

get_column_writers

Returns the ArrowColumnWriter for a given schema