parquet::arrow::arrow_writer

Module byte_array

Macros

downcast_dict_impl 🔒

downcast_dict_op 🔒

downcast_op 🔒

Structs

ByteArrayEncoder

ByteArrayStorage 🔒

[Storage] for the [Interner] used by DictEncoder

DictEncoder 🔒

A dictionary encoder for byte array data

FallbackEncoder 🔒

A fallback encoder, i.e. non-dictionary, for ByteArray

Enums

FallbackEncoderImpl 🔒

The fallback encoder in use

Functions

compute_min_max 🔒

Computes the min and max for the provided array and indices

encode 🔒

Encodes the provided values and indices to encoder

update_geo_stats_accumulator 🔒

Updates geospatial statistics for the provided array and indices