Enum parquet::arrow::arrow_writer::levels::LevelInfoBuilder

enum LevelInfoBuilder {
    Primitive(ArrayLevels),
    List(Box<LevelInfoBuilder>, LevelContext, OffsetBuffer<i32>, Option<NullBuffer>),
    LargeList(Box<LevelInfoBuilder>, LevelContext, OffsetBuffer<i64>, Option<NullBuffer>),
    FixedSizeList(Box<LevelInfoBuilder>, LevelContext, usize, Option<NullBuffer>),
    Struct(Vec<LevelInfoBuilder>, LevelContext, Option<NullBuffer>),
}

A helper to construct ArrayLevels from a potentially nested Field

Variants

Primitive(ArrayLevels)

A primitive, leaf array

List(Box<LevelInfoBuilder>, LevelContext, OffsetBuffer<i32>, Option<NullBuffer>)

A list array

LargeList(Box<LevelInfoBuilder>, LevelContext, OffsetBuffer<i64>, Option<NullBuffer>)

A large list array

FixedSizeList(Box<LevelInfoBuilder>, LevelContext, usize, Option<NullBuffer>)

A fixed size list array

Struct(Vec<LevelInfoBuilder>, LevelContext, Option<NullBuffer>)

A struct array

Implementations

impl LevelInfoBuilder

fn try_new( field: &Field, parent_ctx: LevelContext, array: &ArrayRef ) -> Result<Self>

Create a new LevelInfoBuilder for the given Field and parent LevelContext

fn finish(self) -> Vec<ArrayLevels>

Finish this LevelInfoBuilder returning the ArrayLevels for the leaf columns as enumerated by a depth-first search

fn write(&mut self, range: Range<usize>)

Given an array, write the level data for the elements in range

fn write_list<O: OffsetSizeTrait>( child: &mut LevelInfoBuilder, ctx: &LevelContext, offsets: &[O], nulls: Option<&NullBuffer>, range: Range<usize> )

Write range elements from ListArray array

Note: MapArrays are ListArray<i32> under the hood and so are dispatched to this method

fn write_struct( children: &mut [LevelInfoBuilder], ctx: &LevelContext, nulls: Option<&NullBuffer>, range: Range<usize> )

Write range elements from StructArray array

fn write_fixed_size_list( child: &mut LevelInfoBuilder, ctx: &LevelContext, fixed_size: usize, nulls: Option<&NullBuffer>, range: Range<usize> )

Write range elements from FixedSizeListArray with child data values and null bitmap nulls.

fn write_leaf(info: &mut ArrayLevels, range: Range<usize>)

Write a primitive array, as defined by is_leaf

fn visit_leaves(&mut self, visit: impl Fn(&mut ArrayLevels) + Copy)

Visits all children of this node in depth first order