parquet::arrow::arrow_writer

Struct ArrowRowGroupWriterFactory

pub struct ArrowRowGroupWriterFactory {
    schema: SchemaDescPtr,
    arrow_schema: SchemaRef,
    props: WriterPropertiesPtr,
    page_store_factory: Arc<dyn PageStoreFactory>,
    file_encryptor: Option<Arc<FileEncryptor>>,
}

Factory that creates new column writers for each row group in the Parquet file.

You can create this structure via an ArrowWriter::into_serialized_writer. See the example on ArrowColumnWriter for how to encode columns in parallel

Fields
schema: SchemaDescPtrarrow_schema: SchemaRefprops: WriterPropertiesPtrpage_store_factory: Arc<dyn PageStoreFactory>file_encryptor: Option<Arc<FileEncryptor>>
Implementations

impl ArrowRowGroupWriterFactory

pub fn new<W: Write + Send>( file_writer: &SerializedFileWriter<W>, arrow_schema: SchemaRef, ) -> Self

Create a new ArrowRowGroupWriterFactory for the provided file writer and Arrow schema

pub fn with_page_store_factory( self, page_store_factory: Arc<dyn PageStoreFactory>, ) -> Self

Set the PageStoreFactory used to allocate the buffer for each column chunk, e.g. to spill completed pages to a temp file or object storage instead of the heap. Defaults to InMemoryPageStoreFactory.

fn create_row_group_writer( &self, row_group_index: usize, ) -> Result<ArrowRowGroupWriter>

pub fn create_column_writers( &self, row_group_index: usize, ) -> Result<Vec<ArrowColumnWriter>>

Create column writers for a new row group, with the given row group index

fn column_writer_factory( &self, row_group_idx: usize, ) -> ArrowColumnWriterFactory