parquet::arrow::async_reader::metadata

Trait MetadataSuffixFetch

pub trait MetadataSuffixFetch: MetadataFetch {
    // Required method
    fn fetch_suffix(&mut self, suffix: usize) -> BoxFuture<'_, Result<Bytes>>;
}

A data source that can be used with ParquetMetaDataReader to load ParquetMetaData via suffix requests, without knowing the file size

Required Methods

fn fetch_suffix(&mut self, suffix: usize) -> BoxFuture<'_, Result<Bytes>>

Return a future that fetches the last n bytes asynchronously

Note the returned type is a boxed future, often created by FutureExt::boxed. See the trait documentation for an example