parquet::arrow::async_reader::store

Struct ParquetObjectReader

pub struct ParquetObjectReader {
    store: Arc<dyn ObjectStore>,
    meta: ObjectMeta,
    metadata_size_hint: Option<usize>,
    preload_column_index: bool,
    preload_offset_index: bool,
    runtime: Option<Handle>,
}

Reads Parquet files in object storage using [ObjectStore].

// Populate configuration from environment
let storage_container = Arc::new(MicrosoftAzureBuilder::from_env().build().unwrap());
let location = Path::from("path/to/blob.parquet");
let meta = storage_container.head(&location).await.unwrap();
println!("Found Blob with {}B at {}", meta.size, meta.location);

// Show Parquet metadata
let reader = ParquetObjectReader::new(storage_container, meta);
let builder = ParquetRecordBatchStreamBuilder::new(reader).await.unwrap();
print_parquet_metadata(&mut stdout(), builder.metadata());

Fields
store: Arc<dyn ObjectStore>meta: ObjectMetametadata_size_hint: Option<usize>preload_column_index: boolpreload_offset_index: boolruntime: Option<Handle>
Implementations

impl ParquetObjectReader

pub fn new(store: Arc<dyn ObjectStore>, meta: ObjectMeta) -> Self

Creates a new ParquetObjectReader for the provided [ObjectStore] and [ObjectMeta]

[ObjectMeta] can be obtained using [ObjectStore::list] or [ObjectStore::head]

pub fn with_footer_size_hint(self, hint: usize) -> Self

Provide a hint as to the size of the parquet fileโs footer, see fetch_parquet_metadata

pub fn with_preload_column_index(self, preload_column_index: bool) -> Self

Load the Column Index as part of Self::get_metadata

pub fn with_preload_offset_index(self, preload_offset_index: bool) -> Self

Load the Offset Index as part of Self::get_metadata

pub fn with_runtime(self, handle: Handle) -> Self

Perform IO on the provided tokio runtime

Tokio is a cooperative scheduler, and relies on tasks yielding in a timely manner to service IO. Therefore, running IO and CPU-bound tasks, such as parquet decoding, on the same tokio runtime can lead to degraded throughput, dropped connections and other issues. For more information see here.

fn spawn<F, O, E>(&self, f: F) -> BoxFuture<'_, Result<O>>
where F: for<'a> FnOnce(&'a Arc<dyn ObjectStore>, &'a Path) -> BoxFuture<'a, Result<O, E>> + Send + 'static, O: Send + 'static, E: Into<ParquetError> + Send + 'static,