parquet::arrow::async_writer
Trait AsyncFileWriter

pub trait AsyncFileWriter: Send {
    // Required methods
    fn write(&mut self, bs: Bytes) -> BoxFuture<'_, Result<()>>;
    fn complete(&mut self) -> BoxFuture<'_, Result<()>>;
}

The asynchronous interface used by AsyncArrowWriter to write parquet files.

Required Methods

fn write(&mut self, bs: Bytes) -> BoxFuture<'_, Result<()>>

Write the provided bytes to the underlying writer

The underlying writer CAN decide to buffer the data or write it immediately. This design allows the writer implementer to control the buffering and I/O scheduling.

The underlying writer MAY implement retry logic to prevent breaking users write process.

fn complete(&mut self) -> BoxFuture<'_, Result<()>>

Flush any buffered data to the underlying writer and finish writing process.

After complete returns Ok(()), caller SHOULD not call write again.