parquet::arrow::buffer

Module dictionary_buffer

Macros

offsets_dict_helper 🔒

Structs

BuildPassthroughHasher 🔒

PassthroughHasher 🔒

Enums

DictionaryBuffer

An array of variable length byte arrays that are potentially dictionary encoded and can be converted into a corresponding [ArrayRef]

Functions

hash_byte_slices 🔒

hashes_as_u64 🔒

pack_values_from_offsets 🔒

pack_values_from_offsets_impl 🔒

Builds a [DictionaryArray] directly from a flat OffsetBuffer using pre-computed hashes to deduplicate values in a single pass, avoiding the intermediate StringArray materialization