parquet::arrow::decoder::delta_byte_array

Struct DeltaByteArrayDecoder

pub struct DeltaByteArrayDecoder {
    prefix_lengths: Vec<i32>,
    suffix_lengths: Vec<i32>,
    data: Bytes,
    length_offset: usize,
    data_offset: usize,
    last_value: Vec<u8>,
}

Decoder for Encoding::DELTA_BYTE_ARRAY

Fields
prefix_lengths: Vec<i32>suffix_lengths: Vec<i32>data: Byteslength_offset: usizedata_offset: usizelast_value: Vec<u8>
Implementations

impl DeltaByteArrayDecoder

pub fn new(data: Bytes) -> Result<Self>

Create a new DeltaByteArrayDecoder with the provided data page

pub fn remaining(&self) -> usize

Returns the number of values remaining

pub fn read<F>(&mut self, len: usize, f: F) -> Result<usize>
where F: FnMut(&[u8]) -> Result<()>,

Read up to len values, returning the number of values read and calling f with each decoded byte slice

Will short-circuit and return on error

pub fn skip(&mut self, to_skip: usize) -> Result<usize>

Skip up to to_skip values, returning the number of values skipped