parquet::arrow

Function parquet_to_arrow_schema

pub fn parquet_to_arrow_schema(
    parquet_schema: &SchemaDescriptor,
    key_value_metadata: Option<&Vec<KeyValue>>,
) -> Result<Schema>

Convert Parquet schema to Arrow schema including optional metadata

Attempts to decode any existing Arrow schema metadata, falling back to converting the Parquet schema column-wise