parquet::arrow

Module in_memory_row_group

Structs

ColumnChunkIterator 🔒

Implements PageIterator for a single column chunk, yielding a single PageReader

FetchRanges 🔒

What ranges to fetch for the columns in this row group

InMemoryRowGroup 🔒

An in-memory collection of column chunks

Enums

ColumnChunkData 🔒

An in-memory column chunk. This allows us to hold either dense column chunks or sparse column chunks and easily access them by offset.