parquet::arrow::in_memory_row_group

Struct InMemoryRowGroup

pub(crate) struct InMemoryRowGroup<'a> {
    pub(crate) offset_index: Option<&'a [OffsetIndexMetaData]>,
    pub(crate) column_chunks: Vec<Option<Arc<ColumnChunkData>>>,
    pub(crate) row_count: usize,
    pub(crate) row_group_idx: usize,
    pub(crate) metadata: &'a ParquetMetaData,
}

An in-memory collection of column chunks

Fields
offset_index: Option<&'a [OffsetIndexMetaData]>column_chunks: Vec<Option<Arc<ColumnChunkData>>>

Column chunks for this row group

row_count: usizerow_group_idx: usizemetadata: &'a ParquetMetaData
Implementations

impl InMemoryRowGroup<'_>

pub(crate) fn fetch_ranges( &self, projection: &ProjectionMask, selection: Option<&RowSelection>, batch_size: usize, cache_mask: Option<&ProjectionMask>, ) -> FetchRanges

Returns the byte ranges to fetch for the columns specified in projection and selection.

cache_mask indicates which columns, if any, are being cached by RowGroupCache. The selection for Cached columns is expanded to batch boundaries to simplify accounting for what data is cached.

pub(crate) fn fill_column_chunks<I>( &mut self, projection: &ProjectionMask, page_start_offsets: Option<Vec<Vec<u64>>>, chunk_data: I, )
where I: IntoIterator<Item = Bytes>,

Fills in self.column_chunks with the data fetched from chunk_data.

This function must be called with the data from the ranges returned by fetch_ranges and the corresponding page_start_offsets, with the exact same and selection.