parquet::arrow::push_decoder

Module reader_builder

Modules

data 🔒

DataRequest tracks and holds data needed to construct InMemoryRowGroups

filter 🔒

FilterInfo state machine for evaluating row filters

Structs

NextState 🔒

Result of a state transition

RowGroupInfo 🔒

The current row group being read and the read plan

RowGroupReaderBuilder 🔒

Builder for ParquetRecordBatchReader for a single row group

Enums

RowGroupDecoderState 🔒

This is the inner state machine for reading a single row group.

Functions

override_selector_strategy_if_needed 🔒

Override the selection strategy if needed.