Module parquet::arrow::record_reader

Modules

buffer 🔒

definition_levels 🔒

Structs

GenericRecordReader

A generic stateful column reader that delimits semantic records

Functions

packed_null_mask 🔒

Returns true if we do not need to unpack the nullability for this column, this is only possible if the max definition level is 1, and corresponds to nulls at the leaf level, as opposed to a nullable parent nested type

Type Aliases

ColumnReader 🔒

RecordReader

A RecordReader is a stateful column reader that delimits semantic records.