Trait parquet::arrow::record_reader::buffer::BufferQueue

pub trait BufferQueue: Sized {
    type Output: Sized;
    type Slice: ?Sized;

    // Required methods
    fn consume(&mut self) -> Self::Output;
    fn spare_capacity_mut(&mut self, batch_size: usize) -> &mut Self::Slice;
    fn set_len(&mut self, len: usize);
}

A buffer that supports writing new data to the end, and removing data from the front

Used by RecordReader to buffer up values before returning a potentially smaller number of values, corresponding to a whole number of semantic records

Required Associated Types

type Output: Sized

type Slice: ?Sized

Required Methods

fn consume(&mut self) -> Self::Output

Consumes the contents of this BufferQueue

fn spare_capacity_mut(&mut self, batch_size: usize) -> &mut Self::Slice

Returns a Self::Slice with at least batch_size capacity that can be used to append data to the end of this BufferQueue

NB: writes to the returned slice will not update the length of BufferQueue instead a subsequent call should be made to BufferQueue::set_len

fn set_len(&mut self, len: usize)

Sets the length of the BufferQueue.

Intended to be used in combination with BufferQueue::spare_capacity_mut

Panics

Implementations must panic if len is beyond the initialized length

Implementations may panic if set_len is called with less than what has been written

This distinction is to allow for implementations that return a default initialized BufferQueue::Slice` which doesn’t track capacity and length separately

For example, BufferQueue returns a default-initialized &mut [T], and does not track how much of this slice is actually written to by the caller. This is still safe as the slice is default-initialized.

Object Safety

This trait is not object safe.