parquet::arrow::record_reader
Type Alias ColumnReader

pub(crate) type ColumnReader<CV> = GenericColumnReader<RepetitionLevelDecoderImpl, DefinitionLevelBufferDecoder, CV>;

Aliased Type

struct ColumnReader<CV> {
    descr: Arc<ColumnDescriptor>,
    page_reader: Box<dyn PageReader<Item = Result<Page, ParquetError>>>,
    num_buffered_values: usize,
    num_decoded_values: usize,
    has_record_delimiter: bool,
    def_level_decoder: Option<DefinitionLevelBufferDecoder>,
    rep_level_decoder: Option<RepetitionLevelDecoderImpl>,
    values_decoder: CV,
}

Fields
descr: Arc<ColumnDescriptor>page_reader: Box<dyn PageReader<Item = Result<Page, ParquetError>>>num_buffered_values: usize

The total number of values stored in the data page.

num_decoded_values: usize

The number of values from the current data page that has been decoded into memory so far.

has_record_delimiter: bool

True if the end of the current data page denotes the end of a record

def_level_decoder: Option<DefinitionLevelBufferDecoder>

The decoder for the definition levels if any

rep_level_decoder: Option<RepetitionLevelDecoderImpl>

The decoder for the repetition levels if any

values_decoder: CV

The decoder for the values