parquet::basic

Enum Compression

pub enum Compression {
    UNCOMPRESSED,
    SNAPPY,
    GZIP(GzipLevel),
    LZO,
    BROTLI(BrotliLevel),
    LZ4,
    ZSTD(ZstdLevel),
    LZ4_RAW,
}

Supported block compression algorithms.

Block compression can yield non-trivial improvements to storage efficiency at the expense of potentially significantly worse encode and decode performance. Many applications, especially those making use of high-throughput and low-cost commodity object storage, may find storage efficiency less important than decode throughput, and therefore may wish to not make use of block compression.

The writers in this crate default to no block compression for this reason.

Applications that do still wish to use block compression, will find Compression::ZSTD to provide a good balance of compression, performance, and ecosystem support. Alternatively, Compression::LZ4_RAW provides much faster decompression speeds, at the cost of typically worse compression ratios. However, it is not as widely supported by the ecosystem, with the Hadoop ecosystem historically favoring the non-standard and now deprecated Compression::LZ4.

Variants

UNCOMPRESSED

No compression.

SNAPPY

Snappy compression

GZIP(GzipLevel)

Gzip compression

LZO

LZO compression

BROTLI(BrotliLevel)

Brotli compression

LZ4

LZ4 compression, (deprecated)

ZSTD(ZstdLevel)

ZSTD compression

LZ4_RAW

LZ4 compression.

Implementations

impl Compression

pub(crate) fn codec_to_string(self) -> String

Returns the codec type of this compression setting as a string, without the compression level.