Enum parquet::basic::ConvertedType

pub enum ConvertedType {

    NONE,
    UTF8,
    MAP,
    MAP_KEY_VALUE,
    LIST,
    ENUM,
    DECIMAL,
    DATE,
    TIME_MILLIS,
    TIME_MICROS,
    TIMESTAMP_MILLIS,
    TIMESTAMP_MICROS,
    UINT_8,
    UINT_16,
    UINT_32,
    UINT_64,
    INT_8,
    INT_16,
    INT_32,
    INT_64,
    JSON,
    BSON,
    INTERVAL,

}

Common types (converted types) used by frameworks when using Parquet. This helps map between types in those frameworks to the base types in Parquet. This is only metadata and not needed to read or write the data.

This struct was renamed from LogicalType in version 4.0.0. If targeting Parquet format 2.4.0 or above, please use LogicalType instead.

Variants

NONE

UTF8

A BYTE_ARRAY actually contains UTF8 encoded chars.

MAP

A map is converted as an optional field containing a repeated key/value pair.

MAP_KEY_VALUE

A key/value pair is converted into a group of two fields.

LIST

A list is converted into an optional field containing a repeated field for its values.

ENUM

An enum is converted into a binary field

DECIMAL

A decimal value. This may be used to annotate binary or fixed primitive types. The underlying byte array stores the unscaled value encoded as twoโs complement using big-endian byte order (the most significant byte is the zeroth element).

This must be accompanied by a (maximum) precision and a scale in the SchemaElement. The precision specifies the number of digits in the decimal and the scale stores the location of the decimal point. For example 1.23 would have precision 3 (3 total digits) and scale 2 (the decimal point is 2 digits over).

DATE

A date stored as days since Unix epoch, encoded as the INT32 physical type.

TIME_MILLIS

The total number of milliseconds since midnight. The value is stored as an INT32 physical type.

TIME_MICROS

The total number of microseconds since midnight. The value is stored as an INT64 physical type.

TIMESTAMP_MILLIS

Date and time recorded as milliseconds since the Unix epoch. Recorded as a physical type of INT64.

TIMESTAMP_MICROS

Date and time recorded as microseconds since the Unix epoch. The value is stored as an INT64 physical type.

UINT_8

An unsigned 8 bit integer value stored as INT32 physical type.

UINT_16

An unsigned 16 bit integer value stored as INT32 physical type.

UINT_32

An unsigned 32 bit integer value stored as INT32 physical type.

UINT_64

An unsigned 64 bit integer value stored as INT64 physical type.

INT_8

A signed 8 bit integer value stored as INT32 physical type.

INT_16

A signed 16 bit integer value stored as INT32 physical type.

INT_32

A signed 32 bit integer value stored as INT32 physical type.

INT_64

A signed 64 bit integer value stored as INT64 physical type.

JSON

A JSON document embedded within a single UTF8 column.

BSON

A BSON document embedded within a single BINARY column.

INTERVAL

An interval of time.

This type annotates data stored as a FIXED_LEN_BYTE_ARRAY of length 12. This data is composed of three separate little endian unsigned integers. Each stores a component of a duration of time. The first integer identifies the number of months associated with the duration, the second identifies the number of days associated with the duration and the third identifies the number of milliseconds associated with the provided duration. This duration of time is independent of any particular timezone or date.