parquet::basic

Struct EncodingMask

pub struct EncodingMask(i32);

A bitmask representing the Encodings employed while encoding a Parquet column chunk.

The Parquet ColumnMetaData struct contains an array that indicates what encodings were used when writing that column chunk. For memory and performance reasons, this crate reduces that array to bitmask, where each bit position represents a different Encoding. This struct contains that bitmask, and provides methods to interact with the data.

Example

// read parquet metadata from a file
let file = open_parquet_file("some_path.parquet");
let mut reader = ParquetMetaDataReader::new();
reader.try_parse(&file).unwrap();
let metadata = reader.finish().unwrap();

// find the encodings used by the first column chunk in the first row group
let col_meta = metadata.row_group(0).column(0);
let encodings = col_meta.encodings_mask();

// check to see if a particular encoding was used
let used_rle = encodings.is_set(Encoding::RLE);

// check to see if all of a set of encodings were used
let used_all = encodings.all_set([Encoding::RLE, Encoding::PLAIN].iter());

// convert mask to a Vec<Encoding>
let encodings_vec = encodings.encodings().collect::<Vec<_>>();

Tuple Fields
0: i32
Implementations

impl EncodingMask

const MAX_ENCODING: i32 = Encoding::MAX_DISCRIMINANT

Highest valued discriminant in the Encoding enum

const ALLOWED_MASK: u32

A mask consisting of unused bit positions, used for validation. This includes the never used GROUP_VAR_INT encoding value of 1.

pub fn try_new(val: i32) -> Result<Self>

Attempt to create a new EncodingMask from an integer.

This will return an error if a bit outside the allowable range is set.

pub fn as_i32(&self) -> i32

Return an integer representation of this EncodingMask.

pub fn new_from_encodings<'a>( encodings: impl Iterator<Item = &'a Encoding>, ) -> Self

Create a new EncodingMask from a collection of Encodings.

pub fn insert(&mut self, val: Encoding)

Mark the given Encoding as present in this mask.

pub fn is_set(&self, val: Encoding) -> bool

Test if a given Encoding is present in this mask.

pub fn is_only(&self, val: Encoding) -> bool

Test if this mask has only the bit for the given Encoding set.

pub fn all_set<'a>(&self, encodings: impl Iterator<Item = &'a Encoding>) -> bool

Test if all Encodings in a given set are present in this mask.

pub fn encodings(&self) -> impl Iterator<Item = Encoding>

Return an iterator over all Encodings present in this mask.

fn mask_to_encodings_iter(mask: i32) -> impl Iterator<Item = Encoding>