parquet::basic

Struct GzipLevel

pub struct GzipLevel(u32);

Represents a valid gzip compression level.

Defaults to 6.

0: least compression

9: most compression (that other software can read)

10: most compression (incompatible with other software, see below)

WARNING:

Level 10 compression can offer smallest file size, but Parquet files created with it will not be readable by other โstandardโ paquet readers.

Do NOT use level 10 if you need other software to be able to read the files. Read below for details.

IMPORTANT:

Thereโs often confusion about the compression levels in flate2 vs arrow as highlighted in issue #1011.

flate2 supports two compression backends: miniz_oxide and zlib.

zlib supports levels from 0 to 9.

miniz_oxide supports levels from 0 to 10.

arrow uses flate with rust_backend feature, which provides miniz_oxide as the backend. Therefore 0-10 levels are supported.

flate2 documents this behavior properly with this commit.

Tuple Fields
0: u32
Implementations

impl GzipLevel

pub fn try_new(level: u32) -> Result<Self>

Attempts to create a gzip compression level.

Compression levels must be valid (i.e. be acceptable for flate2::Compression).

pub fn compression_level(&self) -> u32

Returns the compression level.