Trait parquet::column::page::PageWriter

pub trait PageWriter {
    fn write_page(&mut self, page: CompressedPage) -> Result<usize>;

    fn write_metadata(&mut self, metadata: &ColumnChunkMetaData) -> Result<()>;

    fn close(&mut self) -> Result<()>;

    fn dictionary_page_offset(&self) -> Option<u64>;

    fn data_page_offset(&self) -> u64;

    fn total_uncompressed_size(&self) -> u64;

    fn total_compressed_size(&self) -> u64;

    fn num_values(&self) -> u32;
}

API for writing pages in a column chunk.

It is reasonable to assume that all pages will be written in the correct order, e.g. dictionary page followed by data pages, or a set of data pages.

Required Methods

fn write_page(&mut self, page: CompressedPage) -> Result<usize>

Writes a page into the output stream/sink. Returns number of bytes written into the sink.

This method is called for every compressed page we write into underlying buffer, either data page or dictionary page.

fn write_metadata(&mut self, metadata: &ColumnChunkMetaData) -> Result<()>

Writes column chunk metadata into the output stream/sink.

This method is called once per lifetime of page writer, normally when we finalise writes in column writer.

fn close(&mut self) -> Result<()>

Closes resources and flushes underlying sink. Page writer should not be used after this method is called.

fn dictionary_page_offset(&self) -> Option<u64>

Returns dictionary page offset in bytes, if set. This is an absolute offset in bytes in the underlying buffer.

fn data_page_offset(&self) -> u64

Returns data page (either v1 or v2) offset in bytes. This is an absolute offset in bytes in the underlying buffer.

fn total_uncompressed_size(&self) -> u64

Returns total uncompressed size in bytes so far.

fn total_compressed_size(&self) -> u64

Returns total compressed size in bytes so far.

fn num_values(&self) -> u32

Returns number of values so far.