Module parquet::column::reader

Contains column reader API.

Structs

ColumnReaderImpl
Typed value reader for a particular primitive column.

Enums

ColumnReader
Column reader for a Parquet type.

Functions

get_column_reader
Gets a specific column reader corresponding to column descriptor col_descr. The column reader will read from pages in col_page_reader.

get_typed_column_reader
Gets a typed column reader for the specific type T, by "up-casting" col_reader of non-generic type to a generic column reader type ColumnReaderImpl.