Module parquet::column::reader::decoder

Structs

ColumnValueDecoderImpl

An implementation of ColumnValueDecoder for [T::T]

DefinitionLevelDecoderImpl

An implementation of DefinitionLevelDecoder for [i16]

RepetitionLevelDecoderImpl

An implementation of RepetitionLevelDecoder for [i16]

Enums

LevelDecoder 🔒

Constants

REPETITION_LEVELS_BATCH_SIZE 🔒

SKIP_BUFFER_SIZE 🔒

Traits

ColumnLevelDecoder

Decodes level data to a LevelsBufferSlice

ColumnValueDecoder

Decodes value data to a ValuesBufferSlice

DefinitionLevelDecoder

LevelsBufferSlice

A slice of levels buffer data that is written to by a ColumnLevelDecoder

RepetitionLevelDecoder

ValuesBufferSlice

A slice of values buffer data that is written to by a ColumnValueDecoder