parquet::column::reader::decoder

Struct RepetitionLevelDecoderImpl

pub struct RepetitionLevelDecoderImpl {
    decoder: Option<LevelDecoder>,
    bit_width: u8,
    buffer: Box<[i16; 1024]>,
    buffer_len: usize,
    buffer_offset: usize,
    has_partial: bool,
}

An implementation of RepetitionLevelDecoder for [i16]

Fields
decoder: Option<LevelDecoder>bit_width: u8buffer: Box<[i16; 1024]>buffer_len: usizebuffer_offset: usizehas_partial: bool
Implementations

impl RepetitionLevelDecoderImpl

pub fn new(max_level: i16) -> Self

fn fill_buf(&mut self) -> Result<()>

fn count_records( &mut self, records_to_read: usize, num_levels: usize, ) -> (bool, usize, usize)

Inspects the buffered repetition levels in the range self.buffer_offset..self.buffer_len and returns the number of “complete” records along with the corresponding number of values

A “complete” record is one where the buffer contains a subsequent repetition level of 0