Trait parquet::column::reader::decoder::LevelsBufferSlice

pub trait LevelsBufferSlice {
    // Required methods
    fn capacity(&self) -> usize;
    fn count_nulls(&self, range: Range<usize>, max_level: i16) -> usize;
}

A slice of levels buffer data that is written to by a ColumnLevelDecoder

Required Methods

fn capacity(&self) -> usize

Returns the capacity of this slice or usize::MAX if no limit

fn count_nulls(&self, range: Range<usize>, max_level: i16) -> usize

Count the number of levels in range not equal to max_level