Trait parquet::column::reader::decoder::RepetitionLevelDecoder

pub trait RepetitionLevelDecoder: ColumnLevelDecoder {
    // Required methods
    fn read_rep_levels(
        &mut self,
        out: &mut Self::Slice,
        range: Range<usize>,
        max_records: usize
    ) -> Result<(usize, usize)>;
    fn skip_rep_levels(
        &mut self,
        num_records: usize,
        num_levels: usize
    ) -> Result<(usize, usize)>;
}

Required Methods

fn read_rep_levels( &mut self, out: &mut Self::Slice, range: Range<usize>, max_records: usize ) -> Result<(usize, usize)>

Read up to max_records of repetition level data into out[range] returning the number of complete records and levels read

range is provided by the caller to allow for types such as default-initialized [T] that only track capacity and not length

A record only ends when the data contains a subsequent repetition level of 0, it is therefore left to the caller to delimit the final record in a column

Panics

Implementations may panic if range overlaps with already written data

fn skip_rep_levels( &mut self, num_records: usize, num_levels: usize ) -> Result<(usize, usize)>

Skips over up to num_levels repetition levels corresponding to num_records records, where a record is delimited by a repetition level of 0

Returns the number of records skipped, and the number of levels skipped

A record only ends when the data contains a subsequent repetition level of 0, it is therefore left to the caller to delimit the final record in a column