parquet::column

Module writer

Contains column writer API.

Structs

ColumnCloseResult

Metadata returned by GenericColumnWriter::close

GenericColumnWriter

Generic column writer for a primitive column.

Enums

ColumnWriter

Column writer for a Parquet type.

LevelDeprecated

Functions

get_column_writer

Gets a specific column writer corresponding to column descriptor descr.

get_typed_column_writer

Gets a typed column writer for the specific type T, by “up-casting” col_writer of non-generic type to a generic column writer type ColumnWriterImpl.

get_typed_column_writer_mut

Similar to get_typed_column_writer but returns a reference.

get_typed_column_writer_ref

Similar to get_typed_column_writer but returns a reference.

Type Aliases

ColumnWriterImpl

Typed column writer for a primitive column.