parquet::column::writer

Module encoder

Structs

ColumnValueEncoderImpl

DataPageValues

The encoded values for a data page, with optional statistics

DictionaryPage

The encoded data for a dictionary page

Traits

ColumnValueEncoder

A generic encoder of ColumnValues to data and dictionary pages used by super::GenericColumnWriter`

ColumnValues

A collection of ParquetValueType encoded by a ColumnValueEncoder

Functions

get_min_max 🔒

replace_zero 🔒