parquet::column::writer
Function get_typed_column_writer

pub fn get_typed_column_writer<T: DataType>(
    col_writer: ColumnWriter<'_>,
) -> ColumnWriterImpl<'_, T>

Gets a typed column writer for the specific type T, by “up-casting” col_writer of non-generic type to a generic column writer type ColumnWriterImpl.

Panics if actual enum value for col_writer does not match the type T.