parquet::column::writer

Function increment_utf8

fn increment_utf8(data: &str) -> Option<Vec<u8>>

Increment the final character in a UTF-8 string in such a way that the returned result is still a valid UTF-8 string. The returned string may be shorter than the input if the last character(s) cannot be incremented (due to overflow or producing invalid code points). Returns None if the string cannot be incremented.

Note that this implementation will not promote an N-byte code point to (N+1) bytes.