Enum parquet::data_type::Decimal

pub enum Decimal {
    Int32 {
        value: [u8; 4],
        precision: i32,
        scale: i32,
    },
    Int64 {
        value: [u8; 8],
        precision: i32,
        scale: i32,
    },
    Bytes {
        value: ByteArray,
        precision: i32,
        scale: i32,
    },
}

Rust representation for Decimal values.

This is not a representation of Parquet physical type, but rather a wrapper for DECIMAL logical type, and serves as container for raw parts of decimal values: unscaled value in bytes, precision and scale.

Variants
Int32

Decimal backed by i32.

Fields of Int32

value: [u8; 4]precision: i32scale: i32
Int64

Decimal backed by i64.

Fields of Int64

value: [u8; 8]precision: i32scale: i32
Bytes

Decimal backed by byte array.

Fields of Bytes

value: ByteArrayprecision: i32scale: i32

Methods

impl Decimal

pub fn from_i32(value: i32, precision: i32, scale: i32) -> Self

Creates new decimal value from i32.

pub fn from_i64(value: i64, precision: i32, scale: i32) -> Self

Creates new decimal value from i64.

pub fn from_bytes(value: ByteArray, precision: i32, scale: i32) -> Self

Creates new decimal value from ByteArray.

pub fn data(&self) -> &[u8]

Returns bytes of unscaled value.

pub fn precision(&self) -> i32

Returns decimal precision.

pub fn scale(&self) -> i32

Returns decimal scale.