parquet::data_type

Struct ByteArray

pub struct ByteArray {
    data: Option<Bytes>,
}

Rust representation for BYTE_ARRAY and FIXED_LEN_BYTE_ARRAY Parquet physical types. Value is backed by a byte buffer.

Fields
data: Option<Bytes>
Implementations

impl ByteArray

pub fn new() -> Self

Creates new byte array with no data set.

pub fn len(&self) -> usize

Gets length of the underlying byte buffer.

pub fn is_empty(&self) -> bool

Checks if the underlying buffer is empty.

pub fn data(&self) -> &[u8]

Returns slice of data.

pub fn set_data(&mut self, data: Bytes)

Set data from another byte buffer.

pub fn slice(&self, start: usize, len: usize) -> Self

Returns ByteArray instance with slice of values for a data.

pub fn as_utf8(&self) -> Result<&str>

Try to convert the byte array to a utf8 slice