Struct parquet::decoding::DictDecoder

pub struct DictDecoder<T: DataType> { /* fields omitted */ }

Dictionary decoder. The dictionary encoding builds a dictionary of values encountered in a given column. The dictionary is be stored in a dictionary page per column chunk. See DictEncoder for more information.

Methods

impl<T: DataType> DictDecoder<T>

pub fn new() -> Self

Creates new dictionary decoder.

pub fn set_dict(&mut self, decoder: Box<Decoder<T>>) -> Result<()>

Decodes and sets values for dictionary using decoder decoder.