Trait parquet::decoding::Decoder

pub trait Decoder<T: DataType> {
    fn set_data(&mut self, data: ByteBufferPtr, num_values: usize) -> Result<()>;

    fn get(&mut self, buffer: &mut [T::T]) -> Result<usize>;

    fn values_left(&self) -> usize;

    fn encoding(&self) -> Encoding;
}

A Parquet decoder for the data type T.

Required Methods

fn set_data(&mut self, data: ByteBufferPtr, num_values: usize) -> Result<()>

Sets the data to decode to be data, which should contain num_values of values to decode.

fn get(&mut self, buffer: &mut [T::T]) -> Result<usize>

Consumes values from this decoder and write the results to buffer. This will try to fill up buffer.

Returns the actual number of values decoded, which should be equal to buffer.len() unless the remaining number of values is less than buffer.len().

fn values_left(&self) -> usize

Returns the number of values left in this decoder stream.

fn encoding(&self) -> Encoding

Returns the encoding for this decoder.