Struct parquet::encoding::PlainEncoder

pub struct PlainEncoder<T: DataType> { /* fields omitted */ }

Plain encoding that supports all types. Values are encoded back to back. The plain encoding is used whenever a more efficient encoding can not be used. It stores the data in the following format:

BOOLEAN - 1 bit per value, 0 is false; 1 is true.

INT32 - 4 bytes per value, stored as little-endian.

INT64 - 8 bytes per value, stored as little-endian.

FLOAT - 4 bytes per value, stored as IEEE little-endian.

DOUBLE - 8 bytes per value, stored as IEEE little-endian.

BYTE_ARRAY - 4 byte length stored as little endian, followed by bytes.

FIXED_LEN_BYTE_ARRAY - just the bytes are stored.

Methods

impl<T: DataType> PlainEncoder<T>

pub fn new(
desc: ColumnDescPtr,
mem_tracker: MemTrackerPtr,
vec: Vec<u8>
) -> Self

Creates new plain encoder.