parquet::file::metadata::options

Struct ParquetMetaDataOptions

pub struct ParquetMetaDataOptions {
    schema_descr: Option<SchemaDescPtr>,
    encoding_stats_as_mask: bool,
    encoding_stats_policy: ParquetStatisticsPolicy,
    column_stats_policy: ParquetStatisticsPolicy,
    size_stats_policy: ParquetStatisticsPolicy,
}

Options that can be set to control what parts of the Parquet file footer metadata will be decoded and made present in the ParquetMetaData returned by ParquetMetaDataReader and ParquetMetaDataPushDecoder.

Fields
schema_descr: Option<SchemaDescPtr>encoding_stats_as_mask: boolencoding_stats_policy: ParquetStatisticsPolicycolumn_stats_policy: ParquetStatisticsPolicysize_stats_policy: ParquetStatisticsPolicy
Implementations

impl ParquetMetaDataOptions

pub fn new() -> Self

Return a new default ParquetMetaDataOptions.

pub fn schema(&self) -> Option<&SchemaDescPtr>

Returns an optional SchemaDescPtr to use when decoding. If this is not None then the schema in the footer will be skipped.

pub fn set_schema(&mut self, val: SchemaDescPtr)

Provide a schema to use when decoding the metadata.

pub fn with_schema(self, val: SchemaDescPtr) -> Self

Call Self::set_schema and return Self for chaining.

pub fn encoding_stats_as_mask(&self) -> bool

Returns whether to present the encoding_stats field of the Parquet ColumnMetaData as a bitmask (defaults to true).

See ColumnChunkMetaData::page_encoding_stats_mask for an explanation of why this might be desirable.

pub fn set_encoding_stats_as_mask(&mut self, val: bool)

Convert encoding_stats from a vector of PageEncodingStats to a bitmask. This can speed up metadata decoding while still enabling some use cases served by the full stats.

Note that if for a given column both this option and skip_encoding_stats are true, the stats will be skipped and not be returned as a mask.

See ColumnChunkMetaData::page_encoding_stats_mask for more information.

pub fn with_encoding_stats_as_mask(self, val: bool) -> Self

Call Self::set_encoding_stats_as_mask and return Self for chaining.

pub fn skip_encoding_stats(&self, col_index: usize) -> bool

Returns whether to skip decoding the encoding_stats in the Parquet ColumnMetaData for the column indexed by col_index.

pub fn set_encoding_stats_policy(&mut self, policy: ParquetStatisticsPolicy)

Sets the decoding policy for encoding_stats in the Parquet ColumnMetaData.

The default policy is to decode all encoding_stats.

This option takes precedence over Self::encoding_stats_as_mask.

pub fn with_encoding_stats_policy(self, policy: ParquetStatisticsPolicy) -> Self

Call Self::set_encoding_stats_policy and return Self for chaining.

pub fn skip_column_stats(&self, col_index: usize) -> bool

Returns whether to skip decoding the statistics in the Parquet ColumnMetaData for the column indexed by col_index.

pub fn set_column_stats_policy(&mut self, policy: ParquetStatisticsPolicy)

Sets the decoding policy for statistics in the Parquet ColumnMetaData.

The default policy is to decode all statistics.

pub fn with_column_stats_policy(self, policy: ParquetStatisticsPolicy) -> Self

Call Self::set_column_stats_policy and return Self for chaining.

pub fn skip_size_stats(&self, col_index: usize) -> bool

Returns whether to skip decoding the size_statistics in the Parquet ColumnMetaData for the column indexed by col_index.

pub fn set_size_stats_policy(&mut self, policy: ParquetStatisticsPolicy)

Sets the decoding policy for size_statistics in the Parquet ColumnMetaData.

The default policy is to decode all size_statistics.

pub fn with_size_stats_policy(self, policy: ParquetStatisticsPolicy) -> Self

Call Self::set_size_stats_policy and return Self for chaining.