parquet::file::metadata::parser

Function parse_column_index

pub(crate) fn parse_column_index(
    metadata: &mut ParquetMetaData,
    column_index_policy: PageIndexPolicy,
    bytes: &Bytes,
    start_offset: u64,
) -> Result<()>

Parses column index from the provided bytes and adds it to the metadata.

Arguments

metadata - The ParquetMetaData to which the parsed column index will be added.

column_index_policy - The policy for handling column index parsing (e.g., Required, Optional, Skip).

bytes - The byte slice containing the column index data.

start_offset - The offset where bytes begin in the file.