Struct parquet::file::metadata::ColumnChunkMetaData

pub struct ColumnChunkMetaData { /* fields omitted */ }

Metadata for a column chunk.

Methods

impl ColumnChunkMetaData

Represents common operations for a column chunk.

pub fn builder(column_descr: ColumnDescPtr) -> ColumnChunkMetaDataBuilder

Returns builder for column chunk metadata.

pub fn file_path(&self) -> Option<&String>

File where the column chunk is stored.

If not set, assumed to belong to the same file as the metadata. This path is relative to the current file.

pub fn file_offset(&self) -> i64

Byte offset in file_path().

pub fn column_type(&self) -> Type

Type of this column. Must be primitive.

pub fn column_path(&self) -> &ColumnPath

Path (or identifier) of this column.

pub fn column_descr(&self) -> &ColumnDescriptor

Descriptor for this column.

pub fn column_descr_ptr(&self) -> ColumnDescPtr

Reference counted clone of descriptor for this column.

pub fn encodings(&self) -> &Vec<Encoding>

All encodings used for this column.

pub fn num_values(&self) -> i64

Total number of values in this column chunk.

pub fn compression(&self) -> Compression

Compression for this column.

pub fn compressed_size(&self) -> i64

Returns the total compressed data size of this column chunk.

pub fn uncompressed_size(&self) -> i64

Returns the total uncompressed data size of this column chunk.

pub fn data_page_offset(&self) -> i64

Returns the offset for the column data.

pub fn has_index_page(&self) -> bool

Returns true if this column chunk contains a index page, false otherwise.

pub fn index_page_offset(&self) -> Option<i64>

Returns the offset for the index page.

pub fn has_dictionary_page(&self) -> bool

Returns true if this column chunk contains a dictionary page, false otherwise.

pub fn dictionary_page_offset(&self) -> Option<i64>

Returns the offset for the dictionary page, if any.

pub fn statistics(&self) -> Option<&Statistics>

Returns statistics that are set for this column chunk, or None if no statistics are available.

pub fn from_thrift(column_descr: ColumnDescPtr, cc: ColumnChunk) -> Result<Self>

Method to convert from Thrift.

pub fn to_thrift(&self) -> ColumnChunk

Method to convert to Thrift.