parquet::file::metadata

Struct ColumnChunkMetaDataBuilder

pub struct ColumnChunkMetaDataBuilder(ColumnChunkMetaData);

Builder for ColumnChunkMetaData

This builder is used to create a new column chunk metadata or modify an existing one.

Example

let column_chunk_metadata = get_column_chunk_metadata();
// create a new builder from existing column chunk metadata
let builder = ColumnChunkMetaDataBuilder::from(column_chunk_metadata);
// clear the statistics:
let column_chunk_metadata: ColumnChunkMetaData = builder
  .clear_statistics()
  .build()
  .unwrap();

Tuple Fields
0: ColumnChunkMetaData
Implementations

impl ColumnChunkMetaDataBuilder

fn new(column_descr: ColumnDescPtr) -> Self

Creates new column chunk metadata builder.

See also ColumnChunkMetaData::builder

pub fn set_encodings(self, encodings: Vec<Encoding>) -> Self

Sets list of encodings for this column chunk.

pub fn set_file_path(self, value: String) -> Self

Sets optional file path for this column chunk.

pub fn set_file_offset(self, value: i64) -> Self

๐Deprecated since 53.0.0: The Parquet specification requires this field to be 0

Sets file offset in bytes.

This field was meant to provide an alternate to storing ColumnMetadata directly in the ColumnChunkMetadata. However, most Parquet readers assume the ColumnMetadata is stored inline and ignore this field.

pub fn set_num_values(self, value: i64) -> Self

Sets number of values.

pub fn set_compression(self, value: Compression) -> Self

Sets compression.

pub fn set_total_compressed_size(self, value: i64) -> Self

Sets total compressed size in bytes.

pub fn set_total_uncompressed_size(self, value: i64) -> Self

Sets total uncompressed size in bytes.

pub fn set_data_page_offset(self, value: i64) -> Self

Sets data page offset in bytes.

pub fn set_dictionary_page_offset(self, value: Option<i64>) -> Self

Sets optional dictionary page offset in bytes.

pub fn set_index_page_offset(self, value: Option<i64>) -> Self

Sets optional index page offset in bytes.

pub fn set_statistics(self, value: Statistics) -> Self

Sets statistics for this column chunk.

pub fn clear_statistics(self) -> Self

Clears the statistics for this column chunk.

pub fn set_page_encoding_stats(self, value: Vec<PageEncodingStats>) -> Self

Sets page encoding stats for this column chunk.

pub fn clear_page_encoding_stats(self) -> Self

Clears the page encoding stats for this column chunk.

pub fn set_bloom_filter_offset(self, value: Option<i64>) -> Self

Sets optional bloom filter offset in bytes.

pub fn set_bloom_filter_length(self, value: Option<i32>) -> Self

Sets optional bloom filter length in bytes.

pub fn set_offset_index_offset(self, value: Option<i64>) -> Self

Sets optional offset index offset in bytes.

pub fn set_offset_index_length(self, value: Option<i32>) -> Self

Sets optional offset index length in bytes.

pub fn set_column_index_offset(self, value: Option<i64>) -> Self

Sets optional column index offset in bytes.

pub fn set_column_index_length(self, value: Option<i32>) -> Self

Sets optional column index length in bytes.

pub fn set_unencoded_byte_array_data_bytes(self, value: Option<i64>) -> Self

Sets optional length of variable length data in bytes.

pub fn set_repetition_level_histogram( self, value: Option<LevelHistogram>, ) -> Self

Sets optional repetition level histogram

pub fn set_definition_level_histogram( self, value: Option<LevelHistogram>, ) -> Self

Sets optional repetition level histogram

pub fn build(self) -> Result<ColumnChunkMetaData>

Builds column chunk metadata.