parquet::file::metadata

Struct ColumnIndexBuilder

pub struct ColumnIndexBuilder {
    null_pages: Vec<bool>,
    min_values: Vec<Vec<u8>>,
    max_values: Vec<Vec<u8>>,
    null_counts: Vec<i64>,
    boundary_order: BoundaryOrder,
    repetition_level_histograms: Option<Vec<i64>>,
    definition_level_histograms: Option<Vec<i64>>,
    valid: bool,
}

Builder for Parquet ColumnIndex, part of the Parquet PageIndex

Fields
null_pages: Vec<bool>min_values: Vec<Vec<u8>>max_values: Vec<Vec<u8>>null_counts: Vec<i64>boundary_order: BoundaryOrderrepetition_level_histograms: Option<Vec<i64>>

contains the concatenation of the histograms of all pages

definition_level_histograms: Option<Vec<i64>>

contains the concatenation of the histograms of all pages

valid: bool

Is the information in the builder valid?

Set to false if any entry in the page doesn’t have statistics for some reason, so statistics for that page won’t be written to the file. This might happen if the page is entirely null, or is a floating point column without any non-nan values e.g. https://github.com/apache/parquet-format/pull/196

Implementations

impl ColumnIndexBuilder

pub fn new() -> Self

Creates a new column index builder.

pub fn append( &mut self, null_page: bool, min_value: Vec<u8>, max_value: Vec<u8>, null_count: i64, )

Append statistics for the next page

pub fn append_histograms( &mut self, repetition_level_histogram: &Option<LevelHistogram>, definition_level_histogram: &Option<LevelHistogram>, )

Append the given page-level histograms to the ColumnIndex histograms. Does nothing if the ColumnIndexBuilder is not in the valid state.

pub fn set_boundary_order(&mut self, boundary_order: BoundaryOrder)

Set the boundary order of the column index

pub fn to_invalid(&mut self)

Mark this column index as invalid

pub fn valid(&self) -> bool

Is the information in the builder valid?

pub fn build_to_thrift(self) -> ColumnIndex

Build and get the thrift metadata of column index

Note: callers should check Self::valid before calling this method