parquet::file::metadata

Struct RowGroupMetaData

pub struct RowGroupMetaData {
    columns: Vec<ColumnChunkMetaData>,
    num_rows: i64,
    sorting_columns: Option<Vec<SortingColumn>>,
    total_byte_size: i64,
    schema_descr: SchemaDescPtr,
    file_offset: Option<i64>,
    ordinal: Option<i16>,
}

Metadata for a row group

Includes ColumnChunkMetaData for each column in the row group, the number of rows the total byte size of the row group, and the SchemaDescriptor for the row group.

Fields
columns: Vec<ColumnChunkMetaData>num_rows: i64sorting_columns: Option<Vec<SortingColumn>>total_byte_size: i64schema_descr: SchemaDescPtrfile_offset: Option<i64>

We canโt infer from file offset of first column since there may empty columns in row group.

ordinal: Option<i16>

Ordinal position of this row group in file

Implementations

impl RowGroupMetaData

pub fn builder(schema_descr: SchemaDescPtr) -> RowGroupMetaDataBuilder

Returns builder for row group metadata.

pub fn num_columns(&self) -> usize

Number of columns in this row group.

pub fn column(&self, i: usize) -> &ColumnChunkMetaData

Returns column chunk metadata for ith column.

pub fn columns(&self) -> &[ColumnChunkMetaData]

Returns slice of column chunk metadata.

pub fn columns_mut(&mut self) -> &mut [ColumnChunkMetaData]

Returns mutable slice of column chunk metadata.

pub fn num_rows(&self) -> i64

Number of rows in this row group.

pub fn sorting_columns(&self) -> Option<&Vec<SortingColumn>>

Returns the sort ordering of the rows in this RowGroup if any

pub fn total_byte_size(&self) -> i64

Total byte size of all uncompressed column data in this row group.

pub fn compressed_size(&self) -> i64

Total size of all compressed column data in this row group.

pub fn schema_descr(&self) -> &SchemaDescriptor

Returns reference to a schema descriptor.

pub fn schema_descr_ptr(&self) -> SchemaDescPtr

Returns reference counted clone of schema descriptor.

pub fn ordinal(&self) -> Option<i16>

Returns ordinal position of this row group in file.

For example if this is the first row group in the file, this will return 0. If this is the second row group in the file, this will return 1.

pub fn file_offset(&self) -> Option<i64>

Returns file offset of this row group in file.

pub fn from_thrift( schema_descr: SchemaDescPtr, rg: RowGroup, ) -> Result<RowGroupMetaData>

Method to convert from Thrift.

pub fn to_thrift(&self) -> RowGroup

Method to convert to Thrift.

pub fn into_builder(self) -> RowGroupMetaDataBuilder

Converts this RowGroupMetaData into a RowGroupMetaDataBuilder