parquet::file::metadata::thrift::encryption

Function parquet_metadata_with_encryption

pub(crate) fn parquet_metadata_with_encryption(
    file_decryption_properties: Option<&Arc<FileDecryptionProperties>>,
    encrypted_footer: bool,
    buf: &[u8],
    options: Option<&ParquetMetaDataOptions>,
) -> Result<ParquetMetaData>

Decodes ParquetMetaData from the provided bytes, handling metadata that may be encrypted.

Typically this is used to decode the metadata from the end of a parquet file. The format of buf is the Thrift compact binary protocol, as specified by the Parquet Spec. Buffer can be encrypted with AES GCM or AES CTR ciphers as specfied in the Parquet Encryption Spec.