parquet::file::metadata

Module thrift_gen

Structs

AesGcmCtrV1 ğŸ”’

AesGcmV1 ğŸ”’

BoundingBox ğŸ”’

ColumnMetaData ğŸ”’

DataPageHeader ğŸ”’

DataPageHeaderV2 ğŸ”’

DictionaryPageHeader ğŸ”’

FileCryptoMetaData ğŸ”’

Crypto metadata for files with encrypted footer

FileMeta ğŸ”’

GeospatialStatistics ğŸ”’

IndexPageHeader ğŸ”’

PageHeader ğŸ”’

PageStatistics ğŸ”’

Statistics for the page header.

SchemaElement ğŸ”’

SizeStatistics ğŸ”’

Statistics ğŸ”’

Enums

EncryptionAlgorithm

Functions

convert_bounding_box ğŸ”’

convert_geo_stats ğŸ”’

convert_stats ğŸ”’

Create a crate::file::statistics::Statistics from a thrift Statistics object.

get_file_decryptor ğŸ”’

parquet_metadata_from_bytes ğŸ”’

Create ParquetMetaData from thrift input. Note that this only decodes the file metadata in the Parquet footer. Page indexes will need to be added later.

parquet_metadata_with_encryption ğŸ”’

Decodes ParquetMetaData from the provided bytes, handling metadata that may be encrypted.

read_column_chunk ğŸ”’

read_row_group ğŸ”’

row_group_from_encrypted_thrift ğŸ”’

serialize_column_meta_data ğŸ”’

write_schema ğŸ”’

write_schema_helper ğŸ”’

Type Aliases

CompressionCodec ğŸ”’