parquet::file::metadata

Type Alias ParquetOffsetIndex

pub type ParquetOffsetIndex = Vec<Vec<OffsetIndexMetaData>>;

OffsetIndexMetaData for each data page of each row group of each column

This structure is the parsed representation of the OffsetIndex from the Parquet file footer, as described in the Parquet PageIndex documentation.

offset_index[row_group_number][column_number] holds the OffsetIndexMetaData corresponding to column column_numberof row group row_group_number.

Aliased Type

pub struct ParquetOffsetIndex { /* private fields */ }