parquet::file::page_index::column_index

Struct ColumnIndex

pub struct ColumnIndex {
    pub(crate) null_pages: Vec<bool>,
    pub(crate) boundary_order: BoundaryOrder,
    pub(crate) null_counts: Option<Vec<i64>>,
    pub(crate) repetition_level_histograms: Option<Vec<i64>>,
    pub(crate) definition_level_histograms: Option<Vec<i64>>,
}

Common bits of the column index

Fields
null_pages: Vec<bool>boundary_order: BoundaryOrdernull_counts: Option<Vec<i64>>repetition_level_histograms: Option<Vec<i64>>definition_level_histograms: Option<Vec<i64>>
Implementations

impl ColumnIndex

pub fn num_pages(&self) -> u64

Returns the number of pages

pub fn null_count(&self, idx: usize) -> Option<i64>

Returns the number of null values in the page indexed by idx

Returns None if no null counts have been set in the index

pub fn repetition_level_histogram(&self, idx: usize) -> Option<&[i64]>

Returns the repetition level histogram for the page indexed by idx

pub fn definition_level_histogram(&self, idx: usize) -> Option<&[i64]>

Returns the definition level histogram for the page indexed by idx

pub fn is_null_page(&self, idx: usize) -> bool

Returns whether the page indexed by idx consists of all null values