parquet::file::page_index::column_index

Trait ColumnIndexIterators

pub trait ColumnIndexIterators {
    type Item;

    // Required methods
    fn min_values_iter(
        colidx: &ColumnIndexMetaData,
    ) -> impl Iterator<Item = Option<Self::Item>>;
    fn max_values_iter(
        colidx: &ColumnIndexMetaData,
    ) -> impl Iterator<Item = Option<Self::Item>>;
}

Provides iterators over min and max values of a ColumnIndexMetaData

Required Associated Types

type Item

Can be one of bool, i32, i64, Int96, f32, f64, ByteArray, or FixedLenByteArray

Required Methods

fn min_values_iter( colidx: &ColumnIndexMetaData, ) -> impl Iterator<Item = Option<Self::Item>>

Return iterator over the min values for the index

fn max_values_iter( colidx: &ColumnIndexMetaData, ) -> impl Iterator<Item = Option<Self::Item>>

Return iterator over the max values for the index

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.