Module parquet::file::page_index::index

Index structures holding decoded ColumnIndex information

Structs

NativeIndex

Stores the PageIndex for each page of a column

PageIndex

PageIndex Statistics for one data page, as described in Column Index.

Enums

Index

Typed statistics for a data page in a column chunk. This structure is obtained from decoding the ColumnIndex in the parquet file and can be used to skip decoding pages while reading the file data.