parquet::file::page_index::index

Struct PageIndex

pub struct PageIndex<T> {
    pub min: Option<T>,
    pub max: Option<T>,
    pub null_count: Option<i64>,
    pub repetition_level_histogram: Option<LevelHistogram>,
    pub definition_level_histogram: Option<LevelHistogram>,
}

Typed statistics for one data page

See NativeIndex for more details

Fields
min: Option<T>

The minimum value, It is None when all values are null

max: Option<T>

The maximum value, It is None when all values are null

null_count: Option<i64>

Null values in the page

repetition_level_histogram: Option<LevelHistogram>

Repetition level histogram for the page

repetition_level_histogram[i] is a count of how many values are at repetition level i. For example, repetition_level_histogram[0] indicates how many rows the page contains.

definition_level_histogram: Option<LevelHistogram>

Definition level histogram for the page

definition_level_histogram[i] is a count of how many values are at definition level i. For example, definition_level_histogram[max_definition_level] indicates how many non-null values are present in the page.

Implementations

impl<T> PageIndex<T>

pub fn min(&self) -> Option<&T>

Returns the minimum value in the page

It is None when all values are null

pub fn max(&self) -> Option<&T>

Returns the maximum value in the page

It is None when all values are null

pub fn null_count(&self) -> Option<i64>

Returns the number of null values in the page

pub fn repetition_level_histogram(&self) -> Option<&LevelHistogram>

Returns the repetition level histogram for the page

pub fn definition_level_histogram(&self) -> Option<&LevelHistogram>

Returns the definition level histogram for the page

impl<T> PageIndex<T>
where T: AsBytes,

pub fn max_bytes(&self) -> Option<&[u8]>

Returns the minimum value in the page as bytes

It is None when all values are null

pub fn min_bytes(&self) -> Option<&[u8]>

Returns the maximum value in the page as bytes

It is None when all values are null