Module parquet::file::page_index::index_reader

Support for reading Index and PageLocation from parquet metadata.

Functions

read_columns_indexes

Reads per-column Index for all columns of a row group by decoding ColumnIndex .

read_pages_locations

Reads per-page PageLocation for all columns of a row group by decoding the OffsetIndex.