parquet::file::page_index::index_reader

Function read_columns_indexes

pub fn read_columns_indexes<R: ChunkReader>(
    reader: &R,
    chunks: &[ColumnChunkMetaData],
) -> Result<Option<Vec<Index>>, ParquetError>

Reads per-column Index for all columns of a row group by decoding ColumnIndex .

Returns a vector of index[column_number].

Returns None if this row group does not contain a ColumnIndex.

See Page Index Documentation for more details.