parquet::file::page_index::index_reader
Function read_offset_indexes

pub fn read_offset_indexes<R: ChunkReader>(
    reader: &R,
    chunks: &[ColumnChunkMetaData],
) -> Result<Vec<OffsetIndexMetaData>, ParquetError>

Reads per-column OffsetIndexMetaData for all columns of a row group by decoding OffsetIndex .

Returns a vector of offset_index[column_number].

Returns an empty vector if this row group does not contain an OffsetIndex.

See Page Index Documentation for more details.