Function parquet::file::page_index::index_reader::read_pages_locations

pub fn read_pages_locations<R: ChunkReader>(
    reader: &R,
    chunks: &[ColumnChunkMetaData]
) -> Result<Vec<Vec<PageLocation>>, ParquetError>

Reads per-page PageLocation for all columns of a row group by decoding the OffsetIndex.

Returns a vector of location[column_number][page_number]

Return an empty vector if this row group does not contain an [OffsetIndex].

See Column Index Documentation for more details.