parquet::file::properties

Struct ColumnProperties

struct ColumnProperties {
    encoding: Option<Encoding>,
    codec: Option<Compression>,
    dictionary_enabled: Option<bool>,
    statistics_enabled: Option<EnabledStatistics>,
    max_statistics_size: Option<usize>,
    bloom_filter_properties: Option<BloomFilterProperties>,
}

Container for column properties that can be changed as part of writer.

If a field is None, it means that no specific value has been set for this column, so some subsequent or default value must be used.

Fields
encoding: Option<Encoding>codec: Option<Compression>dictionary_enabled: Option<bool>statistics_enabled: Option<EnabledStatistics>max_statistics_size: Option<usize>
๐Deprecated since 54.0.0: Unused; will be removed in 56.0.0
bloom_filter_properties: Option<BloomFilterProperties>

bloom filter related properties

Implementations

impl ColumnProperties

fn set_encoding(&mut self, value: Encoding)

Sets encoding for this column.

If dictionary is not enabled, this is treated as a primary encoding for a column. In case when dictionary is enabled for a column, this value is considered to be a fallback encoding.

Panics if user tries to set dictionary encoding here, regardless of dictionary encoding flag being set. Use set_dictionary_enabled method to enable dictionary for a column.

fn set_compression(&mut self, value: Compression)

Sets compression codec for this column.

fn set_dictionary_enabled(&mut self, enabled: bool)

Sets whether or not dictionary encoding is enabled for this column.

fn set_statistics_enabled(&mut self, enabled: EnabledStatistics)

Sets the statistics level for this column.

fn set_max_statistics_size(&mut self, value: usize)

๐Deprecated since 54.0.0: Unused; will be removed in 56.0.0

Sets max size for statistics for this column.

fn set_bloom_filter_enabled(&mut self, value: bool)

If value is true, sets bloom filter properties to default values if not previously set, otherwise it is a no-op. If value is false, resets bloom filter properties to None.

fn set_bloom_filter_fpp(&mut self, value: f64)

Sets the false positive probability for bloom filter for this column, and implicitly enables bloom filter if not previously enabled.

Panics

Panics if the value is not between 0 and 1 exclusive

fn set_bloom_filter_ndv(&mut self, value: u64)

Sets the number of distinct (unique) values for bloom filter for this column, and implicitly enables bloom filter if not previously enabled.

fn encoding(&self) -> Option<Encoding>

Returns optional encoding for this column.

fn compression(&self) -> Option<Compression>

Returns optional compression codec for this column.

fn dictionary_enabled(&self) -> Option<bool>

Returns Some(true) if dictionary encoding is enabled for this column, if disabled then returns Some(false). If result is None, then no setting has been provided.

fn statistics_enabled(&self) -> Option<EnabledStatistics>

Returns optional statistics level requested for this column. If result is None, then no setting has been provided.

fn max_statistics_size(&self) -> Option<usize>

๐Deprecated since 54.0.0: Unused; will be removed in 56.0.0

Returns optional max size in bytes for statistics.

fn bloom_filter_properties(&self) -> Option<&BloomFilterProperties>

Returns the bloom filter properties, or None if not enabled