Struct parquet::file::properties::ReaderPropertiesBuilder

pub struct ReaderPropertiesBuilder {
    codec_options_builder: CodecOptionsBuilder,
    read_bloom_filter: Option<bool>,
}

Builder for parquet file reader configuration. See example on ReaderProperties

Fields
codec_options_builder: CodecOptionsBuilderread_bloom_filter: Option<bool>
Implementations

impl ReaderPropertiesBuilder

Reader properties builder.

fn with_defaults() -> Self

Returns default state of the builder.

pub fn build(self) -> ReaderProperties

Finalizes the configuration and returns immutable reader properties struct.

pub fn set_backward_compatible_lz4(self, value: bool) -> Self

Enable/disable backward compatible LZ4.

If backward compatible LZ4 is enable, on LZ4_HADOOP error it will fallback to the older versions LZ4 algorithms. That is LZ4_FRAME, for backward compatibility with files generated by older versions of this library, and LZ4_RAW, for backward compatibility with files generated by older versions of parquet-cpp.

If backward compatible LZ4 is disabled, on LZ4_HADOOP error it will return the error.

pub fn set_read_bloom_filter(self, value: bool) -> Self

Enable/disable reading bloom filter

If reading bloom filter is enabled, bloom filter will be read from the file. If reading bloom filter is disabled, bloom filter will not be read from the file.

By default bloom filter is set to be read.