Module parquet::file::reader

Contains file reader API and provides methods to access file metadata, row group readers to read individual column chunks, or access record iterator.

Structs

SerializedFileReader
A serialized implementation for Parquet FileReader.

SerializedPageReader
A serialized implementation for Parquet PageReader.

SerializedRowGroupReader
A serialized implementation for Parquet RowGroupReader.

Traits

FileReader
Parquet file reader API. With this, user can get metadata information about the Parquet file, can get reader for each row group, and access record iterator.

Length
Length should return the amount of bytes that implementor contains. It's mainly used to read the metadata, which is at the end of the source.

ParquetReader
ParquetReader is the interface which needs to be fulfilled to be able to parse a parquet source.

RowGroupReader
Parquet row group reader API. With this, user can get metadata information about the row group, as well as readers for each individual column chunk.

TryClone
TryClone tries to clone the type and should maintain the Seek position of the given instance.