parquet::file::serialized_reader

Enum SerializedPageReaderState

enum SerializedPageReaderState {
    Values {
        offset: u64,
        remaining_bytes: u64,
        next_page_header: Option<Box<PageHeader>>,
        page_index: usize,
        require_dictionary: bool,
    },
    Pages {
        page_locations: VecDeque<PageLocation>,
        dictionary_page: Option<PageLocation>,
        total_rows: usize,
        page_index: usize,
    },
}

Variants

Values

Fields

offset: u64

The current byte offset in the reader Note that offset is u64 (i.e., not usize) to support 32-bit architectures such as WASM

remaining_bytes: u64

The length of the chunk in bytes Note that remaining_bytes is u64 (i.e., not usize) to support 32-bit architectures such as WASM

next_page_header: Option<Box<PageHeader>>

page_index: usize

The index of the data page within this column chunk

require_dictionary: bool

Whether the next page is expected to be a dictionary page

Pages

Fields

page_locations: VecDeque<PageLocation>

Remaining page locations

dictionary_page: Option<PageLocation>

Remaining dictionary location if any

total_rows: usize

The total number of rows in this column chunk

page_index: usize

The index of the data page within this column chunk