parquet::file::serialized_reader

Type Alias ReadGroupPredicate

pub type ReadGroupPredicate = Box<dyn FnMut(&RowGroupMetaData, usize) -> bool>;

A predicate for filtering row groups, invoked with the metadata and index of each row group in the file. Only row groups for which the predicate evaluates to true will be scanned

Aliased Type

struct ReadGroupPredicate(/* private fields */);