parquet::file
Module statistics

Contains definitions for working with Parquet statistics.

Though some common methods are available on enum, use pattern match to extract actual min and max values from statistics, see below:

Examples

use parquet::file::statistics::Statistics;

let stats = Statistics::int32(Some(1), Some(10), None, Some(3), true);
assert_eq!(stats.null_count_opt(), Some(3));
assert!(stats.is_min_max_deprecated());
assert!(stats.min_is_exact());
assert!(stats.max_is_exact());

match stats {
    Statistics::Int32(ref typed) => {
        assert_eq!(typed.min_opt(), Some(&1));
        assert_eq!(typed.max_opt(), Some(&10));
    }
    _ => {}
}

Modules

private 🔒

Macros

statistics_enum_func 🔒

statistics_new_func 🔒

Macro to generate methods to create Statistics.

Structs

ValueStatistics

Typed statistics for one column chunk

Enums

Statistics

Strongly typed statistics for a column chunk within a row group.

Functions

from_thrift

Converts Thrift definition into Statistics.

to_thrift

Convert Statistics into Thrift definition.

Type Aliases

TypedStatistics

Typed implementation for Statistics.