Struct parquet::file::statistics::ValueStatistics

pub struct ValueStatistics<T> {
    min: Option<T>,
    max: Option<T>,
    distinct_count: Option<u64>,
    null_count: u64,
    is_max_value_exact: bool,
    is_min_value_exact: bool,
    is_min_max_deprecated: bool,
    is_min_max_backwards_compatible: bool,
}

Statistics for a particular ParquetValueType

Fields
min: Option<T>max: Option<T>distinct_count: Option<u64>null_count: u64is_max_value_exact: boolis_min_value_exact: boolis_min_max_deprecated: bool

If true populate the deprecated min and max fields instead of min_value and max_value

is_min_max_backwards_compatible: bool

If true the statistics are compatible with the deprecated min and max fields. See ValueStatistics::is_min_max_backwards_compatible

Implementations

impl<T: ParquetValueType> ValueStatistics<T>

pub fn new( min: Option<T>, max: Option<T>, distinct_count: Option<u64>, null_count: u64, is_min_max_deprecated: bool, ) -> Self

Creates new typed statistics.

pub fn with_min_is_exact(self, is_min_value_exact: bool) -> Self

Set whether the stored min field represents the exact minimum, or just a bound on the minimum value.

see Self::min_is_exact

pub fn with_max_is_exact(self, is_max_value_exact: bool) -> Self

Set whether the stored max field represents the exact maximum, or just a bound on the maximum value.

see Self::max_is_exact

pub fn with_backwards_compatible_min_max( self, backwards_compatible: bool, ) -> Self

Set whether to write the deprecated min and max fields for compatibility with older parquet writers

This should only be enabled if the field is signed, see Self::is_min_max_backwards_compatible

pub fn min(&self) -> &T

Returns min value of the statistics.

Panics if min value is not set, e.g. all values are null. Use has_min_max_set method to check that.

pub fn max(&self) -> &T

Returns max value of the statistics.

Panics if max value is not set, e.g. all values are null. Use has_min_max_set method to check that.

pub fn min_bytes(&self) -> &[u8]

Returns min value as bytes of the statistics.

Panics if min value is not set, use has_min_max_set method to check if values are set.

pub fn max_bytes(&self) -> &[u8]

Returns max value as bytes of the statistics.

Panics if max value is not set, use has_min_max_set method to check if values are set.

pub fn has_min_max_set(&self) -> bool

Whether or not min and max values are set. Normally both min/max values will be set to Some(value) or None.

pub fn max_is_exact(&self) -> bool

Whether or not max value is set, and is an exact value.

pub fn min_is_exact(&self) -> bool

Whether or not min value is set, and is an exact value.

pub fn distinct_count(&self) -> Option<u64>

Returns optional value of number of distinct values occurring.

pub fn null_count(&self) -> u64

Returns null count.

fn is_min_max_deprecated(&self) -> bool

Returns true if statistics were created using old min/max fields.

pub fn is_min_max_backwards_compatible(&self) -> bool

Old versions of parquet stored statistics in min and max fields, ordered using signed comparison. This resulted in an undefined ordering for unsigned quantities, such as booleans and unsigned integers.

These fields were therefore deprecated in favour of min_value and max_value, which have a type-defined sort order.

However, not all readers have been updated. For backwards compatibility, this method returns true if the statistics within this have a signed sort order, that is compatible with being stored in the deprecated min and max fields