Type Alias parquet::file::statistics::TypedStatistics

pub type TypedStatistics<T> = ValueStatistics<<T as DataType>::T>;

Typed implementation for Statistics.

Aliased Type

struct TypedStatistics<T> {
    min: Option<<T as DataType>::T>,
    max: Option<<T as DataType>::T>,
    distinct_count: Option<u64>,
    null_count: u64,
    is_min_max_deprecated: bool,
    is_min_max_backwards_compatible: bool,
}

Fields
min: Option<<T as DataType>::T>max: Option<<T as DataType>::T>distinct_count: Option<u64>null_count: u64is_min_max_deprecated: bool

If true populate the deprecated min and max fields instead of min_value and max_value

is_min_max_backwards_compatible: bool

If true the statistics are compatible with the deprecated min and max fields. See ValueStatistics::is_min_max_backwards_compatible

Implementations

impl<T: ParquetValueType> ValueStatistics<T>

pub fn new( min: Option<T>, max: Option<T>, distinct_count: Option<u64>, null_count: u64, is_min_max_deprecated: bool ) -> Self

Creates new typed statistics.

pub fn with_backwards_compatible_min_max( self, backwards_compatible: bool ) -> Self

Set whether to write the deprecated min and max fields for compatibility with older parquet writers

This should only be enabled if the field is signed, see Self::is_min_max_backwards_compatible

pub fn min(&self) -> &T

Returns min value of the statistics.

Panics if min value is not set, e.g. all values are null. Use has_min_max_set method to check that.

pub fn max(&self) -> &T

Returns max value of the statistics.

Panics if max value is not set, e.g. all values are null. Use has_min_max_set method to check that.

pub fn min_bytes(&self) -> &[u8]

Returns min value as bytes of the statistics.

Panics if min value is not set, use has_min_max_set method to check if values are set.

pub fn max_bytes(&self) -> &[u8]

Returns max value as bytes of the statistics.

Panics if max value is not set, use has_min_max_set method to check if values are set.

pub fn has_min_max_set(&self) -> bool

Whether or not min and max values are set. Normally both min/max values will be set to Some(value) or None.

fn distinct_count(&self) -> Option<u64>

Returns optional value of number of distinct values occurring.

fn null_count(&self) -> u64

Returns null count.

fn is_min_max_deprecated(&self) -> bool

Returns true if statistics were created using old min/max fields.

pub fn is_min_max_backwards_compatible(&self) -> bool

Old versions of parquet stored statistics in min and max fields, ordered using signed comparison. This resulted in an undefined ordering for unsigned quantities, such as booleans and unsigned integers.

These fields were therefore deprecated in favour of min_value and max_value, which have a type-defined sort order.

However, not all readers have been updated. For backwards compatibility, this method returns true if the statistics within this have a signed sort order, that is compatible with being stored in the deprecated min and max fields