parquet::file

Module writer

Contains file writer API, and provides methods to write row groups and columns by using row group writers and column writers respectively.

Structs

SerializedColumnWriter

A wrapper around a ColumnWriter that invokes a callback on Self::close

SerializedFileWriter

Parquet file writer API. Provides methods to write row groups sequentially.

SerializedPageWriter

A serialized implementation for Parquet PageWriter. Writes and serializes pages and metadata into output stream.

SerializedRowGroupWriter

Parquet row group writer API. Provides methods to access column writers in an iterator-like fashion, order is guaranteed to match the order of schema leaves (column descriptors).

TrackedWrite

A wrapper around a Write that keeps track of the number of bytes that have been written. The given Write is wrapped with a BufWriter to optimize writing performance.

Type Aliases

OnCloseColumnChunk

Callback invoked on closing a column chunk

OnCloseRowGroup

Callback invoked on closing a row group, arguments are: