Struct parquet::file::writer::InMemoryWriteableCursor

pub struct InMemoryWriteableCursor { /* fields omitted */ }

Use this type to write Parquet to memory rather than a file.

Implementations

impl InMemoryWriteableCursor

pub fn into_inner(self) -> Option<Vec<u8>>

Consume this instance and return the underlying buffer as long as there are no other references to this instance.

pub fn data(&self) -> Vec<u8>

Returns a clone of the underlying buffer