Struct parquet::file::writer::SerializedPageWriter

pub struct SerializedPageWriter<T: Write + Position> { /* fields omitted */ }

Serialized page writer.

Writes and serializes pages and metadata into output stream. SerializedPageWriter should not be used after calling close(), metrics (offsets and sizes) are not reset, create a new page writer instead.

Methods

impl<T: Write + Position> SerializedPageWriter<T>

pub fn new(sink: T) -> Self

Creates new page writer.