parquet::file::writer

Struct TrackedWrite

pub struct TrackedWrite<W: Write> {
    inner: BufWriter<W>,
    bytes_written: usize,
}

A wrapper around a Write that keeps track of the number of bytes that have been written. The given Write is wrapped with a BufWriter to optimize writing performance.

Fields
inner: BufWriter<W>bytes_written: usize
Implementations

impl<W: Write> TrackedWrite<W>

pub fn new(inner: W) -> Self

Create a new TrackedWrite from a Write

pub fn bytes_written(&self) -> usize

Returns the number of bytes written to this instance

pub fn inner(&self) -> &W

Returns a reference to the underlying writer.

pub fn inner_mut(&mut self) -> &mut W

Returns a mutable reference to the underlying writer.

It is inadvisable to directly write to the underlying writer, doing so will likely result in data corruption

pub fn into_inner(self) -> Result<W>

Returns the underlying writer.