Trait parquet::file::writer::TryClone

pub trait TryClone: Sized {
    fn try_clone(&self) -> Result<Self>;
}

TryClone tries to clone the type and should maintain the Seek position of the given instance.

Required methods

fn try_clone(&self) -> Result<Self>

Clones the type returning a new instance or an error if it's not possible to clone it.

Loading content...