parquet

Module geospatial

This module provides functionality for working with geospatial data in Parquet file as defined in the spec.

GeospatialStatistics: describes the geospatial statistics for a Parquet column.

BoundingBox: describes the bounding box values for a geospatial column.

GeospatialStatistics describes the geospatial statistics for a Parquet column.

bbox: the BoundingBox for the geospatial data

geospatial_types: the geospatial types for the geospatial data as specified in specification.

Geospatial bounding box describes the spatial extent of the geospatial data within a Parquet row group.

xmin, xmax: the minimum and maximum longitude values

ymin, ymax: the minimum and maximum latitude values

zmin, zmax: (optional) the minimum and maximum elevation values

mmin, mmax: (optional) the minimum and maximum linear reference values

In 2D representation, where x are points:

 ymax +-----------------------+
      |               x       |
      |      x                |
      |              x        |
      |      x                |
 ymin +-----------------------+
      xmin                    xmax

Modules

accumulator

This module provides implementations and traits for building GeospatialStatistics

bounding_box

Bounding box for GEOMETRY or GEOGRAPHY type in the representation of min/max value pair of coordinates from each axis.

statistics

Geospatial statistics for Parquet files.