parquet::geospatial

Module accumulator

This module provides implementations and traits for building GeospatialStatistics

Structs

DefaultGeoStatsAccumulatorFactory

Default accumulator for GeospatialStatistics

VoidGeoStatsAccumulator

A GeoStatsAccumulator that never computes any GeospatialStatistics

Traits

GeoStatsAccumulator

Dynamic GeospatialStatistics accumulator

GeoStatsAccumulatorFactory

Factory for GeospatialStatistics accumulators

Functions

init_geo_stats_accumulator_factory

Initialize the global GeoStatsAccumulatorFactory

try_new_geo_stats_accumulator

Create a new GeoStatsAccumulator instance if descr represents a Geometry or Geography LogicalType