parquet::geospatial::accumulator

Trait GeoStatsAccumulatorFactory

pub trait GeoStatsAccumulatorFactory: Send + Sync {
    // Required method
    fn new_accumulator(
        &self,
        descr: &ColumnDescPtr,
    ) -> Box<dyn GeoStatsAccumulator>;
}

Factory for GeospatialStatistics accumulators

The GeoStatsAccumulatorFactory is a trait implemented by the global factory that generates new instances of a GeoStatsAccumulator when constructing new encoders for a Geometry or Geography logical type.

Required Methods

fn new_accumulator(&self, descr: &ColumnDescPtr) -> Box<dyn GeoStatsAccumulator>

Create a new GeoStatsAccumulator appropriate for the logical type of a given ColumnDescPtr