parquet
Macro experimental

macro_rules! experimental {
    ($(#[$meta:meta])* $vis:vis mod $module:ident) => { ... };
}

Defines a an item with an experimental public API

The module will not be documented, and will only be public if the experimental feature flag is enabled

Experimental components have no stability guarantees