Struct parquet::memory::BufferPtr

pub struct BufferPtr<T> { /* fields omitted */ }

An representation of a slice on a reference-counting and read-only byte array. Sub-slices can be further created from this. The byte array will be released when all slices are dropped.

Methods

impl<T> BufferPtr<T>

pub fn new(v: Vec<T>) -> Self

Creates new buffer from a vector.

pub fn data(&self) -> &[T]

Returns slice of data in this buffer.

pub fn with_range(self, start: usize, len: usize) -> Self

Updates this buffer with new start position and length len.

Range should be within current start position and length.

pub fn with_mem_tracker(self, mc: MemTrackerPtr) -> Self

Adds memory tracker to this buffer.

pub fn start(&self) -> usize

Returns start position of this buffer.

pub fn len(&self) -> usize

Returns length of this buffer

pub fn is_mem_tracked(&self) -> bool

Returns true if this buffer has memory tracker, false otherwise.

pub fn all(&self) -> BufferPtr<T>

Returns a shallow copy of the buffer. Reference counted pointer to the data is copied.

pub fn start_from(&self, start: usize) -> BufferPtr<T>

Returns a shallow copy of the buffer that starts with start position.

pub fn range(&self, start: usize, len: usize) -> BufferPtr<T>

Returns a shallow copy that is a range slice within this buffer.