parquet

Module parquet_thrift

Structs used for encoding and decoding Parquet Thrift objects.

These include:

ThriftCompactInputProtocol: Trait implemented by Thrift decoders.

ThriftSliceInputProtocol: Thrift decoder that takes a slice of bytes as input.

ThriftReadInputProtocol: Thrift decoder that takes a Read as input.

ReadThrift: Trait implemented by serializable objects.

ThriftCompactOutputProtocol: Thrift encoder.

WriteThrift: Trait implemented by serializable objects.

WriteThriftField: Trait implemented by serializable objects that are fields in Thrift structs.

Structs

FieldIdentifier 🔒

Struct used to describe a thrift struct field during decoding.

ListIdentifier 🔒

Struct used to describe a thrift list.

OrderedF64

Wrapper for thrift double fields. This is used to provide an implementation of Eq for floats. This implementation uses IEEE 754 total order.

ThriftCompactOutputProtocol 🔒

Low-level object used to serialize structs to the Thrift compact output protocol.

ThriftReadInputProtocol 🔒

A Thrift input protocol that wraps a Read object.

ThriftSliceInputProtocol 🔒

A high performance Thrift reader that reads from a slice of bytes.

Enums

ElementType 🔒

FieldType 🔒

ThriftProtocolError 🔒

Traits

ReadThrift 🔒

Trait implemented for objects that can be deserialized from a Thrift input stream. Implementations are provided for Thrift primitive types.

ThriftCompactInputProtocol 🔒

Low-level object used to deserialize structs encoded with the Thrift compact protocol.

WriteThrift 🔒

Trait implemented by objects that are to be serialized to a Thrift compact output protocol stream. Implementations are also provided for primitive Thrift types.

WriteThriftField 🔒

Trait implemented by objects that are fields of Thrift structs.

Functions

read_thrift_vec 🔒

Read a Thrift encoded list from the input protocol object.

Type Aliases

ThriftProtocolResult