parquet::parquet_thrift

Struct ThriftCompactOutputProtocol

pub(crate) struct ThriftCompactOutputProtocol<W: Write> {
    writer: W,
}

Low-level object used to serialize structs to the Thrift compact output protocol.

This struct serves as a wrapper around a Write object, to which thrift encoded data will written. The implementation provides functions to write Thrift primitive types, as well as functions used in the encoding of lists and structs. This should rarely be used directly, but is instead intended for use by implementers of WriteThrift and WriteThriftField.

Fields
writer: W
Implementations

impl<W: Write> ThriftCompactOutputProtocol<W>

pub(crate) fn new(writer: W) -> Self

Create a new ThriftCompactOutputProtocol wrapping the byte sink writer.

fn write_byte(&mut self, b: u8) -> Result<()>

Write a single byte to the output stream.

fn write_vlq(&mut self, val: u64) -> Result<()>

Write the given u64 as a ULEB128 encoded varint.

fn write_zig_zag(&mut self, val: i64) -> Result<()>

Write the given i64 as a zig-zag encoded varint.

pub(crate) fn write_field_begin( &mut self, field_type: FieldType, field_id: i16, last_field_id: i16, ) -> Result<()>

Used to mark the start of a Thrift struct field of type field_type. last_field_id is used to compute a delta to the given field_id per the compact protocol spec.

pub(crate) fn write_list_begin( &mut self, element_type: ElementType, len: usize, ) -> Result<()>

Used to indicate the start of a list of element_type elements.

pub(crate) fn write_struct_end(&mut self) -> Result<()>

Used to mark the end of a struct. This must be called after all fields of the struct have been written.

pub(crate) fn write_bytes(&mut self, val: &[u8]) -> Result<()>

Serialize a slice of u8s. This will encode a length, and then write the bytes without further encoding.

pub(crate) fn write_empty_struct( &mut self, field_id: i16, last_field_id: i16, ) -> Result<i16>

Short-cut method used to encode structs that have no fields (often used in Thrift unions). This simply encodes the field id and then immediately writes the end-of-struct marker.

pub(crate) fn write_bool(&mut self, val: bool) -> Result<()>

Write a boolean value.

pub(crate) fn write_i8(&mut self, val: i8) -> Result<()>

Write a zig-zag encoded i8 value.

pub(crate) fn write_i16(&mut self, val: i16) -> Result<()>

Write a zig-zag encoded i16 value.

pub(crate) fn write_i32(&mut self, val: i32) -> Result<()>

Write a zig-zag encoded i32 value.

pub(crate) fn write_i64(&mut self, val: i64) -> Result<()>

Write a zig-zag encoded i64 value.

pub(crate) fn write_double(&mut self, val: f64) -> Result<()>

Write a double value.