parquet::parquet_thrift

Struct ThriftReadInputProtocol

pub(crate) struct ThriftReadInputProtocol<R: Read> {
    reader: R,
}

A Thrift input protocol that wraps a Read object.

Note that this is only intended for use in reading Parquet page headers. This will panic if Thrift binary data is encountered because a slice of that data cannot be returned.

Fields
reader: R
Implementations

impl<R: Read> ThriftReadInputProtocol<R>

pub(crate) fn new(reader: R) -> Self