parquet::parquet_thrift

Trait WriteThrift

pub(crate) trait WriteThrift {
    const ELEMENT_TYPE: ElementType;

    // Required method
    fn write_thrift<W: Write>(
        &self,
        writer: &mut ThriftCompactOutputProtocol<W>,
    ) -> Result<()>;
}

Trait implemented by objects that are to be serialized to a Thrift compact output protocol stream. Implementations are also provided for primitive Thrift types.

Required Associated Constants

const ELEMENT_TYPE: ElementType

The ElementType to use when a list of this object is written.

Required Methods

fn write_thrift<W: Write>( &self, writer: &mut ThriftCompactOutputProtocol<W>, ) -> Result<()>

Serialize this object to the given writer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.