Module parquet::record

Contains record-based API for reading Parquet files.

Modules

reader

Contains implementation of record assembly and converting Parquet types into Rows.

Structs

List

List represents a list which contains an array of elements.

Map

Map represents a map which contains a list of key->value pairs.

Row

Row represents a nested Parquet record.

RowColumnIter

RowColumnIter represents an iterator over column names and values in a Row.

Enums

Field

API to represent a single field in a Row.

Traits

ListAccessor

Trait for type-safe access of an index for a List. Note that the get_XXX methods do not do bound checking.

MapAccessor

Trait for type-safe access of an index for a Map

RecordReader

read up to max_records records from row_group_reader into self The type parameter T is used to work around the rust orphan rule when implementing on types such as Vec<T>.

RecordWriter

write_to_row_group writes from self into row_group_writer schema builds the schema used by row_group_writer The type parameter T is used to work around the rust orphan rule when implementing on types such as &[T].

RowAccessor

Trait for type-safe convenient access to fields within a Row.

RowFormatter

Trait for formatting fields within a Row.