parquet::record::api

Enum Field

pub enum Field {

    Null,
    Bool(bool),
    Byte(i8),
    Short(i16),
    Int(i32),
    Long(i64),
    UByte(u8),
    UShort(u16),
    UInt(u32),
    ULong(u64),
    Float16(f16),
    Float(f32),
    Double(f64),
    Decimal(Decimal),
    Str(String),
    Bytes(ByteArray),
    Date(i32),
    TimestampMillis(i64),
    TimestampMicros(i64),
    Group(Row),
    ListInternal(List),
    MapInternal(Map),

}

API to represent a single field in a Row.

Variants

Null

Null value.

Bool(bool)

Boolean value (true, false).

Byte(i8)

Signed integer INT_8.

Short(i16)

Signed integer INT_16.

Int(i32)

Signed integer INT_32.

Long(i64)

Signed integer INT_64.

UByte(u8)

Unsigned integer UINT_8.

UShort(u16)

Unsigned integer UINT_16.

UInt(u32)

Unsigned integer UINT_32.

ULong(u64)

Unsigned integer UINT_64.

Float16(f16)

IEEE 16-bit floating point value.

Float(f32)

IEEE 32-bit floating point value.

Double(f64)

IEEE 64-bit floating point value.

Decimal(Decimal)

Decimal value.

Str(String)

UTF-8 encoded character string.

Bytes(ByteArray)

General binary value.

Date(i32)

Date without a time of day, stores the number of days from the Unix epoch, 1 January 1970.

TimestampMillis(i64)

Milliseconds from the Unix epoch, 1 January 1970.

TimestampMicros(i64)

Microseconds from the Unix epoch, 1 January 1970.

Group(Row)

Struct, child elements are tuples of field-value pairs.

ListInternal(List)

List of elements.

MapInternal(Map)

List of key-value pairs.

Implementations

impl Field

fn get_type_name(&self) -> &'static str

Get the type name.

pub fn is_primitive(&self) -> bool

Determines if this Row represents a primitive value.

pub fn convert_bool(_descr: &ColumnDescPtr, value: bool) -> Self

Converts Parquet BOOLEAN type with logical type into bool value.

pub fn convert_int32(descr: &ColumnDescPtr, value: i32) -> Self

Converts Parquet INT32 type with converted type into i32 value.

pub fn convert_int64(descr: &ColumnDescPtr, value: i64) -> Self

Converts Parquet INT64 type with converted type into i64 value.

pub fn convert_int96(_descr: &ColumnDescPtr, value: Int96) -> Self

Converts Parquet INT96 (nanosecond timestamps) type and logical type into Timestamp value.

pub fn convert_float(_descr: &ColumnDescPtr, value: f32) -> Self

Converts Parquet FLOAT type with logical type into f32 value.

pub fn convert_double(_descr: &ColumnDescPtr, value: f64) -> Self

Converts Parquet DOUBLE type with converted type into f64 value.

pub fn convert_byte_array( descr: &ColumnDescPtr, value: ByteArray, ) -> Result<Self>

Converts Parquet BYTE_ARRAY type with converted type into a UTF8 string, decimal, float16, or an array of bytes.

pub fn to_json_value(&self) -> Value

Converts the Parquet field into a JSON Value.