parquet::record::api

Trait RowAccessor

pub trait RowAccessor {

    // Required methods
    fn get_bool(&self, i: usize) -> Result<bool>;
    fn get_byte(&self, i: usize) -> Result<i8>;
    fn get_short(&self, i: usize) -> Result<i16>;
    fn get_int(&self, i: usize) -> Result<i32>;
    fn get_long(&self, i: usize) -> Result<i64>;
    fn get_ubyte(&self, i: usize) -> Result<u8>;
    fn get_ushort(&self, i: usize) -> Result<u16>;
    fn get_uint(&self, i: usize) -> Result<u32>;
    fn get_ulong(&self, i: usize) -> Result<u64>;
    fn get_float16(&self, i: usize) -> Result<f16>;
    fn get_float(&self, i: usize) -> Result<f32>;
    fn get_double(&self, i: usize) -> Result<f64>;
    fn get_timestamp_millis(&self, i: usize) -> Result<i64>;
    fn get_timestamp_micros(&self, i: usize) -> Result<i64>;
    fn get_decimal(&self, i: usize) -> Result<&Decimal>;
    fn get_string(&self, i: usize) -> Result<&String>;
    fn get_bytes(&self, i: usize) -> Result<&ByteArray>;
    fn get_group(&self, i: usize) -> Result<&Row>;
    fn get_list(&self, i: usize) -> Result<&List>;
    fn get_map(&self, i: usize) -> Result<&Map>;

}

Trait for type-safe convenient access to fields within a Row.

Required Methods

fn get_bool(&self, i: usize) -> Result<bool>

Try to get a boolean value at the given index.

fn get_byte(&self, i: usize) -> Result<i8>

Try to get a byte value at the given index.

fn get_short(&self, i: usize) -> Result<i16>

Try to get a short value at the given index.

fn get_int(&self, i: usize) -> Result<i32>

Try to get a int value at the given index.

fn get_long(&self, i: usize) -> Result<i64>

Try to get a long value at the given index.

fn get_ubyte(&self, i: usize) -> Result<u8>

Try to get a ubyte value at the given index.

fn get_ushort(&self, i: usize) -> Result<u16>

Try to get a ushort value at the given index.

fn get_uint(&self, i: usize) -> Result<u32>

Try to get a uint value at the given index.

fn get_ulong(&self, i: usize) -> Result<u64>

Try to get a ulong value at the given index.

fn get_float16(&self, i: usize) -> Result<f16>

Try to get a float16 value at the given index.

fn get_float(&self, i: usize) -> Result<f32>

Try to get a float value at the given index.

fn get_double(&self, i: usize) -> Result<f64>

Try to get a double value at the given index.

fn get_timestamp_millis(&self, i: usize) -> Result<i64>

Try to get a date value at the given index.

fn get_timestamp_micros(&self, i: usize) -> Result<i64>

Try to get a date value at the given index.

fn get_decimal(&self, i: usize) -> Result<&Decimal>

Try to get a decimal value at the given index.

fn get_string(&self, i: usize) -> Result<&String>

Try to get a string value at the given index.

fn get_bytes(&self, i: usize) -> Result<&ByteArray>

Try to get a bytes value at the given index.

fn get_group(&self, i: usize) -> Result<&Row>

Try to get a group value at the given index.

fn get_list(&self, i: usize) -> Result<&List>

Try to get a list value at the given index.

fn get_map(&self, i: usize) -> Result<&Map>

Try to get a map value at the given index.