Module parquet::record::reader

Contains implementation of record assembly and converting Parquet types into Rows.

Structs

ReaderIter
Internal iterator of Rows for a reader.

RowIter
Iterator of Rows. It is used either for a single row group to iterate over data in that row group, or an entire file with auto buffering of all row groups.

TreeBuilder
Tree builder for Reader enum. Serves as a container of options for building a reader tree and a builder, and accessing a records iterator RowIter.

Enums

Reader
Reader tree for record assembly