parquet::record::record_reader

Trait RecordReader

pub trait RecordReader<T> {
    // Required method
    fn read_from_row_group(
        &mut self,
        row_group_reader: &mut dyn RowGroupReader,
        num_records: usize,
    ) -> Result<(), ParquetError>;
}

Read up to num_records records from row_group_reader into self.

The type parameter T is used to work around the rust orphan rule when implementing on types such as Vec<T>.

Required Methods

fn read_from_row_group( &mut self, row_group_reader: &mut dyn RowGroupReader, num_records: usize, ) -> Result<(), ParquetError>

Read up to num_records records from row_group_reader into self.