parquet::record

Struct Row

pub struct Row {
    fields: Vec<(String, Field)>,
}

Row represents a nested Parquet record.

Fields
fields: Vec<(String, Field)>
Implementations

impl Row

pub fn new(fields: Vec<(String, Field)>) -> Row

Constructs a Row from the list of fields and returns it.

pub fn len(&self) -> usize

Get the number of fields in this row.

pub fn into_columns(self) -> Vec<(String, Field)>

Move columns data out of the row. Useful to avoid internal data cloning.

Example

use std::fs::File;
use parquet::record::Row;
use parquet::file::reader::{FileReader, SerializedFileReader};

let file = File::open("/path/to/file").unwrap();
let reader = SerializedFileReader::new(file).unwrap();
let row: Row = reader.get_row_iter(None).unwrap().next().unwrap().unwrap();
let columns = row.into_columns();
println!("row columns: {:?}", columns);

pub fn get_column_iter(&self) -> RowColumnIter<'_>

Get an iterator to go through all columns in the row.

Example

use std::fs::File;
use parquet::record::Row;
use parquet::file::reader::{FileReader, SerializedFileReader};

let file = File::open("/path/to/file").unwrap();
let reader = SerializedFileReader::new(file).unwrap();
let row: Row = reader.get_row_iter(None).unwrap().next().unwrap().unwrap();
for (idx, (name, field)) in row.get_column_iter().enumerate() {
    println!("column index: {}, column name: {}, column value: {}", idx, name, field);
}

pub fn to_json_value(&self) -> Value

Converts the row into a JSON object.