parquet::record
Module triplet

Macros

triplet_enum_func 🔒

Macro to generate simple functions that cover all types of triplet iterator. $func is a function of a typed triplet iterator and $token is a either {ref} or {ref, mut}

Structs

TypedTripletIter

Internal typed triplet iterator as a wrapper for column reader (primitive leaf column), provides per-element access.

Enums

TripletIter

High level API wrapper on column reader. Provides per-element access for each primitive column.