Module parquet::schema::types

Contains structs and methods to build Parquet schema and schema descriptors.

Structs

BasicTypeInfo

Basic type info. This contains information such as the name of the type, the repetition level, the logical type and the kind of the type (group, primitive).

ColumnDescriptor

A descriptor for leaf-level primitive columns. This encapsulates information such as definition and repetition levels and is used to re-assemble nested data.

ColumnPath

Represents a path in a nested schema

GroupTypeBuilder

A builder for group types. All attributes are optional except the name. Note that if not specified explicitly, None is used as the repetition of the group, which means it is a root (message) type.

PrimitiveTypeBuilder

A builder for primitive types. All attributes are optional except the name and physical type. Note that if not specified explicitly, Repetition::OPTIONAL is used.

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

Enums

Type

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Functions

from_thrift

Method to convert from Thrift.

to_thrift

Method to convert to Thrift.

Type Aliases

ColumnDescPtr

Type alias for Arc<ColumnDescriptor>.

SchemaDescPtr

Type alias for Arc<SchemaDescriptor>.

TypePtr

Type alias for Arc<Type>.