Enum parquet::schema::types::Type

pub enum Type {
    PrimitiveType {
        basic_info: BasicTypeInfo,
        physical_type: PhysicalType,
        type_length: i32,
        scale: i32,
        precision: i32,
    },
    GroupType {
        basic_info: BasicTypeInfo,
        fields: Vec<TypePtr>,
    },
}

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

Variants
PrimitiveType
Fields of PrimitiveType

basic_info: BasicTypeInfo

physical_type: PhysicalType

type_length: i32

scale: i32

precision: i32

GroupType
Fields of GroupType

basic_info: BasicTypeInfo

fields: Vec<TypePtr>

Methods

impl Type

pub fn primitive_type_builder(
name: &str,
physical_type: PhysicalType
) -> PrimitiveTypeBuilder

Creates primitive type builder with provided field name and physical type.

pub fn group_type_builder(name: &str) -> GroupTypeBuilder

Creates group type builder with provided column name.

pub fn get_basic_info(&self) -> &BasicTypeInfo

Returns BasicTypeInfo information about the type.

pub fn name(&self) -> &str

Returns this type's field name.

pub fn get_fields(&self) -> &[TypePtr]

Gets the fields from this group type. Note that this will panic if called on a non-group type.

pub fn get_physical_type(&self) -> PhysicalType

Gets physical type of this primitive type. Note that this will panic if called on a non-primitive type.

pub fn check_contains(&self, sub_type: &Type) -> bool

Checks if sub_type schema is part of current schema. This method can be used to check if projected columns are part of the root schema.

pub fn is_primitive(&self) -> bool

Returns true if this type is a primitive type, false otherwise.

pub fn is_group(&self) -> bool

Returns true if this type is a group type, false otherwise.

pub fn is_schema(&self) -> bool

Returns true if this type is the top-level schema type (message type).