parquet::schema::types

Struct PrimitiveTypeBuilder

pub struct PrimitiveTypeBuilder<'a> {
    name: &'a str,
    repetition: Repetition,
    physical_type: Type,
    converted_type: ConvertedType,
    logical_type: Option<LogicalType>,
    length: i32,
    precision: i32,
    scale: i32,
    id: Option<i32>,
}

A builder for primitive types. All attributes are optional except the name and physical type. Note that if not specified explicitly, Repetition::OPTIONAL is used.

Fields
name: &'a strrepetition: Repetitionphysical_type: Typeconverted_type: ConvertedTypelogical_type: Option<LogicalType>length: i32precision: i32scale: i32id: Option<i32>
Implementations

impl<'a> PrimitiveTypeBuilder<'a>

pub fn new(name: &'a str, physical_type: PhysicalType) -> Self

Creates new primitive type builder with provided field name and physical type.

pub fn with_repetition(self, repetition: Repetition) -> Self

Sets Repetition for this field and returns itself.

pub fn with_converted_type(self, converted_type: ConvertedType) -> Self

Sets ConvertedType for this field and returns itself.

pub fn with_logical_type(self, logical_type: Option<LogicalType>) -> Self

Sets LogicalType for this field and returns itself. If only the logical type is populated for a primitive type, the converted type will be automatically populated, and can thus be omitted.

pub fn with_length(self, length: i32) -> Self

Sets type length and returns itself. This is only applied to FIXED_LEN_BYTE_ARRAY and INT96 (INTERVAL) types, because they maintain fixed size underlying byte array. By default, value is 0.

pub fn with_precision(self, precision: i32) -> Self

Sets precision for Parquet DECIMAL physical type and returns itself. By default, it equals to 0 and used only for decimal context.

pub fn with_scale(self, scale: i32) -> Self

Sets scale for Parquet DECIMAL physical type and returns itself. By default, it equals to 0 and used only for decimal context.

pub fn with_id(self, id: Option<i32>) -> Self

Sets optional field id and returns itself.

pub fn build(self) -> Result<Type>

Creates a new PrimitiveType instance from the collected attributes. Returns Err in case of any building conditions are not met.

fn check_decimal_precision_scale(&self) -> Result<()>