Trait parquet::schema::visitor::TypeVisitor

pub trait TypeVisitor<R, C> {
    fn visit_primitive(
        &mut self, 
        primitive_type: TypePtr, 
        context: C
    ) -> Result<R>;

    fn visit_struct(&mut self, struct_type: TypePtr, context: C) -> Result<R>;

    fn visit_map(&mut self, map_type: TypePtr, context: C) -> Result<R>;

    fn visit_list_with_item(
        &mut self, 
        list_type: TypePtr, 
        item_type: TypePtr, 
        context: C
    ) -> Result<R>;

    fn visit_list(&mut self, list_type: TypePtr, context: C) -> Result<R> { ... }

    fn dispatch(&mut self, cur_type: TypePtr, context: C) -> Result<R> { ... }
}

A utility trait to help user to traverse against parquet type.

Required methods

fn visit_primitive(&mut self, primitive_type: TypePtr, context: C) -> Result<R>

Called when a primitive type hit.

fn visit_struct(&mut self, struct_type: TypePtr, context: C) -> Result<R>

Called when a struct type hit.

fn visit_map(&mut self, map_type: TypePtr, context: C) -> Result<R>

Called when a map type hit.

fn visit_list_with_item(
&mut self,
list_type: TypePtr,
item_type: TypePtr,
context: C
) -> Result<R>

Called by visit_list.

Provided methods

fn visit_list(&mut self, list_type: TypePtr, context: C) -> Result<R>

Default implementation when visiting a list.

It checks list type definition and calls visit_list_with_item with extracted item type.

To fully understand this algorithm, please refer to parquet doc.

fn dispatch(&mut self, cur_type: TypePtr, context: C) -> Result<R>

A utility method which detects input type and calls corresponding method.