Struct parquet::util::bit_util::BitReader

pub struct BitReader { /* fields omitted */ }

Implementations

impl BitReader

Utility class to read bit/byte stream. This class can read bits or bytes that are either byte aligned or not.

pub fn new(buffer: ByteBufferPtr) -> Self

pub fn reset(&mut self, buffer: ByteBufferPtr)

pub fn get_byte_offset(&self) -> usize

Gets the current byte offset

pub fn get_value<T: FromBytes>(&mut self, num_bits: usize) -> Option<T>

Reads a value of type T and of size num_bits.

Returns None if there’s not enough data available. Some otherwise.

pub fn get_batch<T: FromBytes>(
&mut self,
batch: &mut [T],
num_bits: usize
) -> usize

pub fn get_aligned<T: FromBytes>(&mut self, num_bytes: usize) -> Option<T>

Reads a num_bytes-sized value from this buffer and return it. T needs to be a little-endian native type. The value is assumed to be byte aligned so the bit reader will be advanced to the start of the next byte before reading the value. Returns Some if there’s enough bytes left to form a value of T. Otherwise None.

pub fn get_vlq_int(&mut self) -> Option<i64>

Reads a VLQ encoded (in little endian order) int from the stream. The encoded int must start at the beginning of a byte.

Returns None if there’s not enough bytes in the stream. Some otherwise.

pub fn get_zigzag_vlq_int(&mut self) -> Option<i64>

Reads a zigzag-VLQ encoded (in little endian order) int from the stream Zigzag-VLQ is a variant of VLQ encoding where negative and positive numbers are encoded in a zigzag fashion. See: https://developers.google.com/protocol-buffers/docs/encoding

Note: the encoded int must start at the beginning of a byte.

Returns None if the number of bytes there’s not enough bytes in the stream. Some otherwise.