Trait parquet::util::io::Position

pub trait Position {
    fn pos(&self) -> u64;
}

Position trait returns the current position in the stream. Should be viewed as a lighter version of Seek that does not allow seek operations, and does not require mutable reference for the current position.

Required methods

fn pos(&self) -> u64

Returns position in the stream.

Loading content...