Struct parquet::util::memory::Buffer

pub struct Buffer<T: Clone> { /* fields omitted */ }

A resize-able buffer class with generic member, with optional memory tracker.

Note that a buffer has two attributes: capacity and size: the former is the total number of space reserved for the buffer, while the latter is the actual number of elements. Invariant: capacity >= size. The total allocated bytes for a buffer equals to capacity * sizeof<T>().

Implementations

impl<T: Clone> Buffer<T>

pub fn new() -> Self

Creates new empty buffer.

pub fn with_mem_tracker(self, mc: MemTrackerPtr) -> Self

Adds MemTracker for this buffer.

pub fn data(&self) -> &[T]

Returns slice of data in this buffer.

pub fn set_data(&mut self, new_data: Vec<T>)

Sets data for this buffer.

pub fn resize(&mut self, new_size: usize, init_value: T)

Resizes underlying data in place to a new length new_size.

If new_size is less than current length, data is truncated, otherwise, it is extended to new_size with provided default value init_value.

Memory tracker is also updated, if available.

pub fn clear(&mut self)

Clears underlying data.

pub fn reserve(&mut self, additional_capacity: usize)

Reserves capacity additional_capacity for underlying data vector.

Memory tracker is also updated, if available.

pub fn consume(&mut self) -> BufferPtr<T>

Returns BufferPtr with buffer data. Buffer data is reset.

pub fn push(&mut self, value: T)

Adds value to the buffer.

pub fn capacity(&self) -> usize

Returns current capacity for the buffer.

pub fn size(&self) -> usize

Returns current size for the buffer.

pub fn is_mem_tracked(&self) -> bool

Returns true if memory tracker is added to buffer, false otherwise.

pub fn mem_tracker(&self) -> &MemTrackerPtr

Returns memory tracker associated with this buffer. This may panic, if memory tracker is not set, use method above to check if memory tracker is available.