parquet

Module variant

⚠️ Experimental Support for reading and writing Variants to / from Parquet files ⚠️

This is a 🚧 Work In Progress

Note: Requires the variant_experimental feature of the parquet crate to be enabled.

Features

Variant represents variant value, which can be an object, list, or primitive.

VariantBuilder for building Variant values.

VariantArray for representing a column of Variant values.

[compute] module with functions for manipulating Variants, such as variant_get to extracting a value by path and functions to convert between Variant and JSON.

Example: Writing a Parquet file with Variant column

 // Use the VariantArrayBuilder to build a VariantArray
 let mut builder = VariantArrayBuilder::new(3);
 // row 1: {"name": "Alice"}
 builder.new_object().with_field("name", "Alice").finish();
 let array = builder.build();

// TODO support writing VariantArray directly
// at the moment it panics when trying to downcast to a struct array
// https://github.com/apache/arrow-rs/issues/8296
//  let array: ArrayRef = Arc::new(array);
let array: ArrayRef = Arc::new(array.into_inner());

 // create a RecordBatch with the VariantArray
 let batch = RecordBatch::try_from_iter(vec![("data", array)])?;

 // write the RecordBatch to a Parquet file
 let file = std::fs::File::create("variant.parquet")?;
 let mut writer = ArrowWriter::try_new(file, batch.schema(), None)?;
 writer.write(&batch)?;
 writer.close()?;

Example: Writing JSON with a Parquet file with Variant column

// Create an array of JSON strings, simulating a column of JSON data
// TODO use StringViewArray when available
let input_array = StringArray::from(vec![
  Some(r#"{"name": "Alice", "age": 30}"#),
  Some(r#"{"name": "Bob", "age": 25, "address": {"city": "New York"}}"#),
  None,
  Some("{}"),
]);
let input_array: ArrayRef = Arc::new(input_array);

// Convert the JSON strings to a VariantArray
let array: VariantArray = json_to_variant(&input_array)?;

// TODO support writing VariantArray directly
// at the moment it panics when trying to downcast to a struct array
// https://github.com/apache/arrow-rs/issues/8296
//  let array: ArrayRef = Arc::new(array);
let array: ArrayRef = Arc::new(array.into_inner());

 // create a RecordBatch with the VariantArray
 let batch = RecordBatch::try_from_iter(vec![("data", array)])?;

 // write the RecordBatch to a Parquet file
 let file = std::fs::File::create("variant-json.parquet")?;
 let mut writer = ArrowWriter::try_new(file, batch.schema(), None)?;
 writer.write(&batch)?;
 writer.close()?;

Example: Reading a Parquet file with Variant column

(TODO: add example)

Re-exports

pub use parquet_variant_compute as compute;

Structs

ListBuilder

A builder for creating Variant::List values.

ListState

Internal state for list building

ObjectBuilder

A builder for creating Variant::Object values.

ObjectFieldBuilder

A VariantBuilderExt that inserts a new field into a variant object.

ObjectState

Internal state for object building

ParentState

Tracks information needed to correctly finalize a nested builder.

ReadOnlyMetadataBuilder

A metadata builder that cannot register new field names, and merely returns the field id associated with a known field name. This is useful for variant unshredding operations, where the metadata column is fixed and – per variant shredding spec – already contains all field names from the typed_value column. It is also useful when projecting a subset of fields from a variant object value, since the bytes can be copied across directly without re-encoding their field ids.

ShortString

A Variant ShortString

Uuid

A Universally Unique Identifier (UUID).

ValueBuilder

Wrapper around a Vec<u8> that provides methods for appending primitive values, variant types, and metadata.

VariantBuilder

Top level builder for Variant values

VariantDecimal4

Represents a 4-byte decimal value in the Variant format.

VariantDecimal8

Represents an 8-byte decimal value in the Variant format.

VariantDecimal16

Represents an 16-byte decimal value in the Variant format.

VariantList

Variant Array.

VariantMetadata

Variant Metadata

VariantObject

A Variant Object (struct with named fields).

VariantPath

Represents a qualified path to a potential subfield or index of a variant value.

WritableMetadataBuilder

Builder for constructing metadata for Variant values.

f16

A 16-bit floating point type implementing the IEEE 754-2008 standard binary16 a.k.a “half” format.

Enums

Variant

Represents a Parquet Variant

VariantPathElement

Element of a VariantPath that can be a field name or an index.

Constants

EMPTY_VARIANT_METADATA

The empty metadata dictionary.

EMPTY_VARIANT_METADATA_BYTES

The canonical byte slice corresponding to an empty metadata dictionary.

Traits

BuilderSpecificState

A trait for managing state specific to different builder types.

MetadataBuilder

A trait for building variant metadata dictionaries, to be used in conjunction with a ValueBuilder. The trait provides methods for managing field names and their IDs, as well as rolling back a failed builder operation that might have created new field ids.

VariantBuilderExt

Extends VariantBuilder to help building nested Variants