parquet::variant

Enum Variant

pub enum Variant<'m, 'v> {

    Null,
    Int8(i8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    Date(NaiveDate),
    TimestampMicros(DateTime<Utc>),
    TimestampNtzMicros(NaiveDateTime),
    TimestampNanos(DateTime<Utc>),
    TimestampNtzNanos(NaiveDateTime),
    Decimal4(VariantDecimal4),
    Decimal8(VariantDecimal8),
    Decimal16(VariantDecimal16),
    Float(f32),
    Double(f64),
    BooleanTrue,
    BooleanFalse,
    Binary(&'v [u8]),
    String(&'v str),
    Time(NaiveTime),
    Uuid(Uuid),
    ShortString(ShortString<'v>),
    Object(VariantObject<'m, 'v>),
    List(VariantList<'m, 'v>),

}

Represents a Parquet Variant

The lifetimes 'm and 'v are for metadata and value buffers, respectively.

Background

The specification says:

The Variant Binary Encoding allows representation of semi-structured data (e.g. JSON) in a form that can be efficiently queried by path. The design is intended to allow efficient access to nested data even in the presence of very wide or deep structures.

Another motivation for the representation is that (aside from metadata) each nested Variant value is contiguous and self-contained. For example, in a Variant containing an Array of Variant values, the representation of an inner Variant value, when paired with the metadata of the full variant, is itself a valid Variant.

When stored in Parquet files, Variant fields can also be shredded. Shredding refers to extracting some elements of the variant into separate columns for more efficient extraction/filter pushdown. The Variant Shredding specification describes the details of shredding Variant values as typed Parquet columns.

A Variant represents a type that contains one of:

Primitive: A type and corresponding value (e.g. INT, STRING)

Array: An ordered list of Variant values

Object: An unordered collection of string/Variant pairs (i.e. key/value pairs). An object may not contain duplicate keys.

Encoding

A Variant is encoded with 2 binary values, the value and the metadata. The metadata stores a header and an optional dictionary of field names which are referred to by offset in the value. The value is a binary representation of the actual data, and varies depending on the type.

Design Goals

The design goals of the Rust API are as follows:

Speed / Zero copy access (no cloneing is required)

Safety

Follow standard Rust conventions

Examples:

Creating Variant from Rust Types

use parquet_variant::Variant;
// variants can be directly constructed
let variant = Variant::Int32(123);
// or constructed via `From` impls
assert_eq!(variant, Variant::from(123i32));

Creating Variant from metadata and value

let metadata = [0x01, 0x00, 0x00];
let value = [0x09, 0x48, 0x49];
// parse the header metadata
assert_eq!(
  Variant::from("HI"),
  Variant::new(&metadata, &value)
);

Using Variant values

// variants can be used in match statements like normal enums
match variant {
  Variant::Int32(i) => println!("Integer: {}", i),
  Variant::String(s) => println!("String: {}", s),
  _ => println!("Other variant"),
}

Validation

Every instance of variant is either valid or invalid. depending on whether the underlying bytes are a valid encoding of a variant value (see below).

Instances produced by Self::try_new, Self::try_new_with_metadata, or Self::with_full_validation are fully validated. They always contain valid data, and infallible accesses such as iteration and indexing are panic-free. The validation cost is O(m + v) where m and v are the number of bytes in the metadata and value buffers, respectively.

Instances produced by Self::new and Self::new_with_metadata are unvalidated and so they may contain either valid or invalid data. Infallible accesses to variant objects and arrays, such as iteration and indexing will panic if the underlying bytes are invalid, and fallible alternatives are provided as panic-free alternatives. Self::with_full_validation can also be used to validate an unvalidated instance, if desired.

Unvalidated instances can be constructed in constant time. This can be useful if the caller knows the underlying bytes were already validated previously, or if the caller intends to perform a small number of (fallible) accesses to a large variant value.

A validated variant value guarantees that the associated metadata and all nested object and array values are valid. Primitive variant subtypes are always valid by construction.

Safety

Even an invalid variant value is still safe to use in the Rust sense. Accessing it with infallible methods may cause panics but will never lead to undefined behavior.

Variants

Null

Primitive type: Null

Int8(i8)

Primitive (type_id=1): INT(8, SIGNED)

Int16(i16)

Primitive (type_id=1): INT(16, SIGNED)

Int32(i32)

Primitive (type_id=1): INT(32, SIGNED)

Int64(i64)

Primitive (type_id=1): INT(64, SIGNED)

Date(NaiveDate)

Primitive (type_id=1): DATE

TimestampMicros(DateTime<Utc>)

Primitive (type_id=1): TIMESTAMP(isAdjustedToUTC=true, MICROS)

TimestampNtzMicros(NaiveDateTime)

Primitive (type_id=1): TIMESTAMP(isAdjustedToUTC=false, MICROS)

TimestampNanos(DateTime<Utc>)

Primitive (type_id=1): TIMESTAMP(isAdjustedToUTC=true, NANOS)

TimestampNtzNanos(NaiveDateTime)

Primitive (type_id=1): TIMESTAMP(isAdjustedToUTC=false, NANOS)

Decimal4(VariantDecimal4)

Primitive (type_id=1): DECIMAL(precision, scale) 32-bits

Decimal8(VariantDecimal8)

Primitive (type_id=1): DECIMAL(precision, scale) 64-bits

Decimal16(VariantDecimal16)

Primitive (type_id=1): DECIMAL(precision, scale) 128-bits

Float(f32)

Primitive (type_id=1): FLOAT

Double(f64)

Primitive (type_id=1): DOUBLE

BooleanTrue

Primitive (type_id=1): BOOLEAN (true)

BooleanFalse

Primitive (type_id=1): BOOLEAN (false)

Binary(&'v [u8])

Primitive (type_id=1): BINARY

String(&'v str)

Primitive (type_id=1): STRING

Time(NaiveTime)

Primitive (type_id=1): TIME(isAdjustedToUTC=false, MICROS)

Uuid(Uuid)

Primitive (type_id=1): UUID

ShortString(ShortString<'v>)

Short String (type_id=2): STRING

Object(VariantObject<'m, 'v>)

Object (type_id=3): N/A

List(VariantList<'m, 'v>)

Array (type_id=4): N/A

Implementations

impl<'m, 'v> Variant<'m, 'v>

pub fn try_new( metadata: &'m [u8], value: &'v [u8], ) -> Result<Variant<'m, 'v>, ArrowError>

Attempts to interpret a metadata and value buffer pair as a new Variant.

The instance is fully validated.

Example

use parquet_variant::{Variant, VariantMetadata};
let metadata = [0x01, 0x00, 0x00];
let value = [0x09, 0x48, 0x49];
// parse the header metadata
assert_eq!(
  Variant::from("HI"),
  Variant::try_new(&metadata, &value).unwrap()
);

pub fn new(metadata: &'m [u8], value: &'v [u8]) -> Variant<'m, 'v>

Attempts to interpret a metadata and value buffer pair as a new Variant.

The instance is unvalidated.

Example

use parquet_variant::{Variant, VariantMetadata};
let metadata = [0x01, 0x00, 0x00];
let value = [0x09, 0x48, 0x49];
// parse the header metadata
assert_eq!(
  Variant::from("HI"),
  Variant::new(&metadata, &value)
);

pub fn try_new_with_metadata( metadata: VariantMetadata<'m>, value: &'v [u8], ) -> Result<Variant<'m, 'v>, ArrowError>

Create a new variant with existing metadata.

The instance is fully validated.

Example

let metadata = [0x01, 0x00, 0x00];
let value = [0x09, 0x48, 0x49];
// parse the header metadata first
let metadata = VariantMetadata::new(&metadata);
assert_eq!(
  Variant::from("HI"),
  Variant::try_new_with_metadata(metadata, &value).unwrap()
);

pub fn new_with_metadata( metadata: VariantMetadata<'m>, value: &'v [u8], ) -> Variant<'m, 'v>

Similar to Self::try_new_with_metadata, but unvalidated.

pub fn is_fully_validated(&self) -> bool

True if this variant instance has already been validated.

pub fn with_full_validation(self) -> Result<Variant<'m, 'v>, ArrowError>

Recursively validates this variant value, ensuring that infallible access will not panic due to invalid bytes.

Variant leaf values are always valid by construction, but objects and arrays can be constructed in unvalidated (and potentially invalid) state.

If Self::is_fully_validated is true, validation is a no-op. Otherwise, the cost is O(m + v) where m and v are the sizes of metadata and value buffers, respectively.

pub fn as_null(&self) -> Option<()>

Converts this variant to () if it is null.

Returns Some(()) for null variants, None for non-null variants.

Examples

use parquet_variant::Variant;

// you can extract `()` from a null variant
let v1 = Variant::from(());
assert_eq!(v1.as_null(), Some(()));

// but not from other variants
let v2 = Variant::from("hello!");
assert_eq!(v2.as_null(), None);

pub fn as_boolean(&self) -> Option<bool>

Converts this variant to a bool if possible.

Returns Some(bool) for boolean variants, None for non-boolean variants.

Examples

use parquet_variant::Variant;

// you can extract a bool from the true variant
let v1 = Variant::from(true);
assert_eq!(v1.as_boolean(), Some(true));

// and the false variant
let v2 = Variant::from(false);
assert_eq!(v2.as_boolean(), Some(false));

// but not from other variants
let v3 = Variant::from("hello!");
assert_eq!(v3.as_boolean(), None);

pub fn as_naive_date(&self) -> Option<NaiveDate>

Converts this variant to a NaiveDate if possible.

Returns Some(NaiveDate) for date variants, None for non-date variants.

Examples

use parquet_variant::Variant;
use chrono::NaiveDate;

// you can extract a NaiveDate from a date variant
let date = NaiveDate::from_ymd_opt(2025, 4, 12).unwrap();
let v1 = Variant::from(date);
assert_eq!(v1.as_naive_date(), Some(date));

// but not from other variants
let v2 = Variant::from("hello!");
assert_eq!(v2.as_naive_date(), None);

pub fn as_timestamp_micros(&self) -> Option<DateTime<Utc>>

Converts this variant to a DateTime<Utc> if possible.

Returns Some(DateTime<Utc>) for Variant::TimestampMicros variants, None for other variants.

Examples

use parquet_variant::Variant;
use chrono::NaiveDate;

// you can extract a DateTime<Utc> from a UTC-adjusted variant
let datetime = NaiveDate::from_ymd_opt(2025, 4, 16)
    .unwrap()
    .and_hms_milli_opt(12, 34, 56, 780)
    .unwrap()
    .and_utc();
let v1 = Variant::from(datetime);
assert_eq!(v1.as_timestamp_micros(), Some(datetime));

// but not for other variants.
let datetime_nanos = NaiveDate::from_ymd_opt(2025, 8, 14)
    .unwrap()
    .and_hms_nano_opt(12, 33, 54, 123456789)
    .unwrap()
    .and_utc();
let v2 = Variant::from(datetime_nanos);
assert_eq!(v2.as_timestamp_micros(), None);

pub fn as_timestamp_ntz_micros(&self) -> Option<NaiveDateTime>

Converts this variant to a NaiveDateTime if possible.

Returns Some(NaiveDateTime) for Variant::TimestampNtzMicros variants, None for other variants.

Examples

use parquet_variant::Variant;
use chrono::NaiveDate;

// you can extract a NaiveDateTime from a non-UTC-adjusted variant
let datetime = NaiveDate::from_ymd_opt(2025, 4, 16)
    .unwrap()
    .and_hms_milli_opt(12, 34, 56, 780)
    .unwrap();
let v1 = Variant::from(datetime);
assert_eq!(v1.as_timestamp_ntz_micros(), Some(datetime));

// but not for other variants.
let datetime_nanos = NaiveDate::from_ymd_opt(2025, 8, 14)
    .unwrap()
    .and_hms_nano_opt(12, 33, 54, 123456789)
    .unwrap();
let v2 = Variant::from(datetime_nanos);
assert_eq!(v2.as_timestamp_micros(), None);

pub fn as_timestamp_nanos(&self) -> Option<DateTime<Utc>>

Converts this variant to a DateTime<Utc> if possible.

Returns Some(DateTime<Utc>) for timestamp variants, None for other variants.

Examples

use parquet_variant::Variant;
use chrono::NaiveDate;

// you can extract a DateTime<Utc> from a UTC-adjusted nanosecond-precision variant
let datetime = NaiveDate::from_ymd_opt(2025, 4, 16)
    .unwrap()
    .and_hms_nano_opt(12, 34, 56, 789123456)
    .unwrap()
    .and_utc();
let v1 = Variant::from(datetime);
assert_eq!(v1.as_timestamp_nanos(), Some(datetime));

// or from UTC-adjusted microsecond-precision variant
let datetime_micros = NaiveDate::from_ymd_opt(2025, 8, 14)
    .unwrap()
    .and_hms_milli_opt(12, 33, 54, 123)
    .unwrap()
    .and_utc();
// this will convert to `Variant::TimestampMicros`.
let v2 = Variant::from(datetime_micros);
assert_eq!(v2.as_timestamp_nanos(), Some(datetime_micros));

// but not for other variants.
let v3 = Variant::from("hello!");
assert_eq!(v3.as_timestamp_nanos(), None);

pub fn as_timestamp_ntz_nanos(&self) -> Option<NaiveDateTime>

Converts this variant to a NaiveDateTime if possible.

Returns Some(NaiveDateTime) for timestamp variants, None for other variants.

Examples

use parquet_variant::Variant;
use chrono::NaiveDate;

// you can extract a NaiveDateTime from a non-UTC-adjusted variant
let datetime = NaiveDate::from_ymd_opt(2025, 4, 16)
    .unwrap()
    .and_hms_nano_opt(12, 34, 56, 789123456)
    .unwrap();
let v1 = Variant::from(datetime);
assert_eq!(v1.as_timestamp_ntz_nanos(), Some(datetime));

// or from a microsecond-precision non-UTC-adjusted variant
let datetime_micros = NaiveDate::from_ymd_opt(2025, 8, 14)
    .unwrap()
    .and_hms_milli_opt(12, 33, 54, 123)
    .unwrap();
// this will convert to `Variant::TimestampMicros`.
let v2 = Variant::from(datetime_micros);
assert_eq!(v2.as_timestamp_ntz_nanos(), Some(datetime_micros));

// but not for other variants.
let v3 = Variant::from("hello!");
assert_eq!(v3.as_timestamp_ntz_nanos(), None);

pub fn as_u8_slice(&'v self) -> Option<&'v [u8]>

Converts this variant to a &[u8] if possible.

Returns Some(&[u8]) for binary variants, None for non-binary variants.

Examples

use parquet_variant::Variant;

// you can extract a byte slice from a binary variant
let data = b"hello!";
let v1 = Variant::Binary(data);
assert_eq!(v1.as_u8_slice(), Some(data.as_slice()));

// but not from other variant types
let v2 = Variant::from(123i64);
assert_eq!(v2.as_u8_slice(), None);

pub fn as_string(&'v self) -> Option<&'v str>

Converts this variant to a &str if possible.

Returns Some(&str) for string variants (both regular and short strings), None for non-string variants.

Examples

use parquet_variant::Variant;

// you can extract a string from string variants
let s = "hello!";
let v1 = Variant::from(s);
assert_eq!(v1.as_string(), Some(s));

// but not from other variants
let v2 = Variant::from(123i64);
assert_eq!(v2.as_string(), None);

pub fn as_uuid(&self) -> Option<Uuid>

Converts this variant to a uuid hyphenated string if possible.

Returns Some(String) for UUID variants, None for non-UUID variants.

Examples

use parquet_variant::Variant;

// You can extract a UUID from a UUID variant
let s = uuid::Uuid::parse_str("67e55044-10b1-426f-9247-bb680e5fe0c8").unwrap();
let v1 = Variant::Uuid(s);
assert_eq!(s, v1.as_uuid().unwrap());
assert_eq!("67e55044-10b1-426f-9247-bb680e5fe0c8", v1.as_uuid().unwrap().to_string());

//but not from other variants
let v2 = Variant::from(1234);
assert_eq!(None, v2.as_uuid())

pub fn as_int8(&self) -> Option<i8>

Converts this variant to an i8 if possible.

Returns Some(i8) for integer variants that fit in i8 range, None for non-integer variants or values that would overflow.

Examples

use parquet_variant::Variant;

// you can read an int64 variant into an i8 if it fits
let v1 = Variant::from(123i64);
assert_eq!(v1.as_int8(), Some(123i8));

// but not if it would overflow
let v2 = Variant::from(1234i64);
assert_eq!(v2.as_int8(), None);

// or if the variant cannot be cast into an integer
let v3 = Variant::from("hello!");
assert_eq!(v3.as_int8(), None);

pub fn as_int16(&self) -> Option<i16>

Converts this variant to an i16 if possible.

Returns Some(i16) for integer variants that fit in i16 range, None for non-integer variants or values that would overflow.

Examples

use parquet_variant::Variant;

// you can read an int64 variant into an i16 if it fits
let v1 = Variant::from(123i64);
assert_eq!(v1.as_int16(), Some(123i16));

// but not if it would overflow
let v2 = Variant::from(123456i64);
assert_eq!(v2.as_int16(), None);

// or if the variant cannot be cast into an integer
let v3 = Variant::from("hello!");
assert_eq!(v3.as_int16(), None);

pub fn as_int32(&self) -> Option<i32>

Converts this variant to an i32 if possible.

Returns Some(i32) for integer variants that fit in i32 range, None for non-integer variants or values that would overflow.

Examples

use parquet_variant::Variant;

// you can read an int64 variant into an i32 if it fits
let v1 = Variant::from(123i64);
assert_eq!(v1.as_int32(), Some(123i32));

// but not if it would overflow
let v2 = Variant::from(12345678901i64);
assert_eq!(v2.as_int32(), None);

// or if the variant cannot be cast into an integer
let v3 = Variant::from("hello!");
assert_eq!(v3.as_int32(), None);

pub fn as_int64(&self) -> Option<i64>

Converts this variant to an i64 if possible.

Returns Some(i64) for integer variants that fit in i64 range, None for non-integer variants or values that would overflow.

Examples

use parquet_variant::Variant;

// you can read an int64 variant into an i64
let v1 = Variant::from(123i64);
assert_eq!(v1.as_int64(), Some(123i64));

// but not a variant that cannot be cast into an integer
let v2 = Variant::from("hello!");
assert_eq!(v2.as_int64(), None);

pub fn as_u8(&self) -> Option<u8>

Converts this variant to a u8 if possible.

Returns Some(u8) for integer variants that fit in u8 None for non-integer variants or values that would overflow.

Examples

 use parquet_variant::{Variant, VariantDecimal4};

 // you can read an int64 variant into an u8
 let v1 = Variant::from(123i64);
 assert_eq!(v1.as_u8(), Some(123u8));

 // or a Decimal4 with scale 0 into u8
 let d = VariantDecimal4::try_new(26, 0).unwrap();
 let v2 = Variant::from(d);
 assert_eq!(v2.as_u8(), Some(26u8));

 // but not a variant that can't fit into the range
 let v3 = Variant::from(-1);
 assert_eq!(v3.as_u8(), None);

 // not a variant that decimal with scale not equal to zero
 let d = VariantDecimal4::try_new(1, 2).unwrap();
 let v4 = Variant::from(d);
 assert_eq!(v4.as_u8(), None);

 // or not a variant that cannot be cast into an integer
 let v5 = Variant::from("hello!");
 assert_eq!(v5.as_u8(), None);

pub fn as_u16(&self) -> Option<u16>

Converts this variant to an u16 if possible.

Returns Some(u16) for integer variants that fit in u16 None for non-integer variants or values that would overflow.

Examples

 use parquet_variant::{Variant, VariantDecimal4};

 // you can read an int64 variant into an u16
 let v1 = Variant::from(123i64);
 assert_eq!(v1.as_u16(), Some(123u16));

 // or a Decimal4 with scale 0 into u8
 let d = VariantDecimal4::try_new(u16::MAX as i32, 0).unwrap();
 let v2 = Variant::from(d);
 assert_eq!(v2.as_u16(), Some(u16::MAX));

 // but not a variant that can't fit into the range
 let v3 = Variant::from(-1);
 assert_eq!(v3.as_u16(), None);

 // not a variant that decimal with scale not equal to zero
 let d = VariantDecimal4::try_new(1, 2).unwrap();
 let v4 = Variant::from(d);
 assert_eq!(v4.as_u16(), None);

 // or not a variant that cannot be cast into an integer
 let v5 = Variant::from("hello!");
 assert_eq!(v5.as_u16(), None);

pub fn as_u32(&self) -> Option<u32>

Converts this variant to an u32 if possible.

Returns Some(u32) for integer variants that fit in u32 None for non-integer variants or values that would overflow.

Examples

 use parquet_variant::{Variant, VariantDecimal8};

 // you can read an int64 variant into an u32
 let v1 = Variant::from(123i64);
 assert_eq!(v1.as_u32(), Some(123u32));

 // or a Decimal4 with scale 0 into u8
 let d = VariantDecimal8::try_new(u32::MAX as i64, 0).unwrap();
 let v2 = Variant::from(d);
 assert_eq!(v2.as_u32(), Some(u32::MAX));

 // but not a variant that can't fit into the range
 let v3 = Variant::from(-1);
 assert_eq!(v3.as_u32(), None);

 // not a variant that decimal with scale not equal to zero
 let d = VariantDecimal8::try_new(1, 2).unwrap();
 let v4 = Variant::from(d);
 assert_eq!(v4.as_u32(), None);

 // or not a variant that cannot be cast into an integer
 let v5 = Variant::from("hello!");
 assert_eq!(v5.as_u32(), None);

pub fn as_u64(&self) -> Option<u64>

Converts this variant to an u64 if possible.

Returns Some(u64) for integer variants that fit in u64 None for non-integer variants or values that would overflow.

Examples

 use parquet_variant::{Variant, VariantDecimal16};

 // you can read an int64 variant into an u64
 let v1 = Variant::from(123i64);
 assert_eq!(v1.as_u64(), Some(123u64));

 // or a Decimal16 with scale 0 into u8
 let d = VariantDecimal16::try_new(u64::MAX as i128, 0).unwrap();
 let v2 = Variant::from(d);
 assert_eq!(v2.as_u64(), Some(u64::MAX));

 // but not a variant that can't fit into the range
 let v3 = Variant::from(-1);
 assert_eq!(v3.as_u64(), None);

 // not a variant that decimal with scale not equal to zero
let d = VariantDecimal16::try_new(1, 2).unwrap();
 let v4 = Variant::from(d);
 assert_eq!(v4.as_u64(), None);

 // or not a variant that cannot be cast into an integer
 let v5 = Variant::from("hello!");
 assert_eq!(v5.as_u64(), None);

pub fn as_decimal4(&self) -> Option<VariantDecimal4>

Converts this variant to tuple with a 4-byte unscaled value if possible.

Returns Some((i32, u8)) for decimal variants where the unscaled value fits in i32 range, None for non-decimal variants or decimal values that would overflow.

Examples

use parquet_variant::{Variant, VariantDecimal4, VariantDecimal8};

// you can extract decimal parts from smaller or equally-sized decimal variants
let v1 = Variant::from(VariantDecimal4::try_new(1234_i32, 2).unwrap());
assert_eq!(v1.as_decimal4(), VariantDecimal4::try_new(1234_i32, 2).ok());

// and from larger decimal variants if they fit
let v2 = Variant::from(VariantDecimal8::try_new(1234_i64, 2).unwrap());
assert_eq!(v2.as_decimal4(), VariantDecimal4::try_new(1234_i32, 2).ok());

// but not if the value would overflow i32
let v3 = Variant::from(VariantDecimal8::try_new(12345678901i64, 2).unwrap());
assert_eq!(v3.as_decimal4(), None);

// or if the variant is not a decimal
let v4 = Variant::from("hello!");
assert_eq!(v4.as_decimal4(), None);

pub fn as_decimal8(&self) -> Option<VariantDecimal8>

Converts this variant to tuple with an 8-byte unscaled value if possible.

Returns Some((i64, u8)) for decimal variants where the unscaled value fits in i64 range, None for non-decimal variants or decimal values that would overflow.

Examples

use parquet_variant::{Variant, VariantDecimal4, VariantDecimal8, VariantDecimal16};

// you can extract decimal parts from smaller or equally-sized decimal variants
let v1 = Variant::from(VariantDecimal4::try_new(1234_i32, 2).unwrap());
assert_eq!(v1.as_decimal8(), VariantDecimal8::try_new(1234_i64, 2).ok());

// and from larger decimal variants if they fit
let v2 = Variant::from(VariantDecimal16::try_new(1234_i128, 2).unwrap());
assert_eq!(v2.as_decimal8(), VariantDecimal8::try_new(1234_i64, 2).ok());

// but not if the value would overflow i64
let v3 = Variant::from(VariantDecimal16::try_new(2e19 as i128, 2).unwrap());
assert_eq!(v3.as_decimal8(), None);

// or if the variant is not a decimal
let v4 = Variant::from("hello!");
assert_eq!(v4.as_decimal8(), None);

pub fn as_decimal16(&self) -> Option<VariantDecimal16>

Converts this variant to tuple with a 16-byte unscaled value if possible.

Returns Some((i128, u8)) for decimal variants where the unscaled value fits in i128 range, None for non-decimal variants or decimal values that would overflow.

Examples

use parquet_variant::{Variant, VariantDecimal16, VariantDecimal4};

// you can extract decimal parts from smaller or equally-sized decimal variants
let v1 = Variant::from(VariantDecimal4::try_new(1234_i32, 2).unwrap());
assert_eq!(v1.as_decimal16(), VariantDecimal16::try_new(1234_i128, 2).ok());

// but not if the variant is not a decimal
let v2 = Variant::from("hello!");
assert_eq!(v2.as_decimal16(), None);

pub fn as_f16(&self) -> Option<f16>

Converts this variant to an f16 if possible.

Returns Some(f16) for floating point values, and integers with up to 11 bits of precision. None otherwise.

Example

use parquet_variant::Variant;
use half::f16;

// you can extract an f16 from a float variant
let v1 = Variant::from(std::f32::consts::PI);
assert_eq!(v1.as_f16(), Some(f16::from_f32(std::f32::consts::PI)));

// and from a double variant (with loss of precision to nearest f16)
let v2 = Variant::from(std::f64::consts::PI);
assert_eq!(v2.as_f16(), Some(f16::from_f64(std::f64::consts::PI)));

// and from integers with no more than 11 bits of precision
let v3 = Variant::from(2047);
assert_eq!(v3.as_f16(), Some(f16::from_f32(2047.0)));

// but not from other variants
let v4 = Variant::from("hello!");
assert_eq!(v4.as_f16(), None);

pub fn as_f32(&self) -> Option<f32>

Converts this variant to an f32 if possible.

Returns Some(f32) for floating point values, and integer values with up to 24 bits of precision. None otherwise.

Examples

use parquet_variant::Variant;

// you can extract an f32 from a float variant
let v1 = Variant::from(std::f32::consts::PI);
assert_eq!(v1.as_f32(), Some(std::f32::consts::PI));

// and from a double variant (with loss of precision to nearest f32)
let v2 = Variant::from(std::f64::consts::PI);
assert_eq!(v2.as_f32(), Some(std::f32::consts::PI));

// and from integers with no more than 24 bits of precision
let v3 = Variant::from(16777215i64);
assert_eq!(v3.as_f32(), Some(16777215.0));

// but not from other variants
let v4 = Variant::from("hello!");
assert_eq!(v4.as_f32(), None);

pub fn as_f64(&self) -> Option<f64>

Converts this variant to an f64 if possible.

Returns Some(f64) for floating point values, and integer values with up to 53 bits of precision. None otherwise.

Examples

use parquet_variant::Variant;

// you can extract an f64 from a float variant
let v1 = Variant::from(std::f32::consts::PI);
assert_eq!(v1.as_f64(), Some(std::f32::consts::PI as f64));

// and from a double variant
let v2 = Variant::from(std::f64::consts::PI);
assert_eq!(v2.as_f64(), Some(std::f64::consts::PI));

// and from integers with no more than 53 bits of precision
let v3 = Variant::from(9007199254740991i64);
assert_eq!(v3.as_f64(), Some(9007199254740991.0));

// but not from other variants
let v4 = Variant::from("hello!");
assert_eq!(v4.as_f64(), None);

pub fn as_object(&'m self) -> Option<&'m VariantObject<'m, 'v>>

Converts this variant to an Object if it is an VariantObject.

Returns Some(&VariantObject) for object variants, None for non-object variants.

See Self::get_path to dynamically traverse objects

Examples

// object that is {"name": "John"}
 let variant = Variant::new(&metadata, &value);
// use the `as_object` method to access the object
let obj = variant.as_object().expect("variant should be an object");
assert_eq!(obj.get("name"), Some(Variant::from("John")));

pub fn get_object_field(&self, field_name: &str) -> Option<Variant<'m, 'v>>

If this is an object and the requested field name exists, retrieves the corresponding field value. Otherwise, returns None.

This is shorthand for Self::as_object followed by VariantObject::get.

Examples

// object that is {"name": "John"}
 let variant = Variant::new(&metadata, &value);
// use the `get_object_field` method to access the object
let obj = variant.get_object_field("name");
assert_eq!(obj, Some(Variant::from("John")));
let obj = variant.get_object_field("foo");
assert!(obj.is_none());

pub fn as_list(&'m self) -> Option<&'m VariantList<'m, 'v>>

Converts this variant to a List if it is a VariantList.

Returns Some(&VariantList) for list variants, None for non-list variants.

See Self::get_path to dynamically traverse lists

Examples

// list that is ["John", "Doe"]
let variant = Variant::new(&metadata, &value);
// use the `as_list` method to access the list
let list = variant.as_list().expect("variant should be a list");
assert_eq!(list.len(), 2);
assert_eq!(list.get(0).unwrap(), Variant::from("John"));
assert_eq!(list.get(1).unwrap(), Variant::from("Doe"));

pub fn as_time_utc(&'m self) -> Option<NaiveTime>

Converts this variant to a NaiveTime if possible.

Returns Some(NaiveTime) for Variant::Time, None for non-Time variants.

Example

use chrono::NaiveTime;
use parquet_variant::Variant;

// you can extract a `NaiveTime` from a `Variant::Time`
let time = NaiveTime::from_hms_micro_opt(1, 2, 3, 4).unwrap();
let v1 = Variant::from(time);
assert_eq!(Some(time), v1.as_time_utc());

// but not from other variants.
let v2 = Variant::from("Hello");
assert_eq!(None, v2.as_time_utc());

pub fn get_list_element(&self, index: usize) -> Option<Variant<'m, 'v>>

If this is a list and the requested index is in bounds, retrieves the corresponding element. Otherwise, returns None.

This is shorthand for Self::as_list followed by VariantList::get.

Examples

// list that is ["John", "Doe"]
let variant = Variant::new(&metadata, &value);
// use the `get_list_element` method to access the list
assert_eq!(variant.get_list_element(0), Some(Variant::from("John")));
assert_eq!(variant.get_list_element(1), Some(Variant::from("Doe")));
assert!(variant.get_list_element(2).is_none());

pub fn metadata(&self) -> &VariantMetadata<'m>

Return the metadata dictionary associated with this variant value.

pub fn get_path(&self, path: &VariantPath<'_>) -> Option<Variant<'_, '_>>

Return a new Variant with the path followed.

If the path is not found, None is returned.

Example

// given a variant like `{"foo": ["bar", "baz"]}`
let variant = Variant::new(&metadata, &value);
// Accessing a non existent path returns None
assert_eq!(variant.get_path(&VariantPath::from("non_existent")), None);
// Access obj["foo"]
let path = VariantPath::from("foo");
let foo = variant.get_path(&path).expect("field `foo` should exist");
assert!(foo.as_list().is_some(), "field `foo` should be a list");
// Access foo[0]
let path = VariantPath::from(0);
let bar = foo.get_path(&path).expect("element 0 should exist");
// bar is a string
assert_eq!(bar.as_string(), Some("bar"));
// You can also access nested paths
let path = VariantPath::from("foo").join(0);
assert_eq!(variant.get_path(&path).unwrap(), bar);