parquet::variant

Struct ListBuilder

pub struct ListBuilder<'a, S>
where
    S: BuilderSpecificState,
{
    parent_state: ParentState<'a, S>,
    offsets: Vec<usize>,
    validate_unique_fields: bool,
}

A builder for creating Variant::List values.

See the examples on VariantBuilder for usage.

Fields
parent_state: ParentState<'a, S>offsets: Vec<usize>validate_unique_fields: bool
Implementations

impl<'a, S> ListBuilder<'a, S>
where S: BuilderSpecificState,

pub fn new( parent_state: ParentState<'a, S>, validate_unique_fields: bool, ) -> ListBuilder<'a, S>

Creates a new list builder, nested on top of the given parent state.

pub fn with_validate_unique_fields( self, validate_unique_fields: bool, ) -> ListBuilder<'a, S>

Enables unique field key validation for objects created within this list.

Propagates the validation flag to any ObjectBuilders created using ListBuilder::new_object.

pub fn new_object(&mut self) -> ObjectBuilder<'_, ListState<'_>>

Returns an object builder that can be used to append a new (nested) object to this list.

WARNING: The builder will have no effect unless/until ObjectBuilder::finish is called.

pub fn new_list(&mut self) -> ListBuilder<'_, ListState<'_>>

Returns a list builder that can be used to append a new (nested) list to this list.

WARNING: The builder will have no effect unless/until ListBuilder::finish is called.

pub fn append_value<'m, 'd, T>(&mut self, value: T)
where T: Into<Variant<'m, 'd>>,

Appends a variant to the list.

Panics

This method will panic if the variant contains duplicate field names in objects when validation is enabled. For a fallible version, use ListBuilder::try_append_value.

pub fn try_append_value<'m, 'd, T>( &mut self, value: T, ) -> Result<(), ArrowError>
where T: Into<Variant<'m, 'd>>,

Appends a new primitive value to this list

pub fn append_value_bytes<'m, 'd>(&mut self, value: impl Into<Variant<'m, 'd>>)

Appends a variant value to this list by copying raw bytes when possible.

For objects and lists, this directly copies their underlying byte representation instead of performing a logical copy. For other variant types, this falls back to the standard append behavior.

The caller must ensure that the metadata dictionary is already built and correct for any objects or lists being appended.

pub fn with_value<'m, 'd, T>(self, value: T) -> ListBuilder<'a, S>
where T: Into<Variant<'m, 'd>>,

Builder-style API for appending a value to the list and returning self to enable method chaining.

Panics

This method will panic if the variant contains duplicate field names in objects when validation is enabled. For a fallible version, use ListBuilder::try_with_value.

pub fn try_with_value<'m, 'd, T>( self, value: T, ) -> Result<ListBuilder<'a, S>, ArrowError>
where T: Into<Variant<'m, 'd>>,

Builder-style API for appending a value to the list and returns self for method chaining.

This is the fallible version of ListBuilder::with_value.

pub fn finish(self)

Finalizes this list and appends it to its parent, which otherwise remains unmodified.