Crate parquet

Re-exports

pub use self::util::memory;

Modules

arrow

Apache Arrow is a cross-language development platform for in-memory data.

basic

Contains Rust mappings for Thrift definition. Refer to parquet.thrift file to see raw definitions.

column

Low level column reader and writer APIs.

compression

Contains codec interface and supported codec implementations.

data_type

Data types that connect Parquet physical types with their Rust-specific representations.

decoding

Contains all supported decoders for Parquet.

encoding

Contains all supported encoders for Parquet.

errors

Common Parquet errors and macros.

file

Main entrypoint for working with Parquet API.

record

Contains record-based API for reading Parquet files.

schema

Parquet schema definitions and methods to print and parse schema.

util