parquet::arrow

Module arrow_reader

Contains reader which reads parquet data into arrow [RecordBatch]

Modules

filter 🔒

metrics

ArrowReaderMetrics for collecting metrics about the Arrow reader

read_plan 🔒

ReadPlan and ReadPlanBuilder for determining which rows to read from a Parquet file

selection 🔒

statistics

StatisticsConverter to convert statistics in parquet format to arrow [ArrayRef].

Structs

ArrowPredicateFn

An ArrowPredicate created from an FnMut and a ProjectionMask

ArrowReaderBuilder

Builder for constructing Parquet readers that decode into Apache Arrow arrays.

ArrowReaderMetadata

The metadata necessary to construct a ArrowReaderBuilder

ArrowReaderOptions

Options that control how ParquetMetaData is read when constructing an Arrow reader.

ParquetRecordBatchReader

Reads Parquet data as Arrow [RecordBatch]es

ReadPlan

A plan reading specific rows from a Parquet Row Group.

ReadPlanBuilder

A builder for ReadPlan

ReaderPageIterator 🔒

ReaderRowGroups 🔒

RowFilter

Filter applied during the parquet read process

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

Enums

RowSelectionCursor

Cursor for iterating a RowSelection during execution within a ReadPlan.

RowSelectionPolicy

Policy for picking a strategy to materialise RowSelection during execution.

Traits

ArrowPredicate

A predicate operating on [RecordBatch]

RowGroups

Interface for reading data pages from the columns of one or more RowGroups.

Type Aliases

ParquetRecordBatchReaderBuilder

Creates ParquetRecordBatchReader for reading Parquet files into Arrow [RecordBatch]es