parquet::arrow

Module arrow_reader

Contains reader which reads parquet data into arrow [RecordBatch]

Modules

filter 🔒

read_plan 🔒

ReadPlan and ReadPlanBuilder for determining which rows to read from a Parquet file

selection 🔒

statistics

StatisticsConverter to convert statistics in parquet format to arrow [ArrayRef].

Structs

ArrowPredicateFn

An ArrowPredicate created from an FnMut and a ProjectionMask

ArrowReaderBuilder

Builder for constructing Parquet readers that decode into Apache Arrow arrays.

ArrowReaderMetadata

The metadata necessary to construct a ArrowReaderBuilder

ArrowReaderOptions

Options that control how metadata is read for a parquet file

ParquetRecordBatchReader

An Iterator<Item = ArrowResult<RecordBatch>> that yields [RecordBatch] read from a parquet data source

ReaderPageIterator 🔒

ReaderRowGroups 🔒

RowFilter

Filter applied during the parquet read process

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

Traits

ArrowPredicate

A predicate operating on [RecordBatch]

RowGroups

A collection of row groups

Type Aliases

ParquetRecordBatchReaderBuilder

A synchronous builder used to construct ParquetRecordBatchReader for a file