Module parquet::arrow::arrow_reader

Contains reader which reads parquet data into arrow [RecordBatch]

Modules

filter 🔒

selection 🔒

Structs

ArrowPredicateFn

An ArrowPredicate created from an FnMut

ArrowReaderBuilder

A generic builder for constructing sync or async arrow parquet readers. This is not intended to be used directly, instead you should use the specialization for the type of reader you wish to use

ArrowReaderMetadata

The cheaply clone-able metadata necessary to construct a ArrowReaderBuilder

ArrowReaderOptions

Options that control how metadata is read for a parquet file

ParquetRecordBatchReader

An Iterator<Item = ArrowResult<RecordBatch>> that yields [RecordBatch] read from a parquet data source

ReaderPageIterator 🔒

ReaderRowGroups 🔒

RowFilter

A RowFilter allows pushing down a filter predicate to skip IO and decode

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

Traits

ArrowPredicate

A predicate operating on [RecordBatch]

RowGroups

A collection of row groups

Functions

apply_range 🔒

Applies an optional offset and limit to an optional RowSelection

evaluate_predicate 🔒

Evaluates an ArrowPredicate, returning a RowSelection indicating which rows to return.

selects_any 🔒

Returns true if selection is None or selects some rows

Type Aliases

ParquetRecordBatchReaderBuilder

A synchronous builder used to construct ParquetRecordBatchReader for a file