parquet::arrow::arrow_reader
Module selection

Structs

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

Functions

intersect_row_selections 🔒

Combine two lists of RowSelection return the intersection of them For example: self: NNYYYYNNYYNYN other: NYNNNNNNY

union_row_selections 🔒

Combine two lists of RowSelector return the union of them For example: self: NNYYYYNNYYNYN other: NYNNNNNNY