parquet::arrow::arrow_reader

Module selection

Structs

MaskChunk

Result of computing the next chunk to read when using a MaskCursor

MaskCursor

Cursor for iterating a mask-backed RowSelection

RowSelection

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

RowSelector

RowSelection is a collection of RowSelector used to skip rows when scanning a parquet file

SelectorsCursor

Cursor for iterating a selector-backed RowSelection

Enums

RowSelectionCursor

Cursor for iterating a RowSelection during execution within a ReadPlan.

RowSelectionPolicy

Policy for picking a strategy to materialise RowSelection during execution.

RowSelectionStrategy 🔒

Fully resolved strategy for materializing RowSelection during execution.

Functions

boolean_mask_from_selectors 🔒

intersect_row_selections 🔒

Combine two lists of RowSelection return the intersection of them For example: self: NNYYYYNNYYNYN other: NYNNNNNNY

union_row_selections 🔒

Combine two lists of RowSelector return the union of them For example: self: NNYYYYNNYYNYN other: NYNNNNNNY