parquet::arrow::arrow_reader::statistics

Macro get_statistics

macro_rules! get_statistics {
    ($stat_type_prefix: ident, $data_type: ident, $iterator: ident, $physical_type: ident) => { ... };
}

Special macro to combine the statistics iterators for min and max using the paste macro. This is used to avoid repeating the same code for min and max statistics extractions

Parameters: stat_type_prefix: The prefix of the statistics iterator type (e.g. Min or Max) data_type: The data type of the statistics (e.g. DataType::Int32) iterator: The iterator of ParquetStatistics to extract the statistics from.