parquet::arrow::arrow_reader::statistics
Macro make_stats_iterator

macro_rules! make_stats_iterator {
    ($iterator_type:ident, $func:ident, $parquet_statistics_type:path, $stat_value_type:ty) => { ... };
}

Define an adapter iterator for extracting statistics from an iterator of ParquetStatistics

Handles checking if the statistics are present and valid with the correct type.

Parameters:

$iterator_type is the name of the iterator type (e.g. MinBooleanStatsIterator)

$func is the function to call to get the value (e.g. min or max)

$parquet_statistics_type is the type of the statistics (e.g. ParquetStatistics::Boolean)

$stat_value_type is the type of the statistics value (e.g. bool)