Struct parquet::arrow::arrow_reader::RowSelection

pub struct RowSelection {
    selectors: Vec<RowSelector>,
}

RowSelection allows selecting or skipping a provided number of rows when scanning the parquet file.

This is applied prior to reading column data, and can therefore be used to skip IO to fetch data into memory

A typical use-case would be using the PageIndex to filter out rows that don’t satisfy a predicate

Example

use parquet::arrow::arrow_reader::{RowSelection, RowSelector};

let selectors = vec![
  RowSelector { row_count: 5, skip: true },
  RowSelector { row_count: 5, skip: false },
  RowSelector { row_count: 5, skip: false },
  RowSelector { row_count: 5, skip: true },
];

// Creating a selection will combine adjacent selectors
let selection: RowSelection = selectors.into();

let expected = vec![
  RowSelector { row_count: 5, skip: true },
  RowSelector { row_count: 10, skip: false },
  RowSelector { row_count: 5, skip: true },
];

let actual: Vec<RowSelector> = selection.into();
assert_eq!(actual, expected);

Fields
selectors: Vec<RowSelector>
Implementations

impl RowSelection

pub fn from_filters(filters: &[BooleanArray]) -> Self

Creates a RowSelection from a slice of BooleanArray

Panic

Panics if any of the BooleanArray contain nulls

pub(crate) fn from_consecutive_ranges<I: Iterator<Item = Range<usize>>>( ranges: I, total_rows: usize ) -> Self

Creates a RowSelection from an iterator of consecutive ranges to keep

fn from_selectors_and_combine(selectors: &[RowSelector]) -> Self

Creates a RowSelection from a slice of uncombined RowSelector: Like [skip(5),skip(5),read(10)]. After combine will return [skip(10),read(10)]

Note

RowSelection must be combined prior to use within offset_index or else the code will panic.

pub fn scan_ranges(&self, page_locations: &[PageLocation]) -> Vec<Range<usize>>

Given an offset index, return the byte ranges for all data pages selected by self

This is useful for determining what byte ranges to fetch from underlying storage

Note: this method does not make any effort to combine consecutive ranges, nor coalesce ranges that are close together. This is instead delegated to the IO subsystem to optimise, e.g. ObjectStore::get_ranges

pub fn split_off(&mut self, row_count: usize) -> Self

Splits off the first row_count from this RowSelection

pub fn and_then(&self, other: &Self) -> Self

Given a RowSelection computed under self, returns the RowSelection representing their conjunction

For example:

self: NNNNNNNNNNNNYYYYYYYYYYYYYYYYYYYYYYNNNYYYYY other: YYYYYNNNNYYYYYYYYYYYYY YYNNN

returned: NNNNNNNNNNNNYYYYYNNNNYYYYYYYYYYYYYNNNYYNNN

Panics

Panics if other does not have a length equal to the number of rows selected by this RowSelection

pub fn intersection(&self, other: &Self) -> Self

Compute the intersection of two RowSelection For example: self: NNYYYYNNYYNYN other: NYNNNNNNY

returned: NNNNNNNNYYNYN

pub fn selects_any(&self) -> bool

Returns true if this RowSelection selects any rows

pub(crate) fn trim(self) -> Self

Trims this RowSelection removing any trailing skips

pub(crate) fn offset(self, offset: usize) -> Self

Applies an offset to this RowSelection, skipping the first offset selected rows

pub(crate) fn limit(self, limit: usize) -> Self

Limit this RowSelection to only select limit rows

pub fn iter(&self) -> impl Iterator<Item = &RowSelector>

Returns an iterator over the RowSelectors for this RowSelection.

pub fn row_count(&self) -> usize

Returns the number of selected rows