Trait parquet::arrow::arrow_reader::ArrowReader

pub trait ArrowReader {
    type RecordReader: RecordBatchReader;

    // Required methods
    fn get_schema(&mut self) -> Result<Schema>;
    fn get_schema_by_columns(&mut self, mask: ProjectionMask) -> Result<Schema>;
    fn get_record_reader(
        &mut self,
        batch_size: usize
    ) -> Result<Self::RecordReader>;
    fn get_record_reader_by_columns(
        &mut self,
        mask: ProjectionMask,
        batch_size: usize
    ) -> Result<Self::RecordReader>;
}

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Arrow reader api. With this api, user can get arrow schema from parquet file, and read parquet data into arrow arrays.

Required Associated Types

type RecordReader: RecordBatchReader

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Required Methods

fn get_schema(&mut self) -> Result<Schema>

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Read parquet schema and convert it into arrow schema.

fn get_schema_by_columns(&mut self, mask: ProjectionMask) -> Result<Schema>

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Read parquet schema and convert it into arrow schema. This schema only includes columns identified by mask.

fn get_record_reader(&mut self, batch_size: usize) -> Result<Self::RecordReader>

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Returns record batch reader from whole parquet file.

Arguments

batch_size: The size of each record batch returned from this reader. Only the last batch may contain records less than this size, otherwise record batches returned from this reader should contains exactly batch_size elements.

fn get_record_reader_by_columns( &mut self, mask: ProjectionMask, batch_size: usize ) -> Result<Self::RecordReader>

👎Deprecated: Use ParquetRecordBatchReaderBuilder instead

Returns record batch reader whose record batch contains columns identified by mask.

Arguments

mask: The columns that should be included in record batches. batch_size: Please refer to get_record_reader.