parquet::arrow::arrow_reader

Trait RowGroups

pub trait RowGroups {
    // Required methods
    fn num_rows(&self) -> usize;
    fn column_chunks(&self, i: usize) -> Result<Box<dyn PageIterator>>;
    fn row_groups(&self) -> Box<dyn Iterator<Item = &RowGroupMetaData> + '_>;
    fn metadata(&self) -> &ParquetMetaData;
}

Interface for reading data pages from the columns of one or more RowGroups.

Required Methods

fn num_rows(&self) -> usize

Get the number of rows in this collection

fn column_chunks(&self, i: usize) -> Result<Box<dyn PageIterator>>

Returns a PageIterator for all pages in the specified column chunk across all row groups in this collection.

fn row_groups(&self) -> Box<dyn Iterator<Item = &RowGroupMetaData> + '_>

Returns an iterator over the row groups in this collection

Note this may not include all row groups in Self::metadata.

fn metadata(&self) -> &ParquetMetaData

Returns the parquet metadata