Type Alias parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder

pub type ParquetRecordBatchReaderBuilder<T> = ArrowReaderBuilder<SyncReader<T>>;

A synchronous builder used to construct ParquetRecordBatchReader for a file

For an async API see crate::arrow::async_reader::ParquetRecordBatchStreamBuilder

See ArrowReaderBuilder for additional member functions

Aliased Type

struct ParquetRecordBatchReaderBuilder<T> {
    pub(crate) input: SyncReader<T>,
    pub(crate) metadata: Arc<ParquetMetaData, Global>,
    pub(crate) schema: Arc<Schema, Global>,
    pub(crate) fields: Option<Arc<ParquetField, Global>>,
    pub(crate) batch_size: usize,
    pub(crate) row_groups: Option<Vec<usize, Global>>,
    pub(crate) projection: ProjectionMask,
    pub(crate) filter: Option<RowFilter>,
    pub(crate) selection: Option<RowSelection>,
    pub(crate) limit: Option<usize>,
    pub(crate) offset: Option<usize>,
}

Fields
input: SyncReader<T>metadata: Arc<ParquetMetaData, Global>schema: Arc<Schema, Global>fields: Option<Arc<ParquetField, Global>>batch_size: usizerow_groups: Option<Vec<usize, Global>>projection: ProjectionMaskfilter: Option<RowFilter>selection: Option<RowSelection>limit: Option<usize>offset: Option<usize>
Implementations

impl<T: ChunkReader + 'static> ParquetRecordBatchReaderBuilder<T>

pub fn try_new(reader: T) -> Result<Self>

Create a new ParquetRecordBatchReaderBuilder

let mut builder = ParquetRecordBatchReaderBuilder::try_new(file).unwrap();

// Inspect metadata
assert_eq!(builder.metadata().num_row_groups(), 1);

// Construct reader
let mut reader: ParquetRecordBatchReader = builder.with_row_groups(vec![0]).build().unwrap();

// Read data
let _batch = reader.next().unwrap().unwrap();

pub fn try_new_with_options( reader: T, options: ArrowReaderOptions ) -> Result<Self>

Create a new ParquetRecordBatchReaderBuilder with ArrowReaderOptions

pub fn new_with_metadata(input: T, metadata: ArrowReaderMetadata) -> Self

Create a ParquetRecordBatchReaderBuilder from the provided ArrowReaderMetadata

This allows loading metadata once and using it to create multiple builders with potentially different settings

let metadata = ArrowReaderMetadata::load(&file, Default::default()).unwrap();
let mut a = ParquetRecordBatchReaderBuilder::new_with_metadata(file.clone(), metadata.clone()).build().unwrap();
let mut b = ParquetRecordBatchReaderBuilder::new_with_metadata(file, metadata).build().unwrap();

// Should be able to read from both in parallel
assert_eq!(a.next().unwrap().unwrap(), b.next().unwrap().unwrap());

pub fn build(self) -> Result<ParquetRecordBatchReader>

Build a ParquetRecordBatchReader

Note: this will eagerly evaluate any RowFilter before returning

impl<T> ArrowReaderBuilder<T>

pub(crate) fn new_builder(input: T, metadata: ArrowReaderMetadata) -> Self

pub fn metadata(&self) -> &Arc<ParquetMetaData>

Returns a reference to the ParquetMetaData for this parquet file

pub fn parquet_schema(&self) -> &SchemaDescriptor

Returns the parquet SchemaDescriptor for this parquet file

pub fn schema(&self) -> &SchemaRef

Returns the arrow SchemaRef for this parquet file

pub fn with_batch_size(self, batch_size: usize) -> Self

Set the size of RecordBatch to produce. Defaults to 1024 If the batch_size more than the file row count, use the file row count.

pub fn with_row_groups(self, row_groups: Vec<usize>) -> Self

Only read data from the provided row group indexes

pub fn with_projection(self, mask: ProjectionMask) -> Self

Only read data from the provided column indexes

pub fn with_row_selection(self, selection: RowSelection) -> Self

Provide a RowSelection to filter out rows, and avoid fetching their data into memory.

Row group filtering is applied prior to this, and therefore rows from skipped row groups should not be included in the RowSelection

An example use case of this would be applying a selection determined by evaluating predicates against the Index

pub fn with_row_filter(self, filter: RowFilter) -> Self

Provide a RowFilter to skip decoding rows

Row filters are applied after row group selection and row selection

pub fn with_limit(self, limit: usize) -> Self

Provide a limit to the number of rows to be read

The limit will be applied after any Self::with_row_selection and Self::with_row_filter allowing it to limit the final set of rows decoded after any pushed down predicates

pub fn with_offset(self, offset: usize) -> Self

Provide an offset to skip over the given number of rows

The offset will be applied after any Self::with_row_selection and Self::with_row_filter allowing it to skip rows after any pushed down predicates