Module parquet::arrow::arrow_writer

Contains writer which writes arrow data into parquet data.

Modules

byte_array 🔒

levels 🔒

Parquet definition and repetition levels

Structs

ArrowColumnChunk

The data for a single column chunk, see ArrowColumnWriter

ArrowColumnChunkData 🔒

A single column chunk produced by ArrowColumnWriter

ArrowColumnChunkReader 🔒

A Read for ArrowColumnChunkData

ArrowColumnWriter

Encodes ArrowLeafColumn to ArrowColumnChunk

ArrowLeafColumn

A leaf column that can be encoded by ArrowColumnWriter

ArrowPageWriter 🔒

ArrowRowGroupWriter 🔒

Encodes [RecordBatch] to a parquet row group

ArrowWriter

Encodes [RecordBatch] to parquet

ArrowWriterOptions

Arrow-specific configuration settings for writing parquet files.

Enums

ArrowColumnWriterImpl 🔒

Functions

compute_leaves

Computes the ArrowLeafColumn for a potentially nested [ArrayRef]

get_arrow_column_writer 🔒

Gets the ArrowColumnWriter for the given data_type

get_bool_array_slice 🔒

get_column_writers

Returns the ArrowColumnWriter for a given schema

get_decimal_128_array_slice 🔒

get_decimal_256_array_slice 🔒

get_float_16_array_slice 🔒

get_fsb_array_slice 🔒

get_interval_dt_array_slice 🔒

Returns 12-byte values representing 3 values of months, days and milliseconds (4-bytes each). An Arrow DayTime interval only stores days and millis, thus the first 4 bytes are not populated.

get_interval_ym_array_slice 🔒

Returns 12-byte values representing 3 values of months, days and milliseconds (4-bytes each). An Arrow YearMonth interval only stores months, thus only the first 4 bytes are populated.

write_leaf 🔒

write_primitive 🔒

Type Aliases

SharedColumnChunk 🔒

A shared ArrowColumnChunkData