parquet::arrow

Module arrow_writer

Contains writer which writes arrow data into parquet data.

Structs

ArrowColumnChunk

The data for a single column chunk, see ArrowColumnWriter

ArrowColumnWriter

Encodes ArrowLeafColumn to ArrowColumnChunk

ArrowLeafColumn

A leaf column that can be encoded by ArrowColumnWriter

ArrowRowGroupWriterFactory

Factory that creates new column writers for each row group in the Parquet file.

ArrowWriter

Encodes RecordBatch to parquet

ArrowWriterOptions

Arrow-specific configuration settings for writing parquet files.

Functions

compute_leaves

Computes the ArrowLeafColumn for a potentially nested ArrayRef

get_column_writersDeprecated

Returns ArrowColumnWriters for each column in a given schema