Module parquet::arrow::arrow_writer

Contains writer which writes arrow data into parquet data.

Modules

byte_array ğŸ”’

levels ğŸ”’

Parquet definition and repetition levels

Structs

ArrowColumnChunk ğŸ”’

A list of [Bytes] comprising a single column chunk

ArrowPageWriter ğŸ”’

ArrowRowGroupWriter ğŸ”’

Encodes RecordBatch to a parquet row group

ArrowWriter

Arrow writer

ChainReader ğŸ”’

A Read for an iterator of [Bytes]

Enums

ArrowColumnWriter ğŸ”’

Encodes a leaf column to ArrowPageWriter

Functions

get_arrow_column_writer ğŸ”’

Get an ArrowColumnWriter along with a reference to its SharedColumnChunk

get_bool_array_slice ğŸ”’

get_decimal_128_array_slice ğŸ”’

get_decimal_256_array_slice ğŸ”’

get_fsb_array_slice ğŸ”’

get_interval_dt_array_slice ğŸ”’

Returns 12-byte values representing 3 values of months, days and milliseconds (4-bytes each). An Arrow DayTime interval only stores days and millis, thus the first 4 bytes are not populated.

get_interval_ym_array_slice ğŸ”’

Returns 12-byte values representing 3 values of months, days and milliseconds (4-bytes each). An Arrow YearMonth interval only stores months, thus only the first 4 bytes are populated.

write_leaf ğŸ”’

write_leaves ğŸ”’

Write the leaves of array in depth-first order to writers with levels

write_primitive ğŸ”’

Type Definitions

SharedColumnChunk ğŸ”’

A shared ArrowColumnChunk