parquet::arrow::arrow_writer::byte_array

Function compute_min_max

fn compute_min_max<T>(
    array: T,
    valid: impl Iterator<Item = usize>,
) -> Option<(ByteArray, ByteArray)>
where
    T: ArrayAccessor,
    T::Item: Copy + Ord + AsRef<[u8]>,

Computes the min and max for the provided array and indices

This is a free function so it can be used with downcast_op!