parquet::arrow::arrow_writer

Struct ArrowRowGroupWriterFactory

pub struct ArrowRowGroupWriterFactory {
    schema: SchemaDescPtr,
    arrow_schema: SchemaRef,
    props: WriterPropertiesPtr,
    file_encryptor: Option<Arc<FileEncryptor>>,
}

Factory that creates new column writers for each row group in the Parquet file.

You can create this structure via an ArrowWriter::into_serialized_writer. See the example on ArrowColumnWriter for how to encode columns in parallel

Fields
schema: SchemaDescPtrarrow_schema: SchemaRefprops: WriterPropertiesPtrfile_encryptor: Option<Arc<FileEncryptor>>
Implementations

impl ArrowRowGroupWriterFactory

pub fn new<W: Write + Send>( file_writer: &SerializedFileWriter<W>, arrow_schema: SchemaRef, ) -> Self

Create a new ArrowRowGroupWriterFactory for the provided file writer and Arrow schema

fn create_row_group_writer( &self, row_group_index: usize, ) -> Result<ArrowRowGroupWriter>

pub fn create_column_writers( &self, row_group_index: usize, ) -> Result<Vec<ArrowColumnWriter>>

Create column writers for a new row group, with the given row group index

fn column_writer_factory( &self, row_group_idx: usize, ) -> ArrowColumnWriterFactory