parquet::arrow

Module async_reader

async API for reading Parquet files as [RecordBatch]es

See the crate-level documentation for more details.

See example on ParquetRecordBatchStreamBuilder::new

Modules

metadata 🔒

store 🔒

Structs

ParquetObjectReader

Reads Parquet files in object storage using [ObjectStore].

ParquetRecordBatchStream

An asynchronous Streamof [RecordBatch] constructed using ParquetRecordBatchStreamBuilder to read parquet files.

Enums

RequestState 🔒

State machine that tracks outstanding requests to fetch data

Traits

AsyncFileReader

The asynchronous interface used by ParquetRecordBatchStream to read parquet files

MetadataFetch

A data source that can be used with ParquetMetaDataReader to load ParquetMetaData

MetadataSuffixFetch

A data source that can be used with ParquetMetaDataReader to load ParquetMetaData via suffix requests, without knowing the file size

Type Aliases

ParquetRecordBatchStreamBuilder

A builder for reading parquet files from an async source as ParquetRecordBatchStream