parquet::arrow

Module async_reader

ParquetRecordBatchStreamBuilder: async API for reading Parquet files as [RecordBatch]es

This can be used to decode a Parquet file in streaming fashion (without downloading the whole file at once) from a remote source, such as an object store.

See example on ParquetRecordBatchStreamBuilder::new

Modules

metadata 🔒

store 🔒

Structs

ColumnChunkIterator 🔒

Implements PageIterator for a single column chunk, yielding a single PageReader

InMemoryRowGroup 🔒

An in-memory collection of column chunks

MetadataLoader

An asynchronous interface to load ParquetMetaData from an async source

ParquetObjectReader

Reads Parquet files in object storage using [ObjectStore].

ParquetRecordBatchStream

An asynchronous Streamof [RecordBatch] constructed using ParquetRecordBatchStreamBuilder to read parquet files.

ReaderFactory 🔒

ReaderFactory is used by ParquetRecordBatchStream to create ParquetRecordBatchReader

Enums

ColumnChunkData 🔒

An in-memory column chunk

StreamState 🔒

Traits

AsyncFileReader

The asynchronous interface used by ParquetRecordBatchStream to read parquet files

MetadataFetch

A data source that can be used with MetadataLoader to load ParquetMetaData

Functions

fetch_parquet_metadataDeprecated

Fetches parquet metadata

Type Aliases

ParquetRecordBatchStreamBuilder

A builder for reading parquet files from an async source as ParquetRecordBatchStream

ReadResult 🔒