Type Alias parquet::arrow::async_reader::ParquetRecordBatchStreamBuilder

pub type ParquetRecordBatchStreamBuilder<T> = ArrowReaderBuilder<AsyncReader<T>>;

A builder used to construct a ParquetRecordBatchStream for a parquet file

In particular, this handles reading the parquet file metadata, allowing consumers to use this information to select what specific columns, row groups, etc… they wish to be read by the resulting stream

See ArrowReaderBuilder for additional member functions

Aliased Type

struct ParquetRecordBatchStreamBuilder<T> {
    pub(crate) input: AsyncReader<T>,
    pub(crate) metadata: Arc<ParquetMetaData>,
    pub(crate) schema: Arc<Schema>,
    pub(crate) fields: Option<Arc<ParquetField>>,
    pub(crate) batch_size: usize,
    pub(crate) row_groups: Option<Vec<usize>>,
    pub(crate) projection: ProjectionMask,
    pub(crate) filter: Option<RowFilter>,
    pub(crate) selection: Option<RowSelection>,
    pub(crate) limit: Option<usize>,
    pub(crate) offset: Option<usize>,
}

Fields
input: AsyncReader<T>metadata: Arc<ParquetMetaData>schema: Arc<Schema>fields: Option<Arc<ParquetField>>batch_size: usizerow_groups: Option<Vec<usize>>projection: ProjectionMaskfilter: Option<RowFilter>selection: Option<RowSelection>limit: Option<usize>offset: Option<usize>
Implementations

impl<T: AsyncFileReader + Send + 'static> ParquetRecordBatchStreamBuilder<T>

pub async fn new(input: T) -> Result<Self>

Create a new ParquetRecordBatchStreamBuilder with the provided parquet file

pub async fn new_with_options( input: T, options: ArrowReaderOptions ) -> Result<Self>

Create a new ParquetRecordBatchStreamBuilder with the provided parquet file and ArrowReaderOptions

pub fn new_with_metadata(input: T, metadata: ArrowReaderMetadata) -> Self

Create a ParquetRecordBatchStreamBuilder from the provided ArrowReaderMetadata

This allows loading metadata once and using it to create multiple builders with potentially different settings

let mut file = tempfile().unwrap();
let mut file = tokio::fs::File::from_std(file);
let meta = ArrowReaderMetadata::load_async(&mut file, Default::default()).await.unwrap();
let mut a = ParquetRecordBatchStreamBuilder::new_with_metadata(
    file.try_clone().await.unwrap(),
    meta.clone()
).build().unwrap();
let mut b = ParquetRecordBatchStreamBuilder::new_with_metadata(file, meta).build().unwrap();

// Should be able to read from both in parallel
assert_eq!(a.next().await.unwrap().unwrap(), b.next().await.unwrap().unwrap());

pub async fn get_row_group_column_bloom_filter( &mut self, row_group_idx: usize, column_idx: usize ) -> Result<Option<Sbbf>>

Read bloom filter for a column in a row group Returns None if the column does not have a bloom filter

We should call this function after other forms pruning, such as projection and predicate pushdown.

pub fn build(self) -> Result<ParquetRecordBatchStream<T>>

Build a new ParquetRecordBatchStream