parquet::arrow::buffer

Module bit_util

Functions

count_set_bits

Counts the number of set bits in the provided range

iter_set_bits_rev

Iterates through the set bit positions in bytes in reverse order

sign_extend_be

Performs big endian sign extension