parquet::arrow::buffer::view_buffer

Struct ViewBuffer

pub struct ViewBuffer {
    pub views: Vec<u128>,
    pub buffers: Vec<Buffer>,
}

A buffer of view type byte arrays that can be converted into GenericByteViewArray

Note this does not reuse GenericByteViewBuilder due to the need to call pad_nulls and reuse the existing logic for Vec in the parquet crate

Fields
views: Vec<u128>buffers: Vec<Buffer>
Implementations

impl ViewBuffer

pub fn is_empty(&self) -> bool

pub fn append_block(&mut self, block: Buffer) -> u32

pub unsafe fn append_view_unchecked( &mut self, block: u32, offset: u32, len: u32, )

Safety

This method is only safe when:

block is a valid index, i.e., the return value of append_block

offset and offset + len are valid indices into the buffer

The (offset, offset + len) is valid value for the native type.

pub unsafe fn append_raw_view_unchecked(&mut self, view: &u128)

Directly append a view to the view array. This is used when we create a StringViewArray from a dictionary whose values are StringViewArray.

Safety

The view must be a valid view as per the ByteView spec.

pub fn into_array( self, null_buffer: Option<Buffer>, data_type: &ArrowType, ) -> ArrayRef

Converts this into an [ArrayRef] with the provided data_type and null_buffer