Module parquet::arrow::converter

Structs

ArrayRefConverter

BinaryArrayConverter

DecimalArrayConverter

DictionaryArrayConverter

FixedSizeArrayConverter

FromConverter

Int96ArrayConverter

IntervalDayTimeArrayConverter

An Arrow Interval converter, which reads the last 8 bytes of a Parquet interval, and interprets it as an i32 value representing the Arrow DayTime value

IntervalYearMonthArrayConverter

An Arrow Interval converter, which reads the first 4 bytes of a Parquet interval, and interprets it as an i32 value representing the Arrow YearMonth value

LargeBinaryArrayConverter

LargeUtf8ArrayConverter

StringDictionaryArrayConverter

Utf8ArrayConverter

Traits

Converter

A converter is used to consume record reader’s content and convert it to arrow primitive array.

Type Definitions

BinaryConverter

DecimalConverter

DictionaryConverter

FixedLenBinaryConverter

Int96Converter

IntervalDayTimeConverter

IntervalYearMonthConverter

LargeBinaryConverter

LargeUtf8Converter

PrimitiveDictionaryConverter

StringDictionaryConverter

Utf8Converter