Trait parquet::arrow::record_reader::buffer::ValuesBuffer

pub trait ValuesBuffer: BufferQueue {
    // Required method
    fn pad_nulls(
        &mut self,
        read_offset: usize,
        values_read: usize,
        levels_read: usize,
        valid_mask: &[u8]
    );
}

A BufferQueue capable of storing column values

Required Methods

fn pad_nulls( &mut self, read_offset: usize, values_read: usize, levels_read: usize, valid_mask: &[u8] )

If a column contains nulls, more level data may be read than value data, as null values are not encoded. Therefore, first the levels data is read, the null count determined, and then the corresponding number of values read to a ValuesBuffer.

It is then necessary to move this values data into positions that correspond to the non-null level positions. This is what this method does.

It is provided with:

read_offset - the offset in ValuesBuffer to start null padding from

values_read - the number of values read

levels_read - the number of levels read

valid_mask - a packed mask of valid levels

Object Safety

This trait is not object safe.