parquet

Module basic

Contains Rust mappings for Thrift definition. This module contains only mappings for thrift enums and unions. Thrift structs are handled elsewhere. Refer to parquet.thrift file to see raw definitions.

Structs

BrotliLevel

Represents a valid brotli compression level.

DecimalType 🔒

EncodingMask

A bitmask representing the Encodings employed while encoding a Parquet column chunk.

GeographyType 🔒

GeometryType 🔒

GzipLevel

Represents a valid gzip compression level.

IntType 🔒

TimestampType 🔒

VariantType 🔒

ZstdLevel

Represents a valid zstd compression level.

Enums

BloomFilterAlgorithm

The algorithm used in Bloom filter.

BloomFilterCompression

The compression used in the Bloom filter.

BloomFilterHash

The hash function used in Bloom filter. This function takes the hash of a column value using plain encoding.

BoundaryOrder

Enum to annotate whether lists of min/max elements inside ColumnIndex are ordered and if so, in which direction.

ColumnOrder

Column order that specifies what method was used to aggregate min/max values for statistics.

Compression

Supported block compression algorithms.

ConvertedType

Common types (converted types) used by frameworks when using Parquet.

EdgeInterpolationAlgorithm

Edge interpolation algorithm for LogicalType::Geography

Encoding

Encodings supported by Parquet.

FieldRepetitionType

Representation of field types in schema.

LogicalType

Logical types used by version 2.4.0+ of the Parquet format.

PageType

Available data pages for Parquet file format. Note that some of the page types may not be supported.

SortOrder

Sort order for page and column statistics.

TimeUnit

Time unit for Time and Timestamp logical types.

Type

Types supported by Parquet.

Functions

check_level_is_none 🔒

i32_to_encoding 🔒

require_level 🔒

split_compression_string 🔒

Type Aliases

Repetition

Type alias for thrift FieldRepetitionType