Struct parquet::bloom_filter::Block

struct Block([u32; 8]);

Each block is 256 bits, broken up into eight contiguous “words”, each consisting of 32 bits. Each word is thought of as an array of bits; each bit is either “set” or “not set”.

Tuple Fields
0: [u32; 8]
Implementations

impl Block

const ZERO: Block = _

fn mask(x: u32) -> Self

takes as its argument a single unsigned 32-bit integer and returns a block in which each word has exactly one bit set.

fn to_le_bytes(self) -> [u8; 32]

fn to_ne_bytes(self) -> [u8; 32]

fn insert(&mut self, hash: u32)

setting every bit in the block that was also set in the result from mask

fn check(&self, hash: u32) -> bool

returns true when every bit that is set in the result of mask is also set in the block.