parquet::bloom_filter

Struct Sbbf

pub struct Sbbf(Vec<Block>);

A split block Bloom filter.

The creation of this structure is based on the crate::file::properties::BloomFilterProperties struct set via crate::file::properties::WriterProperties and is thus hidden by default.

Tuple Fields
0: Vec<Block>
Implementations

impl Sbbf

pub(crate) fn new_with_ndv_fpp(ndv: u64, fpp: f64) -> Result<Self, ParquetError>

Create a new Sbbf with given number of distinct values and false positive probability. Will return an error if fpp is greater than or equal to 1.0 or less than 0.0.

pub(crate) fn new_with_num_of_bytes(num_bytes: usize) -> Self

Create a new Sbbf with given number of bytes, the exact number of bytes will be adjusted to the next power of two bounded by BITSET_MIN_LENGTH and BITSET_MAX_LENGTH.

pub(crate) fn new(bitset: &[u8]) -> Self

pub(crate) fn write<W: Write>(&self, writer: W) -> Result<(), ParquetError>

Write the bloom filter data (header and then bitset) to the output. This doesnโt flush the writer in order to boost performance of bulk writing all blocks. Caller must remember to flush the writer.

fn write_bitset<W: Write>(&self, writer: W) -> Result<(), ParquetError>

Write the bitset in serialized form to the writer.

fn header(&self) -> BloomFilterHeader

Create and populate BloomFilterHeader from this bitset for writing to serialized form

pub(crate) fn read_from_column_chunk<R: ChunkReader>( column_metadata: &ColumnChunkMetaData, reader: Arc<R>, ) -> Result<Option<Self>, ParquetError>

Read a new bloom filter from the given offset in the given reader.

fn hash_to_block_index(&self, hash: u64) -> usize

pub fn insert<T: AsBytes + ?Sized>(&mut self, value: &T)

Insert an AsBytes value into the filter

fn insert_hash(&mut self, hash: u64)

Insert a hash into the filter

pub fn check<T: AsBytes>(&self, value: &T) -> bool

Check if an AsBytes value is probably present or definitely absent in the filter

fn check_hash(&self, hash: u64) -> bool

Check if a hash is in the filter. May return true for values that was never inserted (โfalse positiveโ) but will always return false if a hash has not been inserted.

pub(crate) fn estimated_memory_size(&self) -> usize

Return the total in memory size of this bloom filter in bytes