parquet::column

Module page

Contains Parquet Page definitions and page reader interface.

Structs

CompressedPage

Helper struct to represent pages with potentially compressed buffer (data page v1) or compressed and concatenated buffer (def levels + rep levels + compressed values for data page v2).

PageMetadata

Contains metadata for a page

PageWriteSpec

Contains page write metrics.

Enums

Page

Parquet Page definition.

Traits

PageIterator

An iterator over pages of one specific column in a parquet file.

PageReader

API for reading pages from a column chunk. This offers a iterator like API to get the next page.

PageWriter

API for writing pages in a column chunk.