parquet::column::page

Enum Page

pub enum Page {
    DataPage {
        buf: Bytes,
        num_values: u32,
        encoding: Encoding,
        def_level_encoding: Encoding,
        rep_level_encoding: Encoding,
        statistics: Option<Statistics>,
    },
    DataPageV2 {
        buf: Bytes,
        num_values: u32,
        encoding: Encoding,
        num_nulls: u32,
        num_rows: u32,
        def_levels_byte_len: u32,
        rep_levels_byte_len: u32,
        is_compressed: bool,
        statistics: Option<Statistics>,
    },
    DictionaryPage {
        buf: Bytes,
        num_values: u32,
        encoding: Encoding,
        is_sorted: bool,
    },
}

Parquet Page definition.

List of supported pages. These are 1-to-1 mapped from the equivalent Thrift definitions, except buf which used to store uncompressed bytes of the page.

Variants

DataPage

Data page Parquet format v1.

Fields

buf: Bytes

The underlying data buffer

num_values: u32

Number of values in this page

encoding: Encoding

Encoding for values in this page

def_level_encoding: Encoding

Definition level encoding

rep_level_encoding: Encoding

Repetition level encoding

statistics: Option<Statistics>

Optional statistics for this page

DataPageV2

Data page Parquet format v2.

Fields

buf: Bytes

The underlying data buffer

num_values: u32

Number of values in this page

encoding: Encoding

Encoding for values in this page

num_nulls: u32

Number of null values in this page

num_rows: u32

Number of rows in this page

def_levels_byte_len: u32

Length of definition levels

rep_levels_byte_len: u32

Length of repetition levels

is_compressed: bool

Is this page compressed

statistics: Option<Statistics>

Optional statistics for this page

DictionaryPage

Dictionary page.

Fields

buf: Bytes

The underlying data buffer

num_values: u32

Number of values in this page

encoding: Encoding

Encoding for values in this page

is_sorted: bool

Is dictionary page sorted

Implementations

impl Page

pub fn page_type(&self) -> PageType

Returns PageType for this page.

pub fn buffer(&self) -> &Bytes

Returns internal byte buffer reference for this page.

pub fn num_values(&self) -> u32

Returns number of values in this page.

pub fn encoding(&self) -> Encoding

Returns this page Encoding.

pub fn statistics(&self) -> Option<&Statistics>

Returns optional Statistics.