Enum parquet::column::page::Page

pub enum Page {
    DataPage {
        buf: ByteBufferPtr,
        num_values: u32,
        encoding: Encoding,
        def_level_encoding: Encoding,
        rep_level_encoding: Encoding,
        statistics: Option<Statistics>,
    },
    DataPageV2 {
        buf: ByteBufferPtr,
        num_values: u32,
        encoding: Encoding,
        num_nulls: u32,
        num_rows: u32,
        def_levels_byte_len: u32,
        rep_levels_byte_len: u32,
        is_compressed: bool,
        statistics: Option<Statistics>,
    },
    DictionaryPage {
        buf: ByteBufferPtr,
        num_values: u32,
        encoding: Encoding,
        is_sorted: bool,
    },
}

Parquet Page definition.

List of supported pages. These are 1-to-1 mapped from the equivalent Thrift definitions, except buf which used to store uncompressed bytes of the page.

Variants
DataPage
Fields of DataPage

buf: ByteBufferPtrnum_values: u32encoding: Encodingdef_level_encoding: Encodingrep_level_encoding: Encodingstatistics: Option<Statistics>
DataPageV2
Fields of DataPageV2

buf: ByteBufferPtrnum_values: u32encoding: Encodingnum_nulls: u32num_rows: u32def_levels_byte_len: u32rep_levels_byte_len: u32is_compressed: boolstatistics: Option<Statistics>
DictionaryPage
Fields of DictionaryPage

buf: ByteBufferPtrnum_values: u32encoding: Encodingis_sorted: bool

Methods

impl Page

pub fn page_type(&self) -> PageType

Returns PageType for this page.

pub fn buffer(&self) -> &ByteBufferPtr

Returns internal byte buffer reference for this page.

pub fn num_values(&self) -> u32

Returns number of values in this page.

pub fn encoding(&self) -> Encoding

Returns this page Encoding.

pub fn statistics(&self) -> Option<&Statistics>

Returns optional Statistics.