Struct parquet::column::page::CompressedPage

pub struct CompressedPage { /* fields omitted */ }

Helper struct to represent pages with potentially compressed buffer (data page v1) or compressed and concatenated buffer (def levels + rep levels + compressed values for data page v2).

The difference with Page is that Page buffer is always uncompressed.

Methods

impl CompressedPage

pub fn new(compressed_page: Page, uncompressed_size: usize) -> Self

Creates CompressedPage from a page with potentially compressed buffer and uncompressed size.

pub fn page_type(&self) -> PageType

Returns page type.

pub fn compressed_page(&self) -> &Page

Returns underlying page with potentially compressed buffer.

pub fn uncompressed_size(&self) -> usize

Returns uncompressed size in bytes.

pub fn compressed_size(&self) -> usize

Returns compressed size in bytes.

Note that it is assumed that buffer is compressed, but it may not be. In this case compressed size will be equal to uncompressed size.

pub fn num_values(&self) -> u32

Number of values in page.

pub fn encoding(&self) -> Encoding

Returns encoding for values in page.

pub fn data(&self) -> &[u8]

Returns slice of compressed buffer in the page.