parquet::column::reader::decoder

Trait DefinitionLevelDecoder

pub trait DefinitionLevelDecoder: ColumnLevelDecoder {
    // Required methods
    fn read_def_levels(
        &mut self,
        out: &mut Self::Buffer,
        num_levels: usize,
    ) -> Result<(usize, usize)>;
    fn skip_def_levels(&mut self, num_levels: usize) -> Result<(usize, usize)>;
}

Required Methods

fn read_def_levels( &mut self, out: &mut Self::Buffer, num_levels: usize, ) -> Result<(usize, usize)>

Read up to num_levels definition levels into out.

Returns the number of values read, and the number of levels read.

Panics

Implementations may panic if range overlaps with already written data

fn skip_def_levels(&mut self, num_levels: usize) -> Result<(usize, usize)>

Skips over num_levels definition levels.

Returns the number of values skipped, and the number of levels skipped.