Type Alias parquet::column::reader::ColumnReaderImpl

pub type ColumnReaderImpl<T> = GenericColumnReader<RepetitionLevelDecoderImpl, DefinitionLevelDecoderImpl, ColumnValueDecoderImpl<T>>;

Typed value reader for a particular primitive column.

Aliased Type

struct ColumnReaderImpl<T> {
    descr: Arc<ColumnDescriptor>,
    page_reader: Box<dyn PageReader<Item = Result<Page, ParquetError>>>,
    num_buffered_values: usize,
    num_decoded_values: usize,
    has_record_delimiter: bool,
    def_level_decoder: Option<DefinitionLevelDecoderImpl>,
    rep_level_decoder: Option<RepetitionLevelDecoderImpl>,
    values_decoder: ColumnValueDecoderImpl<T>,
}

Fields
descr: Arc<ColumnDescriptor>page_reader: Box<dyn PageReader<Item = Result<Page, ParquetError>>>num_buffered_values: usize

The total number of values stored in the data page.

num_decoded_values: usize

The number of values from the current data page that has been decoded into memory so far.

has_record_delimiter: bool

True if the end of the current data page denotes the end of a record

def_level_decoder: Option<DefinitionLevelDecoderImpl>

The decoder for the definition levels if any

rep_level_decoder: Option<RepetitionLevelDecoderImpl>

The decoder for the repetition levels if any

values_decoder: ColumnValueDecoderImpl<T>

The decoder for the values