Trait parquet::column::writer::encoder::ColumnValueEncoder

pub trait ColumnValueEncoder {
    type T: ParquetValueType;
    type Values: ColumnValues + ?Sized;

    // Required methods
    fn try_new(descr: &ColumnDescPtr, props: &WriterProperties) -> Result<Self>
       where Self: Sized;
    fn write(
        &mut self,
        values: &Self::Values,
        offset: usize,
        len: usize,
    ) -> Result<()>;
    fn write_gather(
        &mut self,
        values: &Self::Values,
        indices: &[usize],
    ) -> Result<()>;
    fn num_values(&self) -> usize;
    fn has_dictionary(&self) -> bool;
    fn estimated_dict_page_size(&self) -> Option<usize>;
    fn estimated_data_page_size(&self) -> usize;
    fn flush_dict_page(&mut self) -> Result<Option<DictionaryPage>>;
    fn flush_data_page(&mut self) -> Result<DataPageValues<Self::T>>;
    fn flush_bloom_filter(&mut self) -> Option<Sbbf>;
}

A generic encoder of ColumnValues to data and dictionary pages used by super::GenericColumnWriter`

Required Associated Types

type T: ParquetValueType

The underlying value type of Self::Values

Note: this avoids needing to fully qualify <Self::Values as ColumnValues>::T

type Values: ColumnValues + ?Sized

The values encoded by this encoder

Required Methods

fn try_new(descr: &ColumnDescPtr, props: &WriterProperties) -> Result<Self>
where Self: Sized,

Create a new ColumnValueEncoder

fn write( &mut self, values: &Self::Values, offset: usize, len: usize, ) -> Result<()>

Write the corresponding values to this ColumnValueEncoder

fn write_gather( &mut self, values: &Self::Values, indices: &[usize], ) -> Result<()>

Write the values at the indexes in indices to this ColumnValueEncoder

fn num_values(&self) -> usize

Returns the number of buffered values

fn has_dictionary(&self) -> bool

Returns true if this encoder has a dictionary page

fn estimated_dict_page_size(&self) -> Option<usize>

Returns an estimate of the dictionary page size in bytes, or None if no dictionary

fn estimated_data_page_size(&self) -> usize

Returns an estimate of the data page size in bytes

fn flush_dict_page(&mut self) -> Result<Option<DictionaryPage>>

Flush the dictionary page for this column chunk if any. Any subsequent calls to Self::write will not be dictionary encoded

Note: Self::flush_data_page must be called first, as this will error if there are any pending page values

fn flush_data_page(&mut self) -> Result<DataPageValues<Self::T>>

Flush the next data page for this column chunk

fn flush_bloom_filter(&mut self) -> Option<Sbbf>

Flushes bloom filter if enabled and returns it, otherwise returns None. Subsequent writes will not be tracked by the bloom filter as it is empty since. This should be called once near the end of encoding.