parquet

Module data_type

Data types that connect Parquet physical types with their Rust-specific representations.

Modules

private 🔒

Macros

ensure_phys_ty 🔒

Macro to reduce repetition in making type assertions on the physical type against T

gen_as_bytes 🔒

make_type 🔒

unimplemented_slice_as_bytes 🔒

Structs

BoolType

Parquet physical type: BoolType

ByteArray

Rust representation for BYTE_ARRAY and FIXED_LEN_BYTE_ARRAY Parquet physical types. Value is backed by a byte buffer.

ByteArrayType

Parquet physical type: ByteArrayType

DoubleType

Parquet physical type: DoubleType

FixedLenByteArray

Wrapper type for performance reasons, this represents FIXED_LEN_BYTE_ARRAY but in all other considerations behaves the same as ByteArray

FixedLenByteArrayType

Parquet physical type: FixedLenByteArrayType

FloatType

Parquet physical type: FloatType

Int96

Rust representation for logical type INT96, value is backed by an array of u32. The type only takes 12 bytes, without extra padding.

Int32Type

Parquet physical type: Int32Type

Int64Type

Parquet physical type: Int64Type

Int96Type

Parquet physical type: Int96Type

Enums

Decimal

Rust representation for Decimal values.

Traits

AsBytes

Converts an instance of data type to a slice of bytes as u8.

DataType

Contains the Parquet physical type information as well as the Rust primitive type presentation.

SliceAsBytes

Converts an slice of a data type to a slice of bytes.

SliceAsBytesDataTypeDeprecated