Module parquet::data_type

Data types that connect Parquet physical types with their Rust-specific representations.

Structs

BoolType

ByteArray
Rust representation for BYTE_ARRAY and FIXED_LEN_BYTE_ARRAY Parquet physical types. Value is backed by a byte buffer.

ByteArrayType

DoubleType

FixedLenByteArray
Wrapper type for performance reasons, this represents FIXED_LEN_BYTE_ARRAY but in all other considerations behaves the same as ByteArray

FixedLenByteArrayType

FloatType

Int32Type

Int64Type

Int96
Rust representation for logical type INT96, value is backed by an array of u32. The type only takes 12 bytes, without extra padding.

Int96Type

Enums

Decimal
Rust representation for Decimal values.

Traits

AsBytes
Converts an instance of data type to a slice of bytes as u8.

DataType
Contains the Parquet physical type information as well as the Rust primitive type presentation.

SliceAsBytes
Converts an slice of a data type to a slice of bytes.

SliceAsBytesDataType