parquet::data_type

Struct FixedLenByteArray

#[repr(transparent)]
pub struct FixedLenByteArray(ByteArray);

Wrapper type for performance reasons, this represents FIXED_LEN_BYTE_ARRAY but in all other considerations behaves the same as ByteArray

Performance notes:

This type is a little unfortunate, without it the compiler generates code that takes quite a big hit on the CPU pipeline. Essentially the previous version stalls awaiting the result of T::get_physical_type() == Type::FIXED_LEN_BYTE_ARRAY.

Its debatable if this is wanted, it is out of spec for what parquet documents as its base types, although there are code paths in the Rust (and potentially the C++) versions that warrant this.

With this wrapper type the compiler generates more targeted code paths matching the higher level logical types, removing the data-hazard from all decoding and encoding paths.

Tuple Fields
0: ByteArray