Trait parquet::data_type::DataType

pub trait DataType: 'static + Send {
    type T: ParquetValueType;

    // Required methods
    fn get_type_size() -> usize;
    fn get_column_reader(
        column_writer: ColumnReader,
    ) -> Option<ColumnReaderImpl<Self>>
       where Self: Sized;
    fn get_column_writer(
        column_writer: ColumnWriter<'_>,
    ) -> Option<ColumnWriterImpl<'_, Self>>
       where Self: Sized;
    fn get_column_writer_ref<'a, 'b: 'a>(
        column_writer: &'b ColumnWriter<'a>,
    ) -> Option<&'b ColumnWriterImpl<'a, Self>>
       where Self: Sized;
    fn get_column_writer_mut<'a, 'b: 'a>(
        column_writer: &'a mut ColumnWriter<'b>,
    ) -> Option<&'a mut ColumnWriterImpl<'b, Self>>
       where Self: Sized;

    // Provided method
    fn get_physical_type() -> Type { ... }
}

Contains the Parquet physical type information as well as the Rust primitive type presentation.

Required Associated Types

type T: ParquetValueType

Required Methods

fn get_type_size() -> usize

Returns size in bytes for Rust representation of the physical type.

fn get_column_reader( column_writer: ColumnReader, ) -> Option<ColumnReaderImpl<Self>>
where Self: Sized,

fn get_column_writer( column_writer: ColumnWriter<'_>, ) -> Option<ColumnWriterImpl<'_, Self>>
where Self: Sized,

fn get_column_writer_ref<'a, 'b: 'a>( column_writer: &'b ColumnWriter<'a>, ) -> Option<&'b ColumnWriterImpl<'a, Self>>
where Self: Sized,

fn get_column_writer_mut<'a, 'b: 'a>( column_writer: &'a mut ColumnWriter<'b>, ) -> Option<&'a mut ColumnWriterImpl<'b, Self>>
where Self: Sized,

Provided Methods

fn get_physical_type() -> Type

Returns Parquet physical type.

Object Safety

This trait is not object safe.