Trait parquet::data_type::DataType

pub trait DataType {
    type T: PartialEq + Debug + Default + Clone + Rand + AsBytes;
    fn get_physical_type() -> Type;

    fn get_type_size() -> usize;
}

Contains the Parquet physical type information as well as the Rust primitive type presentation.

Associated Types

type T: PartialEq + Debug + Default + Clone + Rand + AsBytes

Required Methods

fn get_physical_type() -> Type

Returns Parquet physical type.

fn get_type_size() -> usize

Returns size in bytes for Rust representation of the physical type.