Trait parquet::data_type::DataType

pub trait DataType: 'static {
    type T: ParquetValueType;
    fn get_physical_type() -> Type;

    fn get_type_size() -> usize;

    fn get_column_reader(
        column_writer: ColumnReader
    ) -> Option<ColumnReaderImpl<Self>>
    where
        Self: Sized;

    fn get_column_writer(
        column_writer: ColumnWriter
    ) -> Option<ColumnWriterImpl<Self>>
    where
        Self: Sized;

    fn get_column_writer_ref(
        column_writer: &ColumnWriter
    ) -> Option<&ColumnWriterImpl<Self>>
    where
        Self: Sized;

    fn get_column_writer_mut(
        column_writer: &mut ColumnWriter
    ) -> Option<&mut ColumnWriterImpl<Self>>
    where
        Self: Sized;
}

Contains the Parquet physical type information as well as the Rust primitive type presentation.

Associated Types

type T: ParquetValueType

Loading content...
Required methods

fn get_physical_type() -> Type

Returns Parquet physical type.

fn get_type_size() -> usize

Returns size in bytes for Rust representation of the physical type.

fn get_column_reader(
column_writer: ColumnReader
) -> Option<ColumnReaderImpl<Self>> where
Self: Sized,

fn get_column_writer(
column_writer: ColumnWriter
) -> Option<ColumnWriterImpl<Self>> where
Self: Sized,

fn get_column_writer_ref(
column_writer: &ColumnWriter
) -> Option<&ColumnWriterImpl<Self>> where
Self: Sized,

fn get_column_writer_mut(
column_writer: &mut ColumnWriter
) -> Option<&mut ColumnWriterImpl<Self>> where
Self: Sized,

Loading content...