Module parquet::encoding

Contains all supported encoders for Parquet.

Structs

DeltaBitPackEncoder

Delta bit packed encoder. Consists of a header followed by blocks of delta encoded values binary packed.

DeltaByteArrayEncoder

Encoding for byte arrays, prefix lengths are encoded using DELTA_BINARY_PACKED encoding, followed by suffixes with DELTA_LENGTH_BYTE_ARRAY encoding.

DeltaLengthByteArrayEncoder

Encoding for byte arrays to separate the length values and the data. The lengths are encoded using DELTA_BINARY_PACKED encoding, data is stored as raw bytes.

DictEncoder

Dictionary encoder. The dictionary encoding builds a dictionary of values encountered in a given column. The dictionary page is written first, before the data pages of the column chunk.

PlainEncoder

Plain encoding that supports all types. Values are encoded back to back. The plain encoding is used whenever a more efficient encoding can not be used. It stores the data in the following format:

RleValueEncoder

RLE/Bit-Packing hybrid encoding for values. Currently is used only for data pages v2 and supports boolean types.

Traits

Encoder

An Parquet encoder for the data type T.

Functions

get_encoder

Gets a encoder for the particular data type T and encoding encoding. Memory usage for the encoder instance is tracked by mem_tracker.