Function parquet::encoding::get_encoder

pub fn get_encoder<T: DataType>(
    desc: ColumnDescPtr, 
    encoding: Encoding, 
    mem_tracker: MemTrackerPtr
) -> Result<Box<dyn Encoder<T>>>

Gets a encoder for the particular data type T and encoding encoding. Memory usage for the encoder instance is tracked by mem_tracker.