parquet::file::footer

Function parse_metadata

pub fn parse_metadata<R: ChunkReader>(
    chunk_reader: &R,
) -> Result<ParquetMetaData>

👎Deprecated since 53.1.0: Use ParquetMetaDataReader

Reads the ParquetMetaData from the footer of the parquet file.

Layout of Parquet file

+---------------------------+-----+---+
|      Rest of file         |  B  | A |
+---------------------------+-----+---+

where

A: parquet footer which stores the length of the metadata.

B: parquet metadata.

I/O

This method first reads the last 8 bytes of the file via ChunkReader::get_read to get the the parquet footer which contains the metadata length.

It then issues a second get_read to read the encoded metadata metadata.

See Also

decode_metadata for decoding the metadata from the bytes. decode_footer for decoding the metadata length from the footer.