Function parquet::file::footer::parse_metadata

pub fn parse_metadata<R: ChunkReader>(
    chunk_reader: &R,
) -> Result<ParquetMetaData>

Layout of Parquet file +—————————+—–+—+ | Rest of file | B | A | +—————————+—–+—+ where A: parquet footer, B: parquet metadata.

The reader first reads DEFAULT_FOOTER_SIZE bytes from the end of the file. If it is not enough according to the length indicated in the footer, it reads more bytes.