Module parquet::file::metadata

Parquet metadata structures

ParquetMetaData: Top level metadata container, read from the Parquet file footer.

FileMetaData: File level metadata such as schema, row counts and version.

RowGroupMetaData: Metadata for each Row Group with a File, such as location and number of rows, and column chunks.

ColumnChunkMetaData: Metadata for each column chunk (primitive leaf) within a Row Group including encoding and compression information, number of values, statistics, etc.

Structs

ColumnChunkMetaData

Metadata for a column chunk.

ColumnChunkMetaDataBuilder

Builder for column chunk metadata.

ColumnIndexBuilder

Builder for column index

FileMetaData

File level metadata for a Parquet file.

OffsetIndexBuilder

Builder for offset index

ParquetMetaData

Parsed metadata for a single Parquet file

RowGroupMetaData

Metadata for a row group

RowGroupMetaDataBuilder

Builder for row group metadata.

Type Aliases

FileMetaDataPtr

Reference counted pointer for FileMetaData.

KeyValue

ParquetColumnIndex

Index for each row group of each column.

ParquetOffsetIndex

PageLocation for each data page of each row group of each column.

RowGroupMetaDataPtr

Reference counted pointer for RowGroupMetaData.