Module parquet::file::metadata

Contains information about available Parquet metadata.

The hierarchy of metadata is as follows:

ParquetMetaData contains FileMetaData and zero or more RowGroupMetaData for each row group.

FileMetaData includes file version, application specific metadata.

Each RowGroupMetaData contains information about row group and one or more ColumnChunkMetaData for each column chunk.

ColumnChunkMetaData has information about column chunk (primitive leaf column), including encoding/compression, number of values, etc.

Structs

ColumnChunkMetaData

Metadata for a column chunk.

ColumnChunkMetaDataBuilder

Builder for column chunk metadata.

ColumnIndexBuilder

Builder for column index

FileMetaData

Metadata for a Parquet file.

OffsetIndexBuilder

Builder for offset index

ParquetMetaData

Global Parquet metadata.

RowGroupMetaData

Metadata for a row group.

RowGroupMetaDataBuilder

Builder for row group metadata.

Type Aliases

FileMetaDataPtr

Reference counted pointer for FileMetaData.

KeyValue

ParquetColumnIndex

Index for each row group of each column.

ParquetOffsetIndex

PageLocation for each datapage of each row group of each column.

RowGroupMetaDataPtr

Reference counted pointer for RowGroupMetaData.