Module parquet::file::metadata

Parquet metadata API

Most users should use these structures to interact with Parquet metadata. The crate::format module contains lower level structures generated from the Parquet thrift definition.

ParquetMetaData: Top level metadata container, read from the Parquet file footer.

FileMetaData: File level metadata such as schema, row counts and version.

RowGroupMetaData: Metadata for each Row Group with a File, such as location and number of rows, and column chunks.

ColumnChunkMetaData: Metadata for each column chunk (primitive leaf) within a Row Group including encoding and compression information, number of values, statistics, etc.

Modules

memory 🔒

Memory calculations for ParquetMetadata::memory_size

Structs

ColumnChunkMetaData

Metadata for a column chunk.

ColumnChunkMetaDataBuilder

Builder for column chunk metadata.

ColumnIndexBuilder

Builder for Parquet ColumnIndex, part of the Parquet PageIndex

FileMetaData

File level metadata for a Parquet file.

LevelHistogram

Histograms for repetition and definition levels.

OffsetIndexBuilder

Builder for offset index, part of the Parquet PageIndex.

ParquetMetaData

Parsed metadata for a single Parquet file

RowGroupMetaData

Metadata for a row group

RowGroupMetaDataBuilder

Builder for row group metadata.

Type Aliases

FileMetaDataPtr

Reference counted pointer for FileMetaData.

KeyValue

ParquetColumnIndex

Page level statistics for each column chunk of each row group.

ParquetOffsetIndex

OffsetIndexMetaData for each data page of each row group of each column

RowGroupMetaDataPtr

Reference counted pointer for RowGroupMetaData.