parquet::file::metadata::memory

Trait HeapSize

pub trait HeapSize {
    // Required method
    fn heap_size(&self) -> usize;
}

Trait for calculating the size of various containers

Required Methods

fn heap_size(&self) -> usize

Return the size of any bytes allocated on the heap by this object, including heap memory in those structures

Note that the size of the type itself is not included in the result – instead, that size is added by the caller (e.g. container).