parquet::file::metadata

Struct ColumnChunkMetaData

pub struct ColumnChunkMetaData {

    column_descr: ColumnDescPtr,
    encodings: Vec<Encoding>,
    file_path: Option<String>,
    file_offset: i64,
    num_values: i64,
    compression: Compression,
    total_compressed_size: i64,
    total_uncompressed_size: i64,
    data_page_offset: i64,
    index_page_offset: Option<i64>,
    dictionary_page_offset: Option<i64>,
    statistics: Option<Statistics>,
    encoding_stats: Option<Vec<PageEncodingStats>>,
    bloom_filter_offset: Option<i64>,
    bloom_filter_length: Option<i32>,
    offset_index_offset: Option<i64>,
    offset_index_length: Option<i32>,
    column_index_offset: Option<i64>,
    column_index_length: Option<i32>,
    unencoded_byte_array_data_bytes: Option<i64>,
    repetition_level_histogram: Option<LevelHistogram>,
    definition_level_histogram: Option<LevelHistogram>,

}

Metadata for a column chunk.

Fields
column_descr: ColumnDescPtrencodings: Vec<Encoding>file_path: Option<String>file_offset: i64num_values: i64compression: Compressiontotal_compressed_size: i64total_uncompressed_size: i64data_page_offset: i64index_page_offset: Option<i64>dictionary_page_offset: Option<i64>statistics: Option<Statistics>encoding_stats: Option<Vec<PageEncodingStats>>bloom_filter_offset: Option<i64>bloom_filter_length: Option<i32>offset_index_offset: Option<i64>offset_index_length: Option<i32>column_index_offset: Option<i64>column_index_length: Option<i32>unencoded_byte_array_data_bytes: Option<i64>repetition_level_histogram: Option<LevelHistogram>definition_level_histogram: Option<LevelHistogram>
Implementations

impl ColumnChunkMetaData

Represents common operations for a column chunk.

pub fn builder(column_descr: ColumnDescPtr) -> ColumnChunkMetaDataBuilder

Returns builder for column chunk metadata.

pub fn file_path(&self) -> Option<&str>

File where the column chunk is stored.

If not set, assumed to belong to the same file as the metadata. This path is relative to the current file.

pub fn file_offset(&self) -> i64

Byte offset of ColumnMetaData in file_path().

Note that the meaning of this field has been inconsistent between implementations so its use has since been deprecated in the Parquet specification. Modern implementations will set this to 0 to indicate that the ColumnMetaData is solely contained in the ColumnChunk struct.

pub fn column_type(&self) -> Type

Type of this column. Must be primitive.

pub fn column_path(&self) -> &ColumnPath

Path (or identifier) of this column.

pub fn column_descr(&self) -> &ColumnDescriptor

Descriptor for this column.

pub fn column_descr_ptr(&self) -> ColumnDescPtr

Reference counted clone of descriptor for this column.

pub fn encodings(&self) -> &Vec<Encoding>

All encodings used for this column.

pub fn num_values(&self) -> i64

Total number of values in this column chunk.

pub fn compression(&self) -> Compression

Compression for this column.

pub fn compressed_size(&self) -> i64

Returns the total compressed data size of this column chunk.

pub fn uncompressed_size(&self) -> i64

Returns the total uncompressed data size of this column chunk.

pub fn data_page_offset(&self) -> i64

Returns the offset for the column data.

pub fn index_page_offset(&self) -> Option<i64>

Returns the offset for the index page.

pub fn dictionary_page_offset(&self) -> Option<i64>

Returns the offset for the dictionary page, if any.

pub fn byte_range(&self) -> (u64, u64)

Returns the offset and length in bytes of the column chunk within the file

pub fn statistics(&self) -> Option<&Statistics>

Returns statistics that are set for this column chunk, or None if no statistics are available.

pub fn page_encoding_stats(&self) -> Option<&Vec<PageEncodingStats>>

Returns the offset for the page encoding stats, or None if no page encoding stats are available.

pub fn bloom_filter_offset(&self) -> Option<i64>

Returns the offset for the bloom filter.

pub fn bloom_filter_length(&self) -> Option<i32>

Returns the offset for the bloom filter.

pub fn column_index_offset(&self) -> Option<i64>

Returns the offset for the column index.

pub fn column_index_length(&self) -> Option<i32>

Returns the offset for the column index length.

pub(crate) fn column_index_range(&self) -> Option<Range<usize>>

Returns the range for the offset index if any

pub fn offset_index_offset(&self) -> Option<i64>

Returns the offset for the offset index.

pub fn offset_index_length(&self) -> Option<i32>

Returns the offset for the offset index length.

pub(crate) fn offset_index_range(&self) -> Option<Range<usize>>

Returns the range for the offset index if any

pub fn unencoded_byte_array_data_bytes(&self) -> Option<i64>

Returns the number of bytes of variable length data after decoding.

Only set for BYTE_ARRAY columns. This field may not be set by older writers.

pub fn repetition_level_histogram(&self) -> Option<&LevelHistogram>

Returns the repetition level histogram.

The returned value vec[i] is how many values are at repetition level i. For example, vec[0] indicates how many rows the page contains. This field may not be set by older writers.

pub fn definition_level_histogram(&self) -> Option<&LevelHistogram>

Returns the definition level histogram.

The returned value vec[i] is how many values are at definition level i. For example, vec[max_definition_level] indicates how many non-null values are present in the page. This field may not be set by older writers.

pub fn from_thrift(column_descr: ColumnDescPtr, cc: ColumnChunk) -> Result<Self>

Method to convert from Thrift.

pub fn to_thrift(&self) -> ColumnChunk

Method to convert to Thrift.

pub fn to_column_metadata_thrift(&self) -> ColumnMetaData

Method to convert to Thrift ColumnMetaData

pub fn into_builder(self) -> ColumnChunkMetaDataBuilder

Converts this ColumnChunkMetaData into a ColumnChunkMetaDataBuilder