Struct parquet::file::metadata::FileMetaData

pub struct FileMetaData { /* private fields */ }

Metadata for a Parquet file.

Implementations

impl FileMetaData

pub fn new( version: i32, num_rows: i64, created_by: Option<String>, key_value_metadata: Option<Vec<KeyValue>>, schema_descr: SchemaDescPtr, column_orders: Option<Vec<ColumnOrder>> ) -> Self

Creates new file metadata.

pub fn version(&self) -> i32

Returns version of this file.

pub fn num_rows(&self) -> i64

Returns number of rows in the file.

pub fn created_by(&self) -> Option<&str>

String message for application that wrote this file.

This should have the following format: <application> version <application version> (build <application build hash>).

parquet-mr version 1.8.0 (build 0fda28af84b9746396014ad6a415b90592a98b3b)

pub fn key_value_metadata(&self) -> Option<&Vec<KeyValue>>

Returns key_value_metadata of this file.

pub fn schema(&self) -> &SchemaType

Returns Parquet [’Type`] that describes schema in this file.

pub fn schema_descr(&self) -> &SchemaDescriptor

Returns a reference to schema descriptor.

pub fn schema_descr_ptr(&self) -> SchemaDescPtr

Returns reference counted clone for schema descriptor.

pub fn column_orders(&self) -> Option<&Vec<ColumnOrder>>

Column (sort) order used for min and max values of each column in this file.

Each column order corresponds to one column, determined by its position in the list, matching the position of the column in the schema.

When None is returned, there are no column orders available, and each column should be assumed to have undefined (legacy) column order.

pub fn column_order(&self, i: usize) -> ColumnOrder

Returns column order for ith column in this file. If column orders are not available, returns undefined (legacy) column order.