parquet::file::metadata::writer

Struct ThriftMetadataWriter

pub(crate) struct ThriftMetadataWriter<'a, W: Write> {
    buf: &'a mut TrackedWrite<W>,
    schema_descr: &'a SchemaDescPtr,
    row_groups: Vec<RowGroupMetaData>,
    column_indexes: Option<Vec<Vec<Option<ColumnIndexMetaData>>>>,
    offset_indexes: Option<Vec<Vec<Option<OffsetIndexMetaData>>>>,
    key_value_metadata: Option<Vec<KeyValue>>,
    created_by: Option<String>,
    object_writer: MetadataObjectWriter,
    writer_version: i32,
}

Writes crate::file::metadata structures to a thrift encoded byte stream

See ParquetMetaDataWriter for background and example.

Fields
buf: &'a mut TrackedWrite<W>schema_descr: &'a SchemaDescPtrrow_groups: Vec<RowGroupMetaData>column_indexes: Option<Vec<Vec<Option<ColumnIndexMetaData>>>>offset_indexes: Option<Vec<Vec<Option<OffsetIndexMetaData>>>>key_value_metadata: Option<Vec<KeyValue>>created_by: Option<String>object_writer: MetadataObjectWriterwriter_version: i32
Implementations

impl<'a, W: Write> ThriftMetadataWriter<'a, W>

fn write_offset_indexes( &mut self, offset_indexes: &[Vec<Option<OffsetIndexMetaData>>], ) -> Result<()>

Serialize all the offset indexes to self.buf,

Note: also updates the ColumnChunk::offset_index_offset and ColumnChunk::offset_index_length to reflect the position and length of the serialized offset indexes.

fn write_column_indexes( &mut self, column_indexes: &[Vec<Option<ColumnIndexMetaData>>], ) -> Result<()>

Serialize all the column indexes to the self.buf

Note: also updates the ColumnChunk::column_index_offset and ColumnChunk::column_index_length to reflect the position and length of the serialized column indexes.

fn finalize_column_indexes(&mut self) -> Result<Option<ParquetColumnIndex>>

Serialize the column indexes and transform to Option<ParquetColumnIndex>

fn finalize_offset_indexes(&mut self) -> Result<Option<ParquetOffsetIndex>>

Serialize the offset indexes and transform to Option<ParquetOffsetIndex>

pub fn finish(self) -> Result<ParquetMetaData>

Assembles and writes the final metadata to self.buf

pub fn new( buf: &'a mut TrackedWrite<W>, schema_descr: &'a SchemaDescPtr, row_groups: Vec<RowGroupMetaData>, created_by: Option<String>, writer_version: i32, ) -> Self

pub fn with_column_indexes( self, column_indexes: Vec<Vec<Option<ColumnIndexMetaData>>>, ) -> Self

pub fn with_offset_indexes( self, offset_indexes: Vec<Vec<Option<OffsetIndexMetaData>>>, ) -> Self

pub fn with_key_value_metadata(self, key_value_metadata: Vec<KeyValue>) -> Self

pub fn with_file_encryptor( self, file_encryptor: Option<Arc<FileEncryptor>>, ) -> Self