parquet::file::metadata::writer

Struct ThriftMetadataWriter

pub(crate) struct ThriftMetadataWriter<'a, W: Write> {
    buf: &'a mut TrackedWrite<W>,
    schema: &'a TypePtr,
    schema_descr: &'a SchemaDescPtr,
    row_groups: Vec<RowGroup>,
    column_indexes: Option<&'a [Vec<Option<ColumnIndex>>]>,
    offset_indexes: Option<&'a [Vec<Option<OffsetIndex>>]>,
    key_value_metadata: Option<Vec<KeyValue>>,
    created_by: Option<String>,
    writer_version: i32,
}

Writes crate::file::metadata structures to a thrift encoded byte stream

See ParquetMetaDataWriter for background and example.

Fields
buf: &'a mut TrackedWrite<W>schema: &'a TypePtrschema_descr: &'a SchemaDescPtrrow_groups: Vec<RowGroup>column_indexes: Option<&'a [Vec<Option<ColumnIndex>>]>offset_indexes: Option<&'a [Vec<Option<OffsetIndex>>]>key_value_metadata: Option<Vec<KeyValue>>created_by: Option<String>writer_version: i32
Implementations

impl<'a, W: Write> ThriftMetadataWriter<'a, W>

fn write_offset_indexes( &mut self, offset_indexes: &[Vec<Option<OffsetIndex>>], ) -> Result<()>

Serialize all the offset indexes to self.buf,

Note: also updates the ColumnChunk::offset_index_offset and ColumnChunk::offset_index_length to reflect the position and length of the serialized offset indexes.

fn write_column_indexes( &mut self, column_indexes: &[Vec<Option<ColumnIndex>>], ) -> Result<()>

Serialize all the column indexes to the self.buf

Note: also updates the ColumnChunk::column_index_offset and ColumnChunk::column_index_length to reflect the position and length of the serialized column indexes.

pub fn finish(self) -> Result<FileMetaData>

Assembles and writes the final metadata to self.buf

pub fn new( buf: &'a mut TrackedWrite<W>, schema: &'a TypePtr, schema_descr: &'a SchemaDescPtr, row_groups: Vec<RowGroup>, created_by: Option<String>, writer_version: i32, ) -> Self

pub fn with_column_indexes( self, column_indexes: &'a [Vec<Option<ColumnIndex>>], ) -> Self

pub fn with_offset_indexes( self, offset_indexes: &'a [Vec<Option<OffsetIndex>>], ) -> Self

pub fn with_key_value_metadata(self, key_value_metadata: Vec<KeyValue>) -> Self