parquet::file::page_index

Module index_reader

Support for reading ColumnIndexMetaData and OffsetIndexMetaData from parquet metadata.

Structs

ThriftColumnIndex 🔒

Functions

acc_range 🔒

Computes the covering range of two optional ranges

decode_column_index 🔒

decode_offset_index 🔒

read_columns_indexesDeprecated

Reads per-column ColumnIndexMetaData for all columns of a row group by decoding ColumnIndex .

read_offset_indexesDeprecated

Reads per-column OffsetIndexMetaData for all columns of a row group by decoding OffsetIndex .