parquet::file::page_index

Module index_reader

Support for reading Index and PageLocation from parquet metadata.

Functions

acc_range 🔒

Computes the covering range of two optional ranges

decode_column_index 🔒

decode_offset_index 🔒

decode_page_locations 🔒

read_columns_indexes

Reads per-column Index for all columns of a row group by decoding ColumnIndex .

read_offset_indexes

Reads per-column OffsetIndexMetaData for all columns of a row group by decoding OffsetIndex .

read_pages_locationsDeprecated

Reads OffsetIndex, per-page PageLocation for all columns of a row group.