parquet::file

Module properties

Configuration via WriterProperties and ReaderProperties

Structs

BloomFilterProperties

Controls the bloom filter to be computed by the writer.

ColumnProperties 🔒

Container for column properties that can be changed as part of writer.

ReaderProperties

Configuration settings for reading parquet files.

ReaderPropertiesBuilder

Builder for parquet file reader configuration. See example on ReaderProperties

WriterProperties

Configuration settings for writing parquet files.

WriterPropertiesBuilder

Builder for WriterProperties parquet writer configuration.

Enums

BloomFilterPosition

Where in the file ArrowWriter should write Bloom filters

EnabledStatistics

Controls the level of statistics to be computed by the writer and stored in the parquet file.

WriterVersion

Parquet writer version.

Constants

DEFAULT_BLOOM_FILTER_FPP

Default value for BloomFilterProperties::fpp

DEFAULT_BLOOM_FILTER_NDV

Default value for BloomFilterProperties::ndv

DEFAULT_BLOOM_FILTER_POSITION

Default value for WriterProperties::bloom_filter_position

DEFAULT_COERCE_TYPES

Default values for WriterProperties::coerce_types

DEFAULT_COLUMN_INDEX_TRUNCATE_LENGTH

Default value for WriterProperties::column_index_truncate_length

DEFAULT_COMPRESSION

Default value for WriterProperties::compression

DEFAULT_CREATED_BY

Default value for WriterProperties::created_by

DEFAULT_DATA_PAGE_ROW_COUNT_LIMIT

Default value for WriterProperties::data_page_row_count_limit

DEFAULT_DICTIONARY_ENABLED

Default value for WriterProperties::dictionary_enabled

DEFAULT_DICTIONARY_PAGE_SIZE_LIMIT

Default value for WriterProperties::dictionary_page_size_limit

DEFAULT_MAX_ROW_GROUP_SIZE

Default value for WriterProperties::max_row_group_size

DEFAULT_MAX_STATISTICS_SIZEDeprecated

Default value for WriterProperties::max_statistics_size

DEFAULT_OFFSET_INDEX_DISABLED

Default value for WriterProperties::offset_index_disabled

DEFAULT_PAGE_SIZE

Default value for WriterProperties::data_page_size_limit

DEFAULT_READ_BLOOM_FILTER 🔒

DEFAULT_STATISTICS_ENABLED

Default value for WriterProperties::statistics_enabled

DEFAULT_STATISTICS_TRUNCATE_LENGTH

Default values for WriterProperties::statistics_truncate_length

DEFAULT_WRITER_VERSION

Default value for WriterProperties::writer_version

DEFAULT_WRITE_BATCH_SIZE

Default value for WriterProperties::write_batch_size

Type Aliases

ReaderPropertiesPtr

Reference counted reader properties.

WriterPropertiesPtr

Reference counted writer properties.