Module parquet::file::properties

Writer properties.

Usage

use parquet::{
    basic::{Compression, Encoding},
    file::properties::*,
    schema::types::ColumnPath,
};

// Create properties with default configuration.
let props = WriterProperties::builder().build();

// Use properties builder to set certain options and assemble the configuration.
let props = WriterProperties::builder()
    .set_writer_version(WriterVersion::PARQUET_1_0)
    .set_encoding(Encoding::PLAIN)
    .set_column_encoding(ColumnPath::from("col1"), Encoding::DELTA_BINARY_PACKED)
    .set_compression(Compression::SNAPPY)
    .build();

assert_eq!(props.writer_version(), WriterVersion::PARQUET_1_0);
assert_eq!(
    props.encoding(&ColumnPath::from("col1")),
    Some(Encoding::DELTA_BINARY_PACKED)
);
assert_eq!(
    props.encoding(&ColumnPath::from("col2")),
    Some(Encoding::PLAIN)
);

Reader properties.

Usage

use parquet::file::properties::ReaderProperties;

// Create properties with default configuration.
let props = ReaderProperties::builder().build();

// Use properties builder to set certain options and assemble the configuration.
let props = ReaderProperties::builder()
    .set_backward_compatible_lz4(false)
    .build();

Structs

BloomFilterProperties

Controls the bloom filter to be computed by the writer.

ReaderProperties

Reader properties.

ReaderPropertiesBuilder

Reader properties builder.

WriterProperties

Writer properties.

WriterPropertiesBuilder

Writer properties builder.

Enums

EnabledStatistics

Controls the level of statistics to be computed by the writer

WriterVersion

Parquet writer version.

Constants

DEFAULT_BLOOM_FILTER_FPP

default value for the false positive probability used in a bloom filter.

DEFAULT_BLOOM_FILTER_NDV

default value for the expected number of distinct values used in a bloom filter.

Type Definitions

ReaderPropertiesPtr

Reference counted reader properties.

WriterPropertiesPtr

Reference counted writer properties.