Struct parquet::file::properties::WriterProperties

pub struct WriterProperties { /* fields omitted */ }

Writer properties.

It is created as an immutable data structure, use WriterPropertiesBuilder to assemble the properties.

Methods

impl WriterProperties

pub fn builder() -> WriterPropertiesBuilder

Returns builder for writer properties with default values.

pub fn data_pagesize_limit(&self) -> usize

Returns data page size limit.

pub fn dictionary_pagesize_limit(&self) -> usize

Returns dictionary page size limit.

pub fn write_batch_size(&self) -> usize

Returns configured batch size for writes.

When writing a batch of data, this setting allows to split it internally into smaller batches so we can better estimate the size of a page currently being written.

pub fn max_row_group_size(&self) -> usize

Returns max size for a row group.

pub fn writer_version(&self) -> WriterVersion

Returns configured writer version.

pub fn created_by(&self) -> &str

Returns created_by string.

pub fn dictionary_data_page_encoding(&self) -> Encoding

Returns encoding for a data page, when dictionary encoding is enabled. This is not configurable.

pub fn dictionary_page_encoding(&self) -> Encoding

Returns encoding for dictionary page, when dictionary encoding is enabled. This is not configurable.

pub fn encoding(&self, col: &ColumnPath) -> Encoding

Returns encoding for a column. In case when dictionary is enabled, returns fallback encoding.

pub fn compression(&self, col: &ColumnPath) -> Compression

Returns compression codec for a column.

pub fn dictionary_enabled(&self, col: &ColumnPath) -> bool

Returns true if dictionary encoding is enabled for a column.

pub fn statistics_enabled(&self, col: &ColumnPath) -> bool

Returns true if statistics are enabled for a column.

pub fn max_statistics_size(&self, col: &ColumnPath) -> usize

Returns max size for statistics. Only applicable if statistics are enabled.