Struct parquet::file::properties::WriterPropertiesBuilder

pub struct WriterPropertiesBuilder { /* fields omitted */ }

Writer properties builder.

Methods

impl WriterPropertiesBuilder

pub fn build(self) -> WriterProperties

Finalizes the configuration and returns immutable writer properties struct.

pub fn set_writer_version(self, value: WriterVersion) -> Self

Sets writer version.

pub fn set_data_pagesize_limit(self, value: usize) -> Self

Sets data page size limit.

pub fn set_dictionary_pagesize_limit(self, value: usize) -> Self

Sets dictionary page size limit.

pub fn set_write_batch_size(self, value: usize) -> Self

Sets write batch size.

pub fn set_max_row_group_size(self, value: usize) -> Self

Sets max size for a row group.

pub fn set_created_by(self, value: String) -> Self

Sets "created by" property.

pub fn set_encoding(self, value: Encoding) -> Self

Sets encoding for any column.

If dictionary is not enabled, this is treated as a primary encoding for all columns. In case when dictionary is enabled for any column, this value is considered to be a fallback encoding for that column.

Panics if user tries to set dictionary encoding here, regardless of dictinoary encoding flag being set.

pub fn set_compression(self, value: Compression) -> Self

Sets compression codec for any column.

pub fn set_dictionary_enabled(self, value: bool) -> Self

Sets flag to enable/disable dictionary encoding for any column.

Use this method to set dictionary encoding, instead of explicitly specifying encoding in set_encoding method.

pub fn set_statistics_enabled(self, value: bool) -> Self

Sets flag to enable/disable statistics for any column.

pub fn set_max_statistics_size(self, value: usize) -> Self

Sets max statistics size for any column. Applicable only if statistics are enabled.

pub fn set_column_encoding(self, col: ColumnPath, value: Encoding) -> Self

Sets encoding for a column. Takes precedence over globally defined settings.

If dictionary is not enabled, this is treated as a primary encoding for this column. In case when dictionary is enabled for this column, either through global defaults or explicitly, this value is considered to be a fallback encoding for this column.

Panics if user tries to set dictionary encoding here, regardless of dictinoary encoding flag being set.

pub fn set_column_compression(self, col: ColumnPath, value: Compression) -> Self

Sets compression codec for a column. Takes precedence over globally defined settings.

pub fn set_column_dictionary_enabled(self, col: ColumnPath, value: bool) -> Self

Sets flag to enable/disable dictionary encoding for a column. Takes precedence over globally defined settings.

pub fn set_column_statistics_enabled(self, col: ColumnPath, value: bool) -> Self

Sets flag to enable/disable statistics for a column. Takes precedence over globally defined settings.

pub fn set_column_max_statistics_size(
self,
col: ColumnPath,
value: usize
) -> Self

Sets max size for statistics for a column. Takes precedence over globally defined settings.