parquet::file
Module reader

File reader API and methods to access file metadata, row group readers to read individual column chunks, or access record iterator.

Re-exports

pub use crate::file::serialized_reader::SerializedFileReader;

pub use crate::file::serialized_reader::SerializedPageReader;

Structs

FilePageIterator

Implementation of page iterator for parquet file.

Traits

ChunkReader

Generates Readers to read chunks of a Parquet data source.

FileReader

Parquet file reader API. With this, user can get metadata information about the Parquet file, can get reader for each row group, and access record iterator.

Length

Length should return the total number of bytes in the input source. It’s mainly used to read the metadata, which is at the end of the source.

RowGroupReader

Parquet row group reader API. With this, user can get metadata information about the row group, as well as readers for each individual column chunk.