Trait parquet::file::reader::ChunkReader

pub trait ChunkReader: Length {
    type T: Read;
    fn get_read(&self, start: u64, length: usize) -> Result<Self::T>;
}

The ChunkReader trait generates readers of chunks of a source. For a file system reader, each chunk might contain a clone of File bounded on a given range. For an object store reader, each read can be mapped to a range request.

Associated Types

type T: Read

Loading content...
Required methods

fn get_read(&self, start: u64, length: usize) -> Result<Self::T>

get a serialy readeable slice of the current reader This should fail if the slice exceeds the current bounds

Loading content...