parquet::file::reader
Trait Length

pub trait Length {
    // Required method
    fn len(&self) -> u64;
}

Length should return the total number of bytes in the input source. It’s mainly used to read the metadata, which is at the end of the source.

Required Methods

fn len(&self) -> u64

Returns the amount of bytes of the inner source.