Trait parquet::file::reader::Length

pub trait Length {
    fn len(&self) -> u64;
}

Length should return the total number of bytes in the input source. It's mainly used to read the metadata, which is at the end of the source.

Required methods

fn len(&self) -> u64

Returns the amount of bytes of the inner source.

Loading content...