parquet::file::reader

Trait RowGroupReader

pub trait RowGroupReader: Send + Sync {
    // Required methods
    fn metadata(&self) -> &RowGroupMetaData;
    fn num_columns(&self) -> usize;
    fn get_column_page_reader(&self, i: usize) -> Result<Box<dyn PageReader>>;
    fn get_column_bloom_filter(&self, i: usize) -> Option<&Sbbf>;
    fn get_row_iter(
        &self,
        projection: Option<SchemaType>,
    ) -> Result<RowIter<'_>>;

    // Provided method
    fn get_column_reader(&self, i: usize) -> Result<ColumnReader> { ... }
}

Parquet row group reader API. With this, user can get metadata information about the row group, as well as readers for each individual column chunk.

Required Methods

fn metadata(&self) -> &RowGroupMetaData

Get metadata information about this row group.

fn num_columns(&self) -> usize

Get the total number of column chunks in this row group.

fn get_column_page_reader(&self, i: usize) -> Result<Box<dyn PageReader>>

Get page reader for the ith column chunk.

fn get_column_bloom_filter(&self, i: usize) -> Option<&Sbbf>

Get bloom filter for the ith column chunk, if present and the reader was configured to read bloom filters.

fn get_row_iter(&self, projection: Option<SchemaType>) -> Result<RowIter<'_>>

Get an iterator over the row in this file, see RowIter for caveats.

Projected schema can be a subset of or equal to the file schema, when it is None, full file schema is assumed.

Provided Methods

fn get_column_reader(&self, i: usize) -> Result<ColumnReader>

Get value reader for the ith column chunk.