Trait parquet::file::reader::RowGroupReader

pub trait RowGroupReader {
    fn metadata(&self) -> &RowGroupMetaData;

    fn num_columns(&self) -> usize;

    fn get_column_page_reader(&self, i: usize) -> Result<Box<dyn PageReader>>;

    fn get_row_iter(
        &self, 
        projection: Option<SchemaType>
    ) -> Result<RowIter<'_>>;

    fn get_column_reader(&self, i: usize) -> Result<ColumnReader> { ... }
}

Parquet row group reader API. With this, user can get metadata information about the row group, as well as readers for each individual column chunk.

Required methods

fn metadata(&self) -> &RowGroupMetaData

Get metadata information about this row group.

fn num_columns(&self) -> usize

Get the total number of column chunks in this row group.

fn get_column_page_reader(&self, i: usize) -> Result<Box<dyn PageReader>>

Get page reader for the ith column chunk.

fn get_row_iter(&self, projection: Option<SchemaType>) -> Result<RowIter<'_>>

Get iterator of Rows from this row group.

Projected schema can be a subset of or equal to the file schema, when it is None, full file schema is assumed.

Provided methods

fn get_column_reader(&self, i: usize) -> Result<ColumnReader>

Get value reader for the ith column chunk.