parquet::file

Module serialized_reader

Contains implementations of the reader traits FileReader, RowGroupReader and PageReader Also contains implementations of the ChunkReader for files (with buffering) and byte arrays (RAM)

Structs

ReadOptions

A collection of options for reading a Parquet file.

ReadOptionsBuilder

A builder for ReadOptions. For the predicates that are added to the builder, they will be chained using ‘AND’ to filter the row groups.

SerializedFileReader

A serialized implementation for Parquet FileReader.

SerializedPageReader

A serialized implementation for Parquet PageReader.

SerializedPageReaderContext 🔒

SerializedRowGroupReader

A serialized implementation for Parquet RowGroupReader.

Enums

SerializedPageReaderState 🔒

Functions

decode_page 🔒

Decodes a Page from the provided buffer

get_midpoint_offset 🔒

Get midpoint offset for a row group

verify_page_header_len 🔒

verify_page_size 🔒

Type Aliases

ReadGroupPredicate

A predicate for filtering row groups, invoked with the metadata and index of each row group in the file. Only row groups for which the predicate evaluates to true will be scanned