Module parquet::file::serialized_reader

Contains implementations of the reader traits FileReader, RowGroupReader and PageReader Also contains implementations of the ChunkReader for files (with buffering) and byte arrays (RAM)

Structs

FileSource

Struct that represents a slice of a file data with independent start position and length. Internally clones provided file handle, wraps with a custom implementation of BufReader that resets position before any read.

ReadOptions

A collection of options for reading a Parquet file.

ReadOptionsBuilder

A builder for ReadOptions. For the predicates that are added to the builder, they will be chained using ‘AND’ to filter the row groups.

SerializedFileReader

A serialized implementation for Parquet FileReader.

SerializedPageReader

A serialized implementation for Parquet PageReader.

SerializedRowGroupReader

A serialized implementation for Parquet RowGroupReader.

Type Definitions

ReadGroupPredicate

A predicate for filtering row groups, invoked with the metadata and index of each row group in the file. Only row groups for which the predicate evaluates to true will be scanned