Enum parquet::file::statistics::Statistics

pub enum Statistics {
    Boolean(TypedStatistics<BoolType>),
    Int32(TypedStatistics<Int32Type>),
    Int64(TypedStatistics<Int64Type>),
    Int96(TypedStatistics<Int96Type>),
    Float(TypedStatistics<FloatType>),
    Double(TypedStatistics<DoubleType>),
    ByteArray(TypedStatistics<ByteArrayType>),
    FixedLenByteArray(TypedStatistics<FixedLenByteArrayType>),
}

Statistics for a column chunk and data page.

Variants

Boolean(TypedStatistics<BoolType>)

Int32(TypedStatistics<Int32Type>)

Int64(TypedStatistics<Int64Type>)

Int96(TypedStatistics<Int96Type>)

Float(TypedStatistics<FloatType>)

Double(TypedStatistics<DoubleType>)

ByteArray(TypedStatistics<ByteArrayType>)

FixedLenByteArray(TypedStatistics<FixedLenByteArrayType>)

Implementations

impl Statistics

pub fn boolean(
min: Option<bool>,
max: Option<bool>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn int32(
min: Option<i32>,
max: Option<i32>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn int64(
min: Option<i64>,
max: Option<i64>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn int96(
min: Option<Int96>,
max: Option<Int96>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn float(
min: Option<f32>,
max: Option<f32>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn double(
min: Option<f64>,
max: Option<f64>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn byte_array(
min: Option<ByteArray>,
max: Option<ByteArray>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn fixed_len_byte_array(
min: Option<FixedLenByteArray>,
max: Option<FixedLenByteArray>,
distinct: Option<u64>,
nulls: u64,
is_deprecated: bool
) -> Self

pub fn is_min_max_deprecated(&self) -> bool

Returns true if statistics have old min and max fields set. This means that the column order is likely to be undefined, which, for old files could mean a signed sort order of values.

Refer to ColumnOrder and SortOrder for more information.

pub fn distinct_count(&self) -> Option<u64>

Returns optional value of number of distinct values occurring. When it is None, the value should be ignored.

pub fn null_count(&self) -> u64

Returns number of null values for the column. Note that this includes all nulls when column is part of the complex type.

pub fn has_nulls(&self) -> bool

Returns true if statistics collected any null values, false otherwise.

pub fn has_min_max_set(&self) -> bool

Returns true if min value and max value are set. Normally both min/max values will be set to Some(value) or None.

pub fn min_bytes(&self) -> &[u8]

Returns slice of bytes that represent min value. Panics if min value is not set.

pub fn max_bytes(&self) -> &[u8]

Returns slice of bytes that represent max value. Panics if max value is not set.

pub fn physical_type(&self) -> Type

Returns physical type associated with statistics.