parquet::file::statistics

Type Alias TypedStatistics

pub type TypedStatistics<T> = ValueStatistics<<T as DataType>::T>;

Typed implementation for Statistics.

Aliased Type

struct TypedStatistics<T> {
    min: Option<<T as DataType>::T>,
    max: Option<<T as DataType>::T>,
    distinct_count: Option<u64>,
    null_count: Option<u64>,
    is_max_value_exact: bool,
    is_min_value_exact: bool,
    is_min_max_deprecated: bool,
    is_min_max_backwards_compatible: bool,
}

Fields
min: Option<<T as DataType>::T>max: Option<<T as DataType>::T>distinct_count: Option<u64>null_count: Option<u64>is_max_value_exact: boolis_min_value_exact: boolis_min_max_deprecated: bool

If true populate the deprecated min and max fields instead of min_value and max_value

is_min_max_backwards_compatible: bool

If true the statistics are compatible with the deprecated min and max fields. See ValueStatistics::is_min_max_backwards_compatible