Module parquet::memory

Utility methods and structs for working with memory.

Structs

Buffer
A resize-able buffer class with generic member, with optional memory tracker.

BufferPtr
An representation of a slice on a reference-counting and read-only byte array. Sub-slices can be further created from this. The byte array will be released when all slices are dropped.

MemTracker
Struct to track memory usage information.

Type Definitions

ByteBuffer
Type alias for Buffer.

ByteBufferPtr
Type alias for BufferPtr.

MemTrackerPtr
Reference counted pointer for MemTracker.

WeakMemTrackerPtr
Non-owning reference for MemTracker.