parquet

Module parquet_macros

This is a collection of macros used to parse Thrift IDL descriptions of structs, unions, and enums into their corresponding Rust types. These macros will also generate the code necessary to serialize and deserialize to/from the Thrift compact protocol.

Further details of how to use them (and other aspects of the Thrift serialization process) can be found in THRIFT.md.