parquet::parquet_thrift

Trait ReadThrift

pub(crate) trait ReadThrift<'a, R: ThriftCompactInputProtocol<'a>> {
    // Required method
    fn read_thrift(prot: &mut R) -> Result<Self>
       where Self: Sized;
}

Trait implemented for objects that can be deserialized from a Thrift input stream. Implementations are provided for Thrift primitive types.

Required Methods

fn read_thrift(prot: &mut R) -> Result<Self>
where Self: Sized,

Read an object of type Self from the input protocol object.