parquet
Module record

Contains record-based API for reading Parquet files.

Modules

api 🔒

Contains Row enum that is used to represent record in Rust.

reader

Contains implementation of record assembly and converting Parquet types into Rows.

record_reader 🔒

record_writer 🔒

triplet 🔒

Structs

List

List represents a list which contains an array of elements.

Map

Map represents a map which contains a list of key->value pairs.

Row

Row represents a nested Parquet record.

RowColumnIter

RowColumnIter represents an iterator over column names and values in a Row.

Enums

Field

API to represent a single field in a Row.

Traits

ListAccessor

Trait for type-safe access of an index for a List. Note that the get_XXX methods do not do bound checking.

MapAccessor

Trait for type-safe access of an index for a Map

RecordReader

Read up to num_records records from row_group_reader into self.

RecordWriter

Trait describing how to write a record (the implementator) to a row group writer.

RowAccessor

Trait for type-safe convenient access to fields within a Row.

RowFormatter

Trait for formatting fields within a Row.