Module parquet::record

Contains record-based API for reading Parquet files.

Modules

reader
Contains implementation of record assembly and converting Parquet types into Rows.

Structs

List
List represents a list which contains an array of elements.

Map
Map represents a map which contains a list of key->value pairs.

Row
Row represents a nested Parquet record.

Enums

Field
API to represent a single field in a Row.

Traits

ListAccessor
Trait for type-safe access of an index for a List. Note that the get_XXX methods do not do bound checking.

MapAccessor
Trait for type-safe access of an index for a Map

RecordWriter

RowAccessor
Trait for type-safe convenient access to fields within a Row.