Module parquet::record::api

Contains Row enum that is used to represent record in Rust.

Macros

list_complex_accessor 🔒

Macro to generate type-safe get_xxx methods for reference types e.g. get_list, get_map

list_primitive_accessor 🔒

Macro to generate type-safe get_xxx methods for primitive types, e.g. get_bool, get_short

map_list_primitive_accessor 🔒

Macro to generate type-safe get_xxx methods for primitive types, e.g. get_bool, get_short

nyi 🔒

Macro as a shortcut to generate ‘not yet implemented’ panic error.

row_complex_accessor 🔒

Macro to generate type-safe get_xxx methods for reference types, e.g. get_list, get_map.

row_primitive_accessor 🔒

Macro to generate type-safe get_xxx methods for primitive types, e.g. get_bool, get_short.

Structs

List

List represents a list which contains an array of elements.

Map

Map represents a map which contains a list of key->value pairs.

MapList 🔒

Row

Row represents a nested Parquet record.

RowColumnIter

RowColumnIter represents an iterator over column names and values in a Row.

Enums

Field

API to represent a single field in a Row.

Traits

ListAccessor

Trait for type-safe access of an index for a List. Note that the get_XXX methods do not do bound checking.

MapAccessor

Trait for type-safe access of an index for a Map

RowAccessor

Trait for type-safe convenient access to fields within a Row.

RowFormatter

Trait for formatting fields within a Row.

Functions

convert_date_to_string 🔒

Helper method to convert Parquet date into a string. Input value is a number of days since the epoch in UTC. Date is displayed in local timezone.

convert_decimal_to_string 🔒

Helper method to convert Parquet decimal into a string. We assert that scale >= 0 and precision > scale, but this will be enforced when constructing Parquet schema.

convert_timestamp_micros_to_string 🔒

Helper method to convert Parquet timestamp into a string. Input value is a number of microseconds since the epoch in UTC. Datetime is displayed in local timezone.

convert_timestamp_millis_to_string 🔒

Helper method to convert Parquet timestamp into a string. Input value is a number of milliseconds since the epoch in UTC. Datetime is displayed in local timezone.

convert_timestamp_secs_to_string 🔒

Helper method to convert Parquet timestamp into a string. Input value is a number of seconds since the epoch in UTC. Datetime is displayed in local timezone.

make_list

Constructs a List from the list of fields and returns it.

make_map

Constructs a Map from the list of entries and returns it.

make_row

Constructs a Row from the list of fields and returns it.