Module parquet::record::reader

Contains implementation of record assembly and converting Parquet types into Rows.

Structs

ReaderIter

Internal iterator of Rows for a reader.

RowIter

Access parquet data as an iterator of Row

TreeBuilder

Tree builder for Reader enum. Serves as a container of options for building a reader tree and a builder, and accessing a records iterator RowIter.

Enums

Either 🔒

The enum Either with variants That represents a reference and a box of FileReader.

Reader

Reader tree for record assembly

Constants

DEFAULT_BATCH_SIZE 🔒

Default batch size for a reader