Trait parquet::record::RecordWriter

pub trait RecordWriter<T> {
    // Required methods
    fn write_to_row_group<W: Write + Send>(
        &self,
        row_group_writer: &mut SerializedRowGroupWriter<'_, W>
    ) -> Result<(), ParquetError>;
    fn schema(&self) -> Result<TypePtr, ParquetError>;
}

write_to_row_group writes from self into row_group_writer schema builds the schema used by row_group_writer The type parameter T is used to work around the rust orphan rule when implementing on types such as &[T].

Required Methods

fn write_to_row_group<W: Write + Send>( &self, row_group_writer: &mut SerializedRowGroupWriter<'_, W> ) -> Result<(), ParquetError>

fn schema(&self) -> Result<TypePtr, ParquetError>

Generated schema

Object Safety

This trait is not object safe.