parquet::schema
Module types

Contains structs and methods to build Parquet schema and schema descriptors.

Structs

BasicTypeInfo

Basic type info. This contains information such as the name of the type, the repetition level, the logical type and the kind of the type (group, primitive).

ColumnDescriptor

Physical type for leaf-level primitive columns.

ColumnPath

Represents the location of a column in a Parquet schema

GroupTypeBuilder

A builder for group types. All attributes are optional except the name. Note that if not specified explicitly, None is used as the repetition of the group, which means it is a root (message) type.

PrimitiveTypeBuilder

A builder for primitive types. All attributes are optional except the name and physical type. Note that if not specified explicitly, Repetition::OPTIONAL is used.

SchemaDescriptor

Schema of a Parquet file.

Enums

Type

Representation of a Parquet type.

Functions

build_tree 🔒

from_thrift

Method to convert from Thrift.

from_thrift_helper 🔒

Constructs a new Type from the elements, starting at index index. The first result is the starting index for the next Type after this one. If it is equal to elements.len(), then this Type is the last one. The second result is the result Type.

to_thrift

Method to convert to Thrift.

to_thrift_helper 🔒

Constructs list of SchemaElement from the schema using depth-first traversal. Here we assume that schema is always valid and starts with group type.

Type Aliases

ColumnDescPtr

Type alias for Arc<ColumnDescriptor>.

SchemaDescPtr

Type alias for Arc<SchemaDescriptor>.

TypePtr

Type alias for Arc<Type>.