parquet::schema::types

Struct ColumnDescriptor

pub struct ColumnDescriptor {
    primitive_type: TypePtr,
    max_def_level: i16,
    max_rep_level: i16,
    path: ColumnPath,
}

Physical type for leaf-level primitive columns.

Also includes the maximum definition and repetition levels required to re-assemble nested data.

Fields
primitive_type: TypePtr

The “leaf” primitive type of this column

max_def_level: i16

The maximum definition level for this column

max_rep_level: i16

The maximum repetition level for this column

path: ColumnPath

The path of this column. For instance, “a.b.c.d”.

Implementations

impl ColumnDescriptor

pub fn new( primitive_type: TypePtr, max_def_level: i16, max_rep_level: i16, path: ColumnPath, ) -> Self

Creates new descriptor for leaf-level column.

pub fn max_def_level(&self) -> i16

Returns maximum definition level for this column.

pub fn max_rep_level(&self) -> i16

Returns maximum repetition level for this column.

pub fn path(&self) -> &ColumnPath

Returns ColumnPath for this column.

pub fn self_type(&self) -> &Type

Returns self type Type for this leaf column.

pub fn self_type_ptr(&self) -> TypePtr

Returns self type TypePtr for this leaf column.

pub fn name(&self) -> &str

Returns column name.

pub fn converted_type(&self) -> ConvertedType

Returns ConvertedType for this column.

pub fn logical_type(&self) -> Option<LogicalType>

Returns LogicalType for this column.

pub fn physical_type(&self) -> PhysicalType

Returns physical type for this column. Note that it will panic if called on a non-primitive type.

pub fn type_length(&self) -> i32

Returns type length for this column. Note that it will panic if called on a non-primitive type.

pub fn type_precision(&self) -> i32

Returns type precision for this column. Note that it will panic if called on a non-primitive type.

pub fn type_scale(&self) -> i32

Returns type scale for this column. Note that it will panic if called on a non-primitive type.

pub fn sort_order(&self) -> SortOrder

Returns the sort order for this column