Struct parquet::schema::types::SchemaDescriptor

pub struct SchemaDescriptor { /* private fields */ }

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

Implementations

impl SchemaDescriptor

pub fn new(tp: TypePtr) -> Self

Creates new schema descriptor from Parquet schema.

pub fn column(&self, i: usize) -> ColumnDescPtr

Returns ColumnDescriptor for a field position.

pub fn columns(&self) -> &[ColumnDescPtr]

Returns slice of ColumnDescriptor.

pub fn num_columns(&self) -> usize

Returns number of leaf-level columns.

pub fn get_column_root(&self, i: usize) -> &Type

Returns column root Type for a leaf position.

pub fn get_column_root_ptr(&self, i: usize) -> TypePtr

Returns column root Type pointer for a leaf position.

pub fn get_column_root_idx(&self, leaf: usize) -> usize

Returns the index of the root column for a field position

pub fn root_schema(&self) -> &Type

Returns schema as Type.

pub fn root_schema_ptr(&self) -> TypePtr

pub fn name(&self) -> &str

Returns schema name.